@@ -4,16 +4,18 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "sparkx"
-version = "1.3.0"
+version = "2.0.0"
 authors = [
-    { name="Niklas Götz", email="goetz@itp.uni-frankfurt.de"},
+    { name="Nils Sass", email="nsass@itp.uni-frankfurt.de" },
     { name="Hendrik Roch", email="hroch@wayne.edu" },
-    { name="Nils Sass", email="nsass@itp.uni-frankfurt.de" }
+    { name="Niklas Götz", email="goetz@itp.uni-frankfurt.de"},
+    { name="Renata Krupczak", email="rkrupczak@physik.uni-bielefeld.de"},
+    { name="Lucas Constantin", email="constantin@itp.uni-frankfurt.de"}
 ]
 description = "Software Package for Analyzing Relativistic Kinematics in Collision eXperiments"
 readme = "README.md"
 license = {file = "LICENSE"}
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Python :: 3",
     "Intended Audience :: Developers",
@@ -24,19 +26,15 @@ classifiers = [
 ]
 
 dependencies = [
-    "particle==0.23.0",
+    "particle>=0.23.0",
     "numpy>=1.23.5",
     "scipy>=1.10.1",
-    "abc-property==1.0",
-    "fastjet==3.4.1.3",
+    "fastjet>=3.4.2.1",
     "matplotlib>=3.7.1"
 ]
 
-[tool.autopep8]
-max_line_length = 80
-in-place = true
-recursive = true
-aggressive = 3
+[tool.black]
+line-length = 80
 
 
 [project.urls]
