@@ -9,12 +9,18 @@
 
 import numpy as np
 import warnings
-from typing import List, Union
+from typing import Union, List
 
 
 class CentralityClasses:
     """
     Class for defining centrality classes based on event multiplicity.
+
+    .. note::
+        It is the user's responsibility to ensure that the amount of events used
+        to determine the centrality classes is sufficient to provide reliable
+        results. The recommended minimum number of events is at least a few
+        hundred.
 
     Parameters
     ----------
@@ -27,7 +33,8 @@ class CentralityClasses:
     Raises
     ------
     TypeError
-        If `events_multiplicity` or `centrality_bins` is not a list or numpy.ndarray.
+        If :code:`events_multiplicity` or :code:`centrality_bins` is not a list
+        or :code:`numpy.ndarray`.
 
     Attributes
     ----------
@@ -57,26 +64,39 @@ class CentralityClasses:
     .. code-block:: python
         :linenos:
 
-        >>> centrality_obj = CentralityClasses(events_multiplicity=[10, 15, 20, 25],
-        ...                                   centrality_bins=[0, 25, 50, 75, 100])
-        >>> centrality_obj.get_centrality_class(18)
-        1
+        >>> from sparkx import *
+        >>> import random as rd
+        >>> rd.seed(0)
+
+        >>> # generate 300 random event multiplicities between 50 and 1500
+        >>> events_multiplicity = [rd.randint(50, 1500) for _ in range(300)]
+        >>> centrality_bins = [0, 10, 30, 50, 70, 90, 100]
+        >>> centrality_obj = CentralityClasses(events_multiplicity=events_multiplicity,
+        ...                                   centrality_bins=centrality_bins)
+        >>> centrality_obj.get_centrality_class(1490)
+        0
         >>> centrality_obj.output_centrality_classes('centrality_output.txt')
     """
 
-    def __init__(self, events_multiplicity: Union[List[float], np.ndarray],
-                 centrality_bins: Union[List[float], np.ndarray]) -> None:
+    def __init__(
+        self,
+        events_multiplicity: Union[List[float], np.ndarray],
+        centrality_bins: Union[List[float], np.ndarray],
+    ) -> None:
         if not isinstance(events_multiplicity, (list, np.ndarray)):
             raise TypeError(
-                "'events_multiplicity' is not list or numpy.ndarray")
+                "'events_multiplicity' is not list or numpy.ndarray"
+            )
         if not isinstance(centrality_bins, (list, np.ndarray)):
             raise TypeError("'centrality_bins' is not list or numpy.ndarray")
-
         # Check if centrality_bins is sorted
-        if not all(centrality_bins[i] <= centrality_bins[i + 1]
-                   for i in range(len(centrality_bins) - 1)):
+        if not all(
+            centrality_bins[i] <= centrality_bins[i + 1]
+            for i in range(len(centrality_bins) - 1)
+        ):
             warnings.warn(
-                "'centrality_bins' is not sorted. Sorting automatically.")
+                "'centrality_bins' is not sorted. Sorting automatically."
+            )
             centrality_bins.sort()
 
         # Check for uniqueness of values
@@ -93,20 +113,22 @@ class CentralityClasses:
 
         if multiple_same_entries:
             warnings.warn(
-                "'centrality_bins' contains duplicate values. They are removed automatically.")
+                "'centrality_bins' contains duplicate values. They are removed automatically."
+            )
 
         # Check for negative values and values greater than 100
         if any(value < 0.0 or value > 100.0 for value in centrality_bins):
             raise ValueError(
-                "'centrality_bins' contains values less than 0 or greater than 100.")
+                "'centrality_bins' contains values less than 0 or greater than 100."
+            )
 
         self.events_multiplicity_ = events_multiplicity
         self.centrality_bins_ = unique_bins
 
-        self.dNchdetaMin_: list[float] = []
-        self.dNchdetaMax_: list[float] = []
-        self.dNchdetaAvg_: list[float] = []
-        self.dNchdetaAvgErr_: list[float] = []
+        self.dNchdetaMin_: List[float] = []
+        self.dNchdetaMax_: List[float] = []
+        self.dNchdetaAvg_: List[float] = []
+        self.dNchdetaAvgErr_: List[float] = []
 
         self.__create_centrality_classes()
 
@@ -148,9 +170,10 @@ class CentralityClasses:
 
         # distribute the numbers evenly
         for i, multiplicity in enumerate(self.events_multiplicity_):
-            if (multiplicity < 0):
+            if multiplicity < 0:
                 raise ValueError(
-                    "Multiplicity in 'events_multiplicity' is negative")
+                    "Multiplicity in 'events_multiplicity' is negative"
+                )
             if i % 4 == 0:
                 event_sample_A.append(multiplicity)
             elif i % 4 == 1:
@@ -167,12 +190,9 @@ class CentralityClasses:
 
         MinRecord = int(number_events / 4 * self.centrality_bins_[0] / 100.0)
         for i in range(1, len(self.centrality_bins_)):
-
             MaxRecord = int(
-                number_events /
-                4 *
-                self.centrality_bins_[i] /
-                100.0)
+                number_events / 4 * self.centrality_bins_[i] / 100.0
+            )
 
             AvgA = np.mean(event_sample_A[MinRecord:MaxRecord])
             AvgB = np.mean(event_sample_B[MinRecord:MaxRecord])
@@ -180,8 +200,15 @@ class CentralityClasses:
             AvgD = np.mean(event_sample_D[MinRecord:MaxRecord])
 
             Avg = (AvgA + AvgB + AvgC + AvgD) / 4.0
-            Err = np.sqrt(((AvgA - Avg)**2 + (AvgB - Avg)**2 +
-                          (AvgC - Avg)**2 + (AvgD - Avg)**2) / 3.0)
+            Err = np.sqrt(
+                (
+                    (AvgA - Avg) ** 2
+                    + (AvgB - Avg) ** 2
+                    + (AvgC - Avg) ** 2
+                    + (AvgD - Avg) ** 2
+                )
+                / 3.0
+            )
 
             self.dNchdetaAvg_.append(Avg)
             self.dNchdetaAvgErr_.append(Err)
@@ -206,6 +233,11 @@ class CentralityClasses:
         This function determines the index of the centrality bin for a given
         multiplicity value based on the predefined centrality classes.
 
+        In the case that the multiplicity input exceeds the largest or smallest
+        value of the multiplicity used to determine the centrality classes, the
+        function returns the index of the most central or most peripheral bin,
+        respectively.
+
         Parameters
         ----------
         dNchdEta : float
@@ -215,17 +247,6 @@ class CentralityClasses:
         -------
         int
             Index of the centrality bin.
-
-        Examples
-        --------
-        .. highlight:: python
-        .. code-block:: python
-            :linenos:
-
-            >>> centrality_obj = CentralityClasses(events_multiplicity=[10, 15, 20, 25],
-            ...                                   centrality_bins=[0, 25, 50, 75, 100])
-            >>> centrality_obj.get_centrality_class(18)
-            1
         """
         # check if the multiplicity is in the most central bin
         if dNchdEta >= self.dNchdetaMin_[0]:
@@ -237,7 +258,8 @@ class CentralityClasses:
         else:
             for i in range(1, len(self.dNchdetaMin_) - 1):
                 if (dNchdEta >= self.dNchdetaMin_[i]) and (
-                        dNchdEta < self.dNchdetaMin_[i - 1]):
+                    dNchdEta < self.dNchdetaMin_[i - 1]
+                ):
                     return i
         # default case to satisfy mypy's requirement for a return statement
         return -1
@@ -254,17 +276,7 @@ class CentralityClasses:
         Raises
         ------
         TypeError
-            If `fname` is not a string.
-
-        Examples
-        --------
-        .. highlight:: python
-        .. code-block:: python
-            :linenos:
-
-            >>> centrality_obj = CentralityClasses(events_multiplicity=[10, 15, 20, 25],
-            ...                                   centrality_bins=[0, 25, 50, 75, 100])
-            >>> centrality_obj.output_centrality_classes('centrality_output.txt')
+            If :code:`fname` is not a string.
 
         Notes
         -----
@@ -274,14 +286,15 @@ class CentralityClasses:
         # Check if fname is a string
         if not isinstance(fname, str):
             raise TypeError("'fname' should be a string.")
-
         # Write the information to the file
-        with open(fname, 'w') as out_stream:
+        with open(fname, "w") as out_stream:
             out_stream.write(
-                "# CentralityMin CentralityMax dNchdEtaMin dNchdEtaMax dNchdEtaAvg dNchdEtaAvgErr\n")
+                "# CentralityMin CentralityMax dNchdEtaMin dNchdEtaMax dNchdEtaAvg dNchdEtaAvgErr\n"
+            )
 
             for i in range(1, len(self.dNchdetaMin_)):
                 out_stream.write(
                     f"{self.centrality_bins_[i - 1]} - {self.centrality_bins_[i]} "
                     f"{self.dNchdetaMin_[i - 1]} {self.dNchdetaMax_[i - 1]} "
-                    f"{self.dNchdetaAvg_[i - 1]} {self.dNchdetaAvgErr_[i - 1]}\n")
+                    f"{self.dNchdetaAvg_[i - 1]} {self.dNchdetaAvgErr_[i - 1]}\n"
+                )
