@@ -10,6 +10,7 @@
 import numpy as np
 import csv
 import warnings
+from typing import Optional, Union, List, Tuple, Dict
 
 
 class Histogram:
@@ -17,7 +18,7 @@ class Histogram:
     Defines a histogram object.
 
     The histograms can be initialized either with a tuple
-    (hist_min,hist_max,num_bins) or a list/numpy.ndarray containing the bin
+    :code:`(hist_min,hist_max,num_bins)` or a list/numpy.ndarray containing the bin
     boundaries, which allows for different bin widths.
     Multiple histograms can be added and averaged.
 
@@ -136,15 +137,20 @@ class Histogram:
         >>> hist = histObj.histogram()
     """
 
-    def __init__(self, bin_boundaries):
-        self.number_of_bins_ = None
-        self.bin_edges_ = None
-        self.number_of_histograms_ = 1
-        self.histograms_ = None
-        self.histograms_raw_count_ = None
-        self.error_ = None
-        self.scaling_ = None
-        self.systematic_error_ = None
+    def __init__(
+        self,
+        bin_boundaries: Union[
+            Tuple[float, float, int], List[float], np.ndarray
+        ],
+    ) -> None:
+        self.number_of_bins_: Optional[int] = None
+        self.bin_edges_: Optional[np.ndarray] = None
+        self.number_of_histograms_: int = 1
+        self.histograms_: Optional[np.ndarray] = None
+        self.histograms_raw_count_: Optional[np.ndarray] = None
+        self.error_: Optional[np.ndarray] = None
+        self.scaling_: Optional[np.ndarray] = None
+        self.systematic_error_: Optional[np.ndarray] = None
 
         if isinstance(bin_boundaries, tuple) and len(bin_boundaries) == 3:
             hist_min = bin_boundaries[0]
@@ -152,10 +158,10 @@ class Histogram:
             num_bins = bin_boundaries[2]
 
             if hist_min > hist_max or hist_min == hist_max:
-                raise ValueError('hist_min must be smaller than hist_max')
+                raise ValueError("hist_min must be smaller than hist_max")
 
             elif not isinstance(num_bins, int) or num_bins <= 0:
-                raise ValueError('Number of bins must be a positive integer')
+                raise ValueError("Number of bins must be a positive integer")
 
             self.number_of_bins_ = num_bins
             self.bin_edges_ = np.linspace(hist_min, hist_max, num=num_bins + 1)
@@ -166,23 +172,25 @@ class Histogram:
             self.systematic_error_ = np.asarray([np.zeros(num_bins)])
 
         elif isinstance(bin_boundaries, (list, np.ndarray)):
-
             self.number_of_bins_ = len(bin_boundaries) - 1
             self.bin_edges_ = np.asarray(bin_boundaries)
             self.histograms_ = np.asarray([np.zeros(self.number_of_bins_)])
             self.histograms_raw_count_ = np.asarray(
-                [np.zeros(self.number_of_bins_)])
+                [np.zeros(self.number_of_bins_)]
+            )
             self.scaling_ = np.asarray([np.ones(self.number_of_bins_)])
             self.error_ = np.asarray([np.zeros(self.number_of_bins_)])
             self.systematic_error_ = np.asarray(
-                [np.zeros(self.number_of_bins_)])
+                [np.zeros(self.number_of_bins_)]
+            )
 
         else:
             raise TypeError(
-                'Input must be a tuple (hist_min, hist_max, num_bins) ' +
-                'or a list/numpy.ndarray containing the bin edges!')
+                "Input must be a tuple (hist_min, hist_max, num_bins) "
+                + "or a list/numpy.ndarray containing the bin edges!"
+            )
 
-    def histogram(self):
+    def histogram(self) -> np.ndarray:
         """
         Get the current histogram(s).
 
@@ -191,9 +199,14 @@ class Histogram:
         `histograms_`: numpy.ndarray
             Array containing the histogram(s).
         """
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'histogram' function."
+            )
+
         return self.histograms_
 
-    def histogram_raw_counts(self):
+    def histogram_raw_counts(self) -> np.ndarray:
         """
         Get the raw bin counts of the histogram(s), even after the original
         histograms are scaled or averaged. If weights are used, then they
@@ -204,9 +217,14 @@ class Histogram:
         `histograms_raw_count_`: numpy.ndarray
             Array containing the raw counts of the histogram(s)
         """
+        if self.histograms_raw_count_ is None:
+            raise TypeError(
+                "'histograms_raw_count_' is None. It must be initialized before calling the 'histogram_raw_counts' function."
+            )
+
         return self.histograms_raw_count_
 
-    def number_of_histograms(self):
+    def number_of_histograms(self) -> int:
         """
         Get the number of current histograms.
 
@@ -217,7 +235,7 @@ class Histogram:
         """
         return self.number_of_histograms_
 
-    def bin_centers(self):
+    def bin_centers(self) -> np.ndarray:
         """
         Get the bin centers.
 
@@ -226,9 +244,14 @@ class Histogram:
         numpy.ndarray
             Array containing the bin centers.
         """
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'bin_centers' function."
+            )
+
         return (self.bin_edges_[:-1] + self.bin_edges_[1:]) / 2.0
 
-    def bin_width(self):
+    def bin_width(self) -> np.ndarray:
         """
         Get the bin widths.
 
@@ -237,9 +260,14 @@ class Histogram:
         numpy.ndarray
             Array containing the bin widths.
         """
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'bin_width' function."
+            )
+
         return self.bin_edges_[1:] - self.bin_edges_[:-1]
 
-    def bin_bounds_left(self):
+    def bin_bounds_left(self) -> np.ndarray:
         """
         Extract the lower bounds of the individual bins.
 
@@ -248,9 +276,14 @@ class Histogram:
         numpy.ndarray
             Array containing the lower bin boundaries.
         """
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'bin_bounds_left' function."
+            )
+
         return self.bin_edges_[:-1]
 
-    def bin_bounds_right(self):
+    def bin_bounds_right(self) -> np.ndarray:
         """
         Extract the upper bounds of the individual bins.
 
@@ -259,9 +292,14 @@ class Histogram:
         numpy.ndarray
             Array containing the upper bin boundaries.
         """
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'bin_bounds_right' function."
+            )
+
         return self.bin_edges_[1:]
 
-    def bin_boundaries(self):
+    def bin_boundaries(self) -> np.ndarray:
         """
         Get the bin boundaries.
 
@@ -270,9 +308,14 @@ class Histogram:
         numpy.ndarray
             Array containing the bin boundaries.
         """
-        return np.asarray(self.bin_edges_)
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'bin_boundaries' function."
+            )
 
-    def remove_bin(self, index):
+        return self.bin_edges_
+
+    def remove_bin(self, index: int) -> "Histogram":
         """
         Remove a bin from all histograms, starting from the 0th bin.
 
@@ -286,12 +329,36 @@ class Histogram:
         ValueError
             If `index` is out of range.
         """
+        if self.number_of_bins_ is None:
+            raise TypeError(
+                "'number_of_bins_' is None. It must be initialized before calling the'remove_bin' function."
+            )
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'remove_bin' function."
+            )
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'remove_bin' function."
+            )
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'remove_bin' function."
+            )
+        if self.histograms_raw_count_ is None:
+            raise TypeError(
+                "'histograms_raw_count_' is None. It must be initialized before calling the 'remove_bin' function."
+            )
+        if self.systematic_error_ is None:
+            raise TypeError(
+                "'systematic_error_' is None. It must be initialized before calling the 'remove_bin' function."
+            )
 
         if isinstance(index, (int)):
             if np.isnan(index):
                 raise ValueError("Bin number in remove_bin is NaN.")
 
-            if (index < 0 or index >= len(self.bin_edges_)):
+            if index < 0 or index >= len(self.bin_edges_):
                 raise ValueError("Bin number in remove_bin is out of range.")
         else:
             raise TypeError("Bin number in remove_bin must be an integer.")
@@ -299,20 +366,23 @@ class Histogram:
         self.number_of_bins_ -= 1
         self.bin_edges_ = np.delete(self.bin_edges_, index)
 
-        self.histograms_ = [np.delete(hist, index) for hist in self.histograms_]
-        self.error_ = [np.delete(err, index) for err in self.error_]
-        self.histograms_raw_count_ = [
-            np.delete(
-                raw_count,
-                index) for raw_count in self.histograms_raw_count_]
-        self.systematic_error_ = [
-            np.delete(
-                sys_err,
-                index) for sys_err in self.systematic_error_]
+        self.histograms_ = np.asarray(
+            [np.delete(hist, index) for hist in self.histograms_]
+        )
+        self.error_ = np.asarray([np.delete(err, index) for err in self.error_])
+        self.histograms_raw_count_ = np.asarray(
+            [
+                np.delete(raw_count, index)
+                for raw_count in self.histograms_raw_count_
+            ]
+        )
+        self.systematic_error_ = np.asarray(
+            [np.delete(sys_err, index) for sys_err in self.systematic_error_]
+        )
 
         return self
 
-    def add_bin(self, index, bin_edge):
+    def add_bin(self, index: int, bin_edge: float) -> "Histogram":
         """
         Add a bin to all histograms at the specified index.
         Attention: If values were added to bins before inserting a new bin, the information about its value is lost.
@@ -334,11 +404,35 @@ class Histogram:
         ValueError
             If `index` is out of range or the bin edges are not monotonically increasing.
         """
-
         if not isinstance(index, int):
             raise TypeError("Index in add_bin must be an integer.")
         if not isinstance(bin_edge, (int, float)):
             raise TypeError("Bin edge in add_bin must be a float or int.")
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'add_bin' function."
+            )
+        if self.number_of_bins_ is None:
+            raise TypeError(
+                "'number_of_bins_' is None. It must be initialized before calling the 'add_bin' function."
+            )
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'add_bin' function."
+            )
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'add_bin' function."
+            )
+        if self.histograms_raw_count_ is None:
+            raise TypeError(
+                "'histograms_raw_count_' is None. It must be initialized before calling the 'add_bin' function."
+            )
+        if self.systematic_error_ is None:
+            raise TypeError(
+                "'systematic_error_' is None. It must be initialized before calling the 'add_bin' function."
+            )
+
         if index < 0 or index >= len(self.bin_edges_):
             raise ValueError("Index in add_bin is out of range.")
         if index > 0 and bin_edge <= self.bin_edges_[index - 1]:
@@ -350,22 +444,36 @@ class Histogram:
         self.bin_edges_ = np.insert(self.bin_edges_, index, bin_edge)
 
         self.histograms_ = np.asarray(
-            [np.insert(hist, index, 0) for hist in self.histograms_])
-        self.error_ = np.asarray([np.insert(err, index, 0)
-                                 for err in self.error_])
-        self.histograms_raw_count_ = np.asarray([np.insert(
-            raw_count, index, 0).tolist() for raw_count in self.histograms_raw_count_])
+            [np.insert(hist, index, 0) for hist in self.histograms_]
+        )
+        self.error_ = np.asarray(
+            [np.insert(err, index, 0) for err in self.error_]
+        )
+        self.histograms_raw_count_ = np.asarray(
+            [
+                np.insert(raw_count, index, 0).tolist()
+                for raw_count in self.histograms_raw_count_
+            ]
+        )
         self.systematic_error_ = np.asarray(
-            [np.insert(sys_err, index, 0).tolist() for sys_err in self.systematic_error_])
+            [
+                np.insert(sys_err, index, 0).tolist()
+                for sys_err in self.systematic_error_
+            ]
+        )
 
         return self
 
-    def add_value(self, value, weight=None):
+    def add_value(
+        self,
+        value: Union[float, List[float], np.ndarray],
+        weight: Optional[Union[float, List[float], np.ndarray]] = None,
+    ) -> None:
         """
         Add value(s) to the latest histogram.
 
-        Different cases, if there is just one number added or a whole list/
-        array of numbers.
+        Different cases, if there is just one number added or a whole
+        list/array of numbers.
 
         Parameters
         ----------
@@ -380,27 +488,47 @@ class Histogram:
         TypeError
             if the input is not a number or numpy.ndarray or list
         ValueError
-            if an input `value` is `np.nan`
+            if an input :code:`value` is :code:`np.nan`
         ValueError
-            if the input `weight` has not the same dimension as `value`
+            if the input :code:`weight` has not the same dimension as :code:`value`
         ValueError
-            if a `weight` value is `np.nan`
+            if a :code:`weight` value is :code:`np.nan`
         """
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'add_value' function."
+            )
+        if self.number_of_bins_ is None:
+            raise TypeError(
+                "'number_of_bins_' is None. It must be initialized before calling the 'add_value' function."
+            )
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'add_value' function."
+            )
+        if self.histograms_raw_count_ is None:
+            raise TypeError(
+                "'histograms_raw_count_' is None. It must be initialized before calling the 'add_value' function."
+            )
+
         # Check if weight has the same length as value
         if weight is not None:
             if isinstance(weight, (int, float, np.number)):
                 if not isinstance(value, (int, float, np.number)):
                     raise ValueError(
-                        "Value must be numeric when weight is scalar.")
+                        "Value must be numeric when weight is scalar."
+                    )
                 if np.isnan(weight):
                     raise ValueError(
-                        "Value cannot be NaN when weight is scalar.")
+                        "Value cannot be NaN when weight is scalar."
+                    )
             elif len(weight) != np.atleast_1d(value).shape[0]:
                 raise ValueError("Weight must have the same length as value.")
             else:
                 if np.isnan(value).any():
                     raise ValueError(
-                        "Value cannot contain NaN when weight is not scalar.")
+                        "Value cannot contain NaN when weight is not scalar."
+                    )
 
         # Case 1.1: value is a single number
         if isinstance(value, (int, float, np.number)):
@@ -408,11 +536,17 @@ class Histogram:
                 raise ValueError("Input value in add_value is NaN.")
 
             counter_warnings = 0
-            if (value < self.bin_edges_[
-                    0] or value > self.bin_edges_[-1]) and counter_warnings == 0:
-                warn_msg = 'One or more values lie outside the histogram ' +\
-                    'range [' + str(self.bin_edges_[0]) + ',' + str(self.bin_edges_[-1]) +\
-                    ']. Exceeding values are ignored. Increase histogram range!'
+            if (
+                value < self.bin_edges_[0] or value > self.bin_edges_[-1]
+            ) and counter_warnings == 0:
+                warn_msg = (
+                    "One or more values lie outside the histogram "
+                    + "range ["
+                    + str(self.bin_edges_[0])
+                    + ","
+                    + str(self.bin_edges_[-1])
+                    + "]. Exceeding values are ignored. Increase histogram range!"
+                )
                 warnings.warn(warn_msg)
 
             bin_index = np.digitize(value, self.bin_edges_)
@@ -430,21 +564,28 @@ class Histogram:
         elif isinstance(value, (list, np.ndarray)):
             if np.isnan(value).any():
                 raise ValueError(
-                    "At least one input value in add_value is NaN.")
+                    "At least one input value in add_value is NaN."
+                )
             if weight is not None:
-                for element, w in zip(value, weight):
-                    self.add_value(element, weight=w)
+                if isinstance(weight, (int, float, np.number)):
+                    for element in value:
+                        self.add_value(element, weight=weight)
+                else:
+                    for element, w in zip(value, weight):
+                        self.add_value(element, weight=w)
             else:
                 for element in value:
                     self.add_value(element)
 
         # Case 1.3: value has an invalid input type
         else:
-            err_msg = 'Invalid input type! Input value must have one of the ' +\
-                      'following types: (int, float, np.number, list, np.ndarray)'
+            err_msg = (
+                "Invalid input type! Input value must have one of the "
+                + "following types: (int, float, np.number, list, np.ndarray)"
+            )
             raise TypeError(err_msg)
 
-    def make_density(self):
+    def make_density(self) -> None:
         """
         Make a probability density from the last histogram.
         The result represents the probability density function in a given bin.
@@ -458,6 +599,11 @@ class Histogram:
         ValueError
             if the integral over the histogram is zero
         """
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'make_density' function."
+            )
+
         if self.number_of_histograms_ == 0:
             raise ValueError("No histograms available to compute density.")
 
@@ -468,32 +614,60 @@ class Histogram:
         integral = np.sum(density * bin_widths)
         if integral == 0:
             raise ValueError("Integral over the histogram is zero.")
-        scale_factor = 1. / integral
+        scale_factor = 1.0 / integral
 
         self.statistical_error()
         self.scale_histogram(scale_factor)
 
-    def add_histogram(self):
+    def add_histogram(self) -> "Histogram":
         """
         Add a new histogram to the Histogram class instance.
 
         If new values are added to the histogram afterwards, these are added
         to the last histogram.
         """
+        if self.number_of_bins_ is None:
+            raise TypeError(
+                "'number_of_bins_' is None. It must be initialized before calling the 'add_histogram' function."
+            )
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'add_histogram' function."
+            )
+        if self.histograms_raw_count_ is None:
+            raise TypeError(
+                "'histograms_raw_count_' is None. It must be initialized before calling the 'add_histogram' function."
+            )
+        if self.scaling_ is None:
+            raise TypeError(
+                "'scaling_' is None. It must be initialized before calling the 'add_histogram' function."
+            )
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'add_histogram' function."
+            )
+        if self.systematic_error_ is None:
+            raise TypeError(
+                "'systematic_error_' is None. It must be initialized before calling the 'add_histogram' function."
+            )
+
         empty_histogram = np.zeros(self.number_of_bins_)
         self.histograms_ = np.vstack((self.histograms_, empty_histogram))
         self.histograms_raw_count_ = np.vstack(
-            (self.histograms_raw_count_, empty_histogram))
+            (self.histograms_raw_count_, empty_histogram)
+        )
         self.scaling_ = np.vstack(
-            (self.scaling_, np.ones(self.number_of_bins_)))
+            (self.scaling_, np.ones(self.number_of_bins_))
+        )
         self.error_ = np.vstack((self.error_, np.zeros(self.number_of_bins_)))
         self.systematic_error_ = np.vstack(
-            (self.systematic_error_, np.zeros(self.number_of_bins_)))
+            (self.systematic_error_, np.zeros(self.number_of_bins_))
+        )
         self.number_of_histograms_ += 1
 
         return self
 
-    def average(self):
+    def average(self) -> "Histogram":
         """
         Average over all histograms.
 
@@ -513,7 +687,7 @@ class Histogram:
         self.average_weighted(np.ones(self.number_of_histograms_))
         return self
 
-    def average_weighted(self, weights):
+    def average_weighted(self, weights: np.ndarray) -> "Histogram":
         """
         Weighted average over all histograms.
 
@@ -535,27 +709,47 @@ class Histogram:
             Returns a Histogram object.
 
         """
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+        if self.systematic_error_ is None:
+            raise TypeError(
+                "'systematic_error_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+        if self.histograms_raw_count_ is None:
+            raise TypeError(
+                "'histograms_raw_count_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+        if self.scaling_ is None:
+            raise TypeError(
+                "'scaling_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+
         average = np.average(self.histograms_, axis=0, weights=weights)
         variance = np.average(
-            (self.histograms_ - average)**2.,
-            axis=0,
-            weights=weights)
+            (self.histograms_ - average) ** 2.0, axis=0, weights=weights
+        )
+        # Ensure the result is a 2D array
+        if average.ndim == 1:
+            average = average.reshape(1, -1)
 
         self.histograms_ = average
-        self.error_ = np.sqrt(variance)
+        self.error_ = np.sqrt(variance).reshape(1, -1)
         self.systematic_error_ = np.sqrt(
-            np.average(
-                self.systematic_error_**2.,
-                axis=0,
-                weights=weights))
+            np.average(self.systematic_error_**2.0, axis=0, weights=weights)
+        )
         self.histogram_raw_count_ = np.sum(self.histograms_raw_count_, axis=0)
-        self.scaling_ = self.scaling_[0]
+        self.scaling_ = np.asarray(self.scaling_[0])
+
+        if self.scaling_.ndim == 1:
+            self.scaling_ = self.scaling_.reshape(1, -1)
 
         self.number_of_histograms_ = 1
 
         return self
 
-    def average_weighted_by_error(self):
+    def average_weighted_by_error(self) -> "Histogram":
         """
         Weighted average over all histograms, where each entry is weighted by its associated error.
 
@@ -576,28 +770,57 @@ class Histogram:
         TypeError
             if the error is zero for any entry.
         """
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+        if self.systematic_error_ is None:
+            raise TypeError(
+                "'systematic_error_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+        if self.histograms_raw_count_ is None:
+            raise TypeError(
+                "'histograms_raw_count_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+        if self.scaling_ is None:
+            raise TypeError(
+                "'scaling_' is None. It must be initialized before calling the 'average_weighted' function."
+            )
+
         if np.any(self.error_ == 0):
             raise TypeError(
-                "Error cannot be zero for any entry when averaging by error.")
+                "Error cannot be zero for any entry when averaging by error."
+            )
 
         weights = 1 / self.error_**2
         average = np.average(self.histograms_, axis=0, weights=weights)
 
+        # Ensure the result is a 2D array
+        if average.ndim == 1:
+            average = average.reshape(1, -1)
+
         self.histograms_ = average
-        self.error_ = np.sqrt(1. / np.sum(1. / np.square(self.error_), axis=0))
+        self.error_ = np.sqrt(
+            1.0 / np.sum(1.0 / np.square(self.error_), axis=0)
+        )
         self.systematic_error_ = np.sqrt(
-            np.average(
-                self.systematic_error_**2.,
-                axis=0,
-                weights=weights))
+            np.average(self.systematic_error_**2.0, axis=0, weights=weights)
+        )
         self.histogram_raw_count_ = np.sum(self.histograms_raw_count_, axis=0)
-        self.scaling_ = self.scaling_[0]
+        self.scaling_ = np.asarray(self.scaling_[0])
+
+        if self.scaling_.ndim == 1:
+            self.scaling_ = self.scaling_.reshape(1, -1)
 
         self.number_of_histograms_ = 1
 
         return self
 
-    def standard_error(self):
+    def standard_error(self) -> np.ndarray:
         """
         Get the standard deviation over all histogram counts for each bin.
 
@@ -606,9 +829,14 @@ class Histogram:
         numpy.ndarray
             Array containing the standard deviation for each bin.
         """
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'standard_error' function."
+            )
+
         return self.error_
 
-    def statistical_error(self):
+    def statistical_error(self) -> np.ndarray:
         """
         Compute the statistical error of all histogram bins for all histograms.
         This assumes Poisson distributed counts in each bin and independent draws.
@@ -619,13 +847,20 @@ class Histogram:
             2D Array containing the statistical error (standard deviation) for
             each bin and histogram.
         """
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'statistical_error' function."
+            )
+
         counter_histogram = 0
         for histogram in self.histogram():
             self.error_[counter_histogram] = np.sqrt(histogram)
             counter_histogram += 1
         return self.error_
 
-    def scale_histogram(self, value):
+    def scale_histogram(
+        self, value: Union[int, float, np.number, List[float], np.ndarray]
+    ) -> None:
         """
         Scale the latest histogram by a factor.
 
@@ -639,15 +874,37 @@ class Histogram:
         value: int, float, np.number, list, numpy.ndarray
             Scaling factor for the histogram.
         """
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'scale_histogram' function."
+            )
+        if self.scaling_ is None:
+            raise TypeError(
+                "'scaling_' is None. It must be initialized before calling the 'scale_histogram' function."
+            )
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'scale_histogram' function."
+            )
+
         if isinstance(value, (int, float, np.number)) and value < 0:
             raise ValueError(
-                "The scaling factor of the histogram cannot be negative")
-        elif isinstance(value, (list, np.ndarray)) and sum(1 for number in value if number < 0) > 0:
+                "The scaling factor of the histogram cannot be negative"
+            )
+        elif (
+            isinstance(value, (list, np.ndarray))
+            and sum(1 for number in value if number < 0) > 0
+        ):
             raise ValueError(
-                "The scaling factor of the histogram cannot be negative")
-        elif isinstance(value, (list, np.ndarray)) and len(value) != self.number_of_bins_:
+                "The scaling factor of the histogram cannot be negative"
+            )
+        elif (
+            isinstance(value, (list, np.ndarray))
+            and len(value) != self.number_of_bins_
+        ):
             raise ValueError(
-                "The length of list/array not compatible with number_of_bins_ of the histogram")
+                "The length of list/array not compatible with number_of_bins_ of the histogram"
+            )
 
         if isinstance(value, (int, float, np.number)):
             self.histograms_[-1] *= value
@@ -655,11 +912,16 @@ class Histogram:
             self.error_[-1] *= value
 
         elif isinstance(value, (list, np.ndarray)):
+            if np.asarray(value).shape != self.histograms_[-1].shape:
+                raise ValueError(
+                    "The shape of the scaling factor array is not compatible with the histogram shape"
+                )
+
             self.histograms_[-1] *= np.asarray(value)
             self.scaling_[-1] *= np.asarray(value)
-            self.scaling_[-1] *= np.asarray(value)
+            self.error_[-1] *= np.asarray(value)
 
-    def set_error(self, own_error):
+    def set_error(self, own_error: Union[List[float], np.ndarray]) -> None:
         """
         Sets the histogram error by hand. This is helpful for weighted
         histograms where the weight has also an uncertainty.
@@ -672,14 +934,24 @@ class Histogram:
         value: list, numpy.ndarray
             Values for the uncertainties of the individual bins.
         """
-        if len(own_error) != self.number_of_bins_ or\
-                not isinstance(own_error, (list, np.ndarray)):
-            error_message = "The input error has a different length than the"\
+        if len(own_error) != self.number_of_bins_ or not isinstance(
+            own_error, (list, np.ndarray)
+        ):
+            error_message = (
+                "The input error has a different length than the"
                 + " number of histogram bins or it is not a list/numpy.ndarray"
+            )
             raise ValueError(error_message)
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'set_error' function."
+            )
+
         self.error_[-1] = own_error
 
-    def set_systematic_error(self, own_error):
+    def set_systematic_error(
+        self, own_error: Union[List[float], np.ndarray]
+    ) -> None:
         """
         Sets the systematic histogram error of the last created histogram by hand.
 
@@ -688,25 +960,57 @@ class Histogram:
         value: list, numpy.ndarray
             Values for the systematic uncertainties of the individual bins.
         """
-        if len(own_error) != self.number_of_bins_ or\
-                not isinstance(own_error, (list, np.ndarray)):
-            error_message = "The input error has a different length than the"\
+        if len(own_error) != self.number_of_bins_ or not isinstance(
+            own_error, (list, np.ndarray)
+        ):
+            error_message = (
+                "The input error has a different length than the"
                 + " number of histogram bins or it is not a list/numpy.ndarray"
+            )
             raise ValueError(error_message)
+        if self.systematic_error_ is None:
+            raise TypeError(
+                "'systematic_error_' is None. It must be initialized before calling the 'set_systematic_error' function."
+            )
 
         self.systematic_error_[-1] = own_error
 
-    def print_histogram(self):
+    def print_histogram(self) -> None:
         """Print the histograms to the terminal."""
+        if self.number_of_histograms_ is None:
+            raise TypeError(
+                "'number_of_histograms_' is None. It must be initialized before calling the 'print_histogram' function."
+            )
+        if self.number_of_bins_ is None:
+            raise TypeError(
+                "'number_of_bins_' is None. It must be initialized before calling the 'print_histogram' function."
+            )
+        if self.bin_edges_ is None:
+            raise TypeError(
+                "'bin_edges_' is None. It must be initialized before calling the 'print_histogram' function."
+            )
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'print_histogram' function."
+            )
+
         print("bin_low,bin_high,bin_value")
         for hist in range(self.number_of_histograms_):
             print(f"{hist}. histogram:")
             for bin in range(self.number_of_bins_):
-                print(f'[{self.bin_edges_[bin]},{self.bin_edges_[bin+1]}):\
-                          {self.histograms_[hist][bin]}')
+                print(
+                    f"[{self.bin_edges_[bin]},{self.bin_edges_[bin+1]}):\
+                          {self.histograms_[hist][bin]}"
+                )
             print("")
 
-    def write_to_file(self, filename, hist_labels, comment='', columns=None):
+    def write_to_file(
+        self,
+        filename: str,
+        hist_labels: List[Dict[str, str]],
+        comment: str = "",
+        columns: Optional[List[str]] = None,
+    ) -> None:
         """
         Write multiple histograms to a CSV file along with their headers.
 
@@ -762,57 +1066,75 @@ class Histogram:
             is greater than 1 and the number of provided dictionaries is 1.
             Then the same dictionary is used for all histograms.
         """
-        if not isinstance(
-                hist_labels,
-                list) or not all(
-                isinstance(
-                hist_label,
-                dict) for hist_label in hist_labels):
+        if not isinstance(hist_labels, list) or not all(
+            isinstance(hist_label, dict) for hist_label in hist_labels
+        ):
             raise TypeError("hist_labels must be a list of dictionaries")
 
         if columns is not None and (
-            not isinstance(
-                columns,
-                list) or not all(
-                isinstance(
-                col,
-                str) for col in columns)):
+            not isinstance(columns, list)
+            or not all(isinstance(col, str) for col in columns)
+        ):
             raise TypeError("columns must be a list of strings")
+        if self.number_of_bins_ is None:
+            raise TypeError(
+                "'number_of_bins_' is None. It must be initialized before calling the 'write_to_file' function."
+            )
+        if self.histograms_ is None:
+            raise TypeError(
+                "'histograms_' is None. It must be initialized before calling the 'write_to_file' function."
+            )
+        if self.error_ is None:
+            raise TypeError(
+                "'error_' is None. It must be initialized before calling the 'write_to_file' function."
+            )
+        if self.systematic_error_ is None:
+            raise TypeError(
+                "'systematic_error_' is None. It must be initialized before calling the 'write_to_file' function."
+            )
 
         if columns is not None and not all(
-                col in hist_labels[0].keys() for col in columns):
+            col in hist_labels[0].keys() for col in columns
+        ):
             raise TypeError(
-                "columns must contain only keys present in hist_labels")
+                "columns must contain only keys present in hist_labels"
+            )
 
         if self.number_of_histograms_ > 1 and len(hist_labels) == 1:
-            error_message = "Print multiple histograms to file, only one header"\
-                            + " provided. Use the header for all histograms."
+            error_message = (
+                "Print multiple histograms to file, only one header"
+                + " provided. Use the header for all histograms."
+            )
             warnings.warn(error_message)
-        elif self.number_of_histograms_ > 1 and (len(hist_labels) > 1 and len(hist_labels) < self.number_of_histograms_):
+        elif self.number_of_histograms_ > 1 and (
+            len(hist_labels) > 1
+            and len(hist_labels) < self.number_of_histograms_
+        ):
             raise ValueError(
-                "Print multiple histograms to file, more than one," +
-                " but less than number of histograms headers provided.")
+                "Print multiple histograms to file, more than one,"
+                + " but less than number of histograms headers provided."
+            )
 
         if columns is None:
             columns = [
-                'bin_center',
-                'bin_low',
-                'bin_high',
-                'distribution',
-                'stat_err+',
-                'stat_err-',
-                'sys_err+',
-                'sys_err-']
+                "bin_center",
+                "bin_low",
+                "bin_high",
+                "distribution",
+                "stat_err+",
+                "stat_err-",
+                "sys_err+",
+                "sys_err-",
+            ]
 
-        with open(filename, 'w') as f:
+        with open(filename, "w") as f:
             writer = csv.writer(f)
-            if comment != '':
+            if comment != "":
                 f.write(comment)
-                f.write('\n')
+                f.write("\n")
 
             for idx in range(self.number_of_histograms_):
-                header = [hist_labels[0][col] if len(
-                    hist_labels) == 1 else hist_labels[idx][col] for col in columns]
+                header = [hist_labels[idx][col] for col in columns]
                 writer.writerow(header)
                 for i in range(self.number_of_bins_):
                     data = [
@@ -823,7 +1145,8 @@ class Histogram:
                         self.error_[idx][i],
                         self.error_[idx][i],
                         self.systematic_error_[idx][i],
-                        self.systematic_error_[idx][i]]
+                        self.systematic_error_[idx][i],
+                    ]
                     data = [data[columns.index(col)] for col in columns]
                     writer.writerow(data)
-                f.write('\n')
+                f.write("\n")
