@@ -67,12 +67,12 @@ class Jackknife:
     Raises
     ------
     ValueError
-        If delete_fraction is less than 0 or greater than or equal to 1.
-        If number_samples is less than 1.
+        If :code:`delete_fraction` is less than 0 or greater than or equal to 1.
+        If :code:`number_samples` is less than 1.
     TypeError
-        If delete_fraction is not a float.
-        If number_samples is not an integer.
-        If seed is not an integer.
+        If :code:`delete_fraction` is not a float.
+        If :code:`number_samples` is not an integer.
+        If :code:`seed` is not an integer.
 
     Examples
     --------
@@ -91,10 +91,8 @@ class Jackknife:
     """
 
     def __init__(
-            self,
-            delete_fraction: float,
-            number_samples: int,
-            seed: int = 42) -> None:
+        self, delete_fraction: float, number_samples: int, seed: int = 42
+    ) -> None:
 
         if not isinstance(delete_fraction, float):
             raise TypeError("delete_fraction must be a float.")
@@ -103,15 +101,21 @@ class Jackknife:
         if not isinstance(seed, int):
             raise TypeError("seed must be an integer.")
 
+        if not isinstance(delete_fraction, float):
+            raise TypeError("delete_fraction must be a float.")
+        if not isinstance(number_samples, int):
+            raise TypeError("number_samples must be an integer.")
+        if not isinstance(seed, int):
+            raise TypeError("seed must be an integer.")
         if delete_fraction < 0.0 or delete_fraction >= 1.0:
             raise ValueError("delete_fraction must be between 0 and 1.")
         if number_samples < 1:
             raise ValueError("number_samples must be greater than 0.")
 
-        self.delete_fraction: float = delete_fraction
-        self.number_samples: int = number_samples
+        self.delete_fraction = delete_fraction
+        self.number_samples = number_samples
 
-        self.seed: int = seed
+        self.seed = seed
         self._init_random()
 
     def _init_random(self) -> None:
@@ -136,17 +140,19 @@ class Jackknife:
             The remaining data after randomly deleting a fraction of it.
         """
         data = data.copy()
-        delete_indices = rd.sample(range(len(data)),
-                                   int(self.delete_fraction * len(data)))
+        delete_indices = rd.sample(
+            range(len(data)), int(self.delete_fraction * len(data))
+        )
         data = np.delete(data, delete_indices, axis=0)
         return data
 
-    def _apply_function_to_reduced_data(self,
-                                        reduced_data: np.ndarray,
-                                        function: Callable[...,
-                                                           Any],
-                                        *args,
-                                        **kwargs) -> Any:
+    def _apply_function_to_reduced_data(
+        self,
+        reduced_data: np.ndarray,
+        function: Callable[..., Any],
+        *args: tuple,
+        **kwargs: Any,
+    ) -> Any:
         """
         Apply a function to the reduced data.
 
@@ -170,7 +176,12 @@ class Jackknife:
         return function(reduced_data, *args, **kwargs)
 
     def _compute_one_jackknife_sample(
-            self, data: np.ndarray, function: Callable[..., Any], *args, **kwargs) -> Any:
+        self,
+        data: np.ndarray,
+        function: Callable[..., Any],
+        *args: tuple,
+        **kwargs: Any,
+    ) -> Any:
         """
         Compute one Jackknife sample.
 
@@ -193,16 +204,18 @@ class Jackknife:
         """
         reduced_data = self._randomly_delete_data(data)
         return self._apply_function_to_reduced_data(
-            reduced_data, function, *args, **kwargs)
+            reduced_data, function, *args, **kwargs
+        )
 
     @staticmethod
-    def _helper_unpack(instance,
-                       index: int,
-                       data: np.ndarray,
-                       function: Callable[...,
-                                          Any],
-                       args: tuple,
-                       kwargs: dict) -> Any:
+    def _helper_unpack(
+        instance: "Jackknife",
+        index: int,
+        data: np.ndarray,
+        function: Callable[..., Any],
+        args: tuple,
+        kwargs: Any,
+    ) -> Any:
         """
         Helper function to unpack the arguments for parallel processing.
 
@@ -229,7 +242,8 @@ class Jackknife:
         """
         rd.seed(instance.seed + index)
         return instance._compute_one_jackknife_sample(
-            data, function, *args, **kwargs)
+            data, function, *args, **kwargs
+        )
 
     def _init_random_subprocess(self, seed: int) -> None:
         """
@@ -246,13 +260,14 @@ class Jackknife:
         """
         rd.seed(seed)
 
-    def _compute_jackknife_samples(self,
-                                   data: np.ndarray,
-                                   function: Callable[...,
-                                                      Any],
-                                   num_cores: Optional[int] = None,
-                                   *args,
-                                   **kwargs) -> np.ndarray:
+    def _compute_jackknife_samples(
+        self,
+        data: np.ndarray,
+        function: Callable[..., Any],
+        num_cores: Optional[int] = None,
+        *args: tuple,
+        **kwargs: Any,
+    ) -> np.ndarray:
         """
         Compute all Jackknife samples.
 
@@ -286,25 +301,34 @@ class Jackknife:
         if num_cores is None:
             num_cores = os.cpu_count()
 
-        with Pool(num_cores, initializer=self._init_random_subprocess, initargs=(self.seed,)) as pool:
-            results = pool.starmap(self._helper_unpack,
-                                   [(self, index, data, function, args, kwargs)
-                                    for index in range(self.number_samples)]
-                                   )
+        with Pool(
+            num_cores,
+            initializer=self._init_random_subprocess,
+            initargs=(self.seed,),
+        ) as pool:
+            results = pool.starmap(
+                self._helper_unpack,
+                [
+                    (self, index, data, function, args, kwargs)
+                    for index in range(self.number_samples)
+                ],
+            )
         return np.array(results)
 
-    def compute_jackknife_estimates(self,
-                                    data: np.ndarray,
-                                    function: Callable[...,
-                                                       Any] = np.mean,
-                                    num_cores: Optional[int] = None,
-                                    *args,
-                                    **kwargs) -> float:
+    def compute_jackknife_estimates(
+        self,
+        data: np.ndarray,
+        function: Callable[..., Any] = np.mean,
+        num_cores: Optional[int] = None,
+        *args: tuple,
+        **kwargs: Any,
+    ) -> float:
         """
         Compute the Jackknife uncertainty estimates for a function applied to
-        a data array. The default function is np.mean, but it can be changed to
-        any other function that accepts a numpy array as input. Multiple
-        other arguments can be passed to the function as args and kwargs.
+        a data array. The default function is :code:`np.mean`, but it can be
+        changed to any other function that accepts a numpy array as input.
+        Multiple other arguments can be passed to the function as args and
+        kwargs.
 
         Parameters
         ----------
@@ -332,7 +356,7 @@ class Jackknife:
         Raises
         ------
         ValueError
-            If delete_n_points is less than 1 (delete_fraction too small).
+            If :code:`delete_n_points` is less than 1 (:code:`delete_fraction` too small).
         TypeError
             If data is not a numpy array or if function is not callable.
         TypeError
@@ -347,25 +371,30 @@ class Jackknife:
             raise TypeError("function must be a callable object.")
 
         # Check if the function returns a single number
-        test_result = function(data[:max(1, len(data) // 100)], *args, **kwargs)
+        test_result = function(
+            data[: max(1, len(data) // 100)], *args, **kwargs
+        )
         if not isinstance(test_result, (int, float)):
             raise TypeError("function must return a single number.")
 
         jackknife_samples = self._compute_jackknife_samples(
-            data, function, num_cores, *args, **kwargs)
+            data, function, num_cores, *args, **kwargs
+        )
 
         # mean of jackknife samples
         mean_samples = np.mean(jackknife_samples)
         # variance of the jackknife samples
         variance_samples = 0.0
         for i in range(len(jackknife_samples)):
-            variance_samples = (jackknife_samples[i] - mean_samples)**2.
+            variance_samples = (jackknife_samples[i] - mean_samples) ** 2.0
 
         if delete_n_points == 1:
-            variance_samples *= ((len(jackknife_samples) -
-                                 1) / len(jackknife_samples))
+            variance_samples *= (len(jackknife_samples) - 1) / len(
+                jackknife_samples
+            )
         else:
-            variance_samples *= (len(data) - delete_n_points) / \
-                (delete_n_points * len(jackknife_samples))
+            variance_samples *= (len(data) - delete_n_points) / (
+                delete_n_points * len(jackknife_samples)
+            )
 
         return np.sqrt(variance_samples)
