@@ -8,10 +8,11 @@
 # ===================================================
 
 import numpy as np
-import fastjet as fj
+import fastjet as fj  # type: ignore
 import csv
 import warnings
 from sparkx.Particle import Particle
+from typing import Optional, List, Tuple, Any
 
 
 class JetAnalysis:
@@ -34,7 +35,7 @@ class JetAnalysis:
         Jet radius parameter.
     jet_eta_range_: tuple
         Minimum and maximum pseudorapidity for jet selection.
-    jet_pt_range_: tuple
+    jet_pT_range_: tuple
         Minimum transverse momentum for jet selection and maximum transverse
         momentum to write out the jet.
     jet_data_: list
@@ -86,7 +87,7 @@ class JetAnalysis:
         >>> jet_analysis = JetAnalysis()
         >>>
         >>> # Perform the jet analysis
-        >>> jet_analysis.perform_jet_finding(hadron_data, jet_R=0.4, jet_eta_range=(-2., 2.), jet_pt_range=(10., None), output_filename=JET_ANALYSIS_OUTPUT_PATH)
+        >>> jet_analysis.perform_jet_finding(hadron_data, jet_R=0.4, jet_eta_range=(-2., 2.), jet_pT_range=(10., None), output_filename=JET_ANALYSIS_OUTPUT_PATH)
 
     If you want to analyze the jets further, you have to read in the jet data
     from the file like:
@@ -138,19 +139,20 @@ class JetAnalysis:
            <br />
     """
 
-    def __init__(self):
-        self.hadron_data_ = None
-        self.jet_R_ = None
-        self.jet_eta_range_ = None
-        self.jet_pt_range_ = None
-        self.jet_data_ = None
+    def __init__(self) -> None:
+        self.hadron_data_: Optional[List[List[Particle]]] = None
+        self.jet_R_: Optional[float] = None
+        self.jet_eta_range_: Optional[tuple] = None
+        self.jet_pT_range_: Optional[tuple] = None
+        self.jet_data_: Optional[List[List[Any]]] = None
 
     def __initialize_and_check_parameters(
-            self,
-            hadron_data,
-            jet_R,
-            jet_eta_range,
-            jet_pt_range):
+        self,
+        hadron_data: List[List[Particle]],
+        jet_R: float,
+        jet_eta_range: Tuple[Optional[float], Optional[float]],
+        jet_pT_range: Tuple[Optional[float], Optional[float]],
+    ) -> None:
         """
         Initialize and check the parameters for jet analysis.
 
@@ -166,7 +168,7 @@ class JetAnalysis:
             Minimum and maximum pseudorapidity for jet selection.
             :code:`None` values are allowed and are exchanged
             by :math:`-\\infty` or :math:`+\\infty` automatically.
-        jet_pt_range: tuple
+        jet_pT_range: tuple
             Minimum transverse momentum for jet finding algorithm and maximum
             transverse momentum to write out the jet to a file.
             Values can be :code:`None`, then the lower bound is set to zero
@@ -175,47 +177,68 @@ class JetAnalysis:
         self.hadron_data_ = hadron_data
 
         # check jet radius
-        if jet_R <= 0.:
+        if jet_R <= 0.0:
             raise ValueError("jet_R must be larger than 0")
         self.jet_R_ = jet_R
 
         # check jet eta range
         if not isinstance(jet_eta_range, tuple):
-            raise TypeError("jet_eta_range is not a tuple. " +
-                            "It must contain either values or None.")
+            raise TypeError(
+                "jet_eta_range is not a tuple. "
+                + "It must contain either values or None."
+            )
         if len(jet_eta_range) != 2:
             raise ValueError("jet_eta_range must contain exactly two values.")
 
-        lower_cut = float(
-            '-inf') if jet_eta_range[0] is None else jet_eta_range[0]
-        upper_cut = float(
-            'inf') if jet_eta_range[1] is None else jet_eta_range[1]
+        # check jet eta range
+        if not isinstance(jet_eta_range, tuple):
+            raise TypeError(
+                "jet_eta_range is not a tuple. "
+                + "It must contain either values or None."
+            )
+        if len(jet_eta_range) != 2:
+            raise ValueError("jet_eta_range must contain exactly two values.")
+
+        lower_cut = (
+            float("-inf") if jet_eta_range[0] is None else jet_eta_range[0]
+        )
+        upper_cut = (
+            float("inf") if jet_eta_range[1] is None else jet_eta_range[1]
+        )
         if lower_cut < upper_cut:
             self.jet_eta_range_ = (lower_cut, upper_cut)
         else:
             self.jet_eta_range_ = (upper_cut, lower_cut)
-            warnings.warn("The lower jet eta cut value is larger than the " +
-                          "upper one. They are interchanged automatically.")
+            warnings.warn(
+                "The lower jet eta cut value is larger than the "
+                + "upper one. They are interchanged automatically."
+            )
 
         # check the jet pt range
-        if not isinstance(jet_pt_range, tuple):
-            raise TypeError("jet_pt_range is not a tuple. " +
-                            "It must contain either values or None.")
-        if len(jet_pt_range) != 2:
-            raise ValueError("jet_pt_range must contain exactly two values.")
-        if any(pt is not None and pt < 0 for pt in jet_pt_range):
-            raise ValueError("All values in jet_pt_range must be non-negative.")
+        if not isinstance(jet_pT_range, tuple):
+            raise TypeError(
+                "jet_pT_range is not a tuple. "
+                + "It must contain either values or None."
+            )
+        if len(jet_pT_range) != 2:
+            raise ValueError("jet_pT_range must contain exactly two values.")
+        if any(pT is not None and pT < 0 for pT in jet_pT_range):
+            raise ValueError("All values in jet_pT_range must be non-negative.")
 
-        lower_cut = 0. if jet_pt_range[0] is None else jet_pt_range[0]
-        upper_cut = float('inf') if jet_pt_range[1] is None else jet_pt_range[1]
+        lower_cut = 0.0 if jet_pT_range[0] is None else jet_pT_range[0]
+        upper_cut = float("inf") if jet_pT_range[1] is None else jet_pT_range[1]
         if lower_cut < upper_cut:
-            self.jet_pt_range_ = (lower_cut, upper_cut)
+            self.jet_pT_range_ = (lower_cut, upper_cut)
         else:
-            self.jet_pt_range_ = (upper_cut, lower_cut)
-            warnings.warn("The lower jet pt cut value is larger than the " +
-                          "upper one. They are interchanged automatically.")
+            self.jet_pT_range_ = (upper_cut, lower_cut)
+            warnings.warn(
+                "The lower jet pT cut value is larger than the "
+                + "upper one. They are interchanged automatically."
+            )
 
-    def create_fastjet_PseudoJets(self, event_hadrons):
+    def create_fastjet_PseudoJets(
+        self, event_hadrons: List[Particle]
+    ) -> List[fj.PseudoJet]:
         """
         Convert hadron data to a list of fastjet.PseudoJet objects.
 
@@ -236,11 +259,12 @@ class JetAnalysis:
         return event_list_PseudoJets
 
     def fill_associated_particles(
-            self,
-            jet,
-            event,
-            status_selection,
-            only_charged):
+        self,
+        jet: fj.PseudoJet,
+        event: int,
+        status_selection: str,
+        only_charged: bool,
+    ) -> List[Particle]:
         """
         Select particles in the jet cone.
 
@@ -261,26 +285,39 @@ class JetAnalysis:
         list
             List of associated particles in the jet cone.
         """
+        if self.hadron_data_ is None:
+            raise TypeError(
+                "'hadron_data_' is None. It must be initialized before calling the 'hadron_data_' function."
+            )
+
+        if not (0 <= event < len(self.hadron_data_)):
+            raise IndexError(f"Event index {event} is out of range.")
+
         associated_hadrons = []
         for hadron in self.hadron_data_[event]:
             if np.isnan(hadron.status):
                 raise ValueError("Hadron status not set")
-            if (status_selection == 'negative' and hadron.status >= 0) or \
-                (status_selection == 'positive' and hadron.status < 0) or \
-                    (only_charged and hadron.charge == 0):
+            if (
+                (status_selection == "negative" and hadron.status >= 0)
+                or (status_selection == "positive" and hadron.status < 0)
+                or (only_charged and hadron.charge == 0)
+            ):
                 continue
 
-            fastjet_particle = fj.PseudoJet(hadron.px, hadron.py,
-                                            hadron.pz, hadron.E)
+            fastjet_particle = fj.PseudoJet(
+                hadron.px, hadron.py, hadron.pz, hadron.E
+            )
             delta_eta = fastjet_particle.eta() - jet.eta()
             delta_phi = fastjet_particle.delta_phi_to(jet)
-            delta_r = np.sqrt(delta_eta**2. + delta_phi**2.)
+            delta_r = np.sqrt(delta_eta**2.0 + delta_phi**2.0)
 
             if delta_r < self.jet_R_:
                 associated_hadrons.append(hadron)
         return associated_hadrons
 
-    def jet_hole_subtraction(self, jet, holes):
+    def jet_hole_subtraction(
+        self, jet: fj.PseudoJet, holes: List[Particle]
+    ) -> fj.PseudoJet:
         """
         Subtract energy-momentum contributions from holes in the jet.
 
@@ -318,8 +355,14 @@ class JetAnalysis:
         jet.reset(jet_px, jet_py, jet_pz, jet_E)
         return jet
 
-    def write_jet_output(self, output_filename, jet, associated_hadrons, event,
-                         new_file=False):
+    def write_jet_output(
+        self,
+        output_filename: str,
+        jet: fj.PseudoJet,
+        associated_hadrons: List[Particle],
+        event: int,
+        new_file: Optional[bool] = False,
+    ) -> bool:
         """
         Write the jet and associated hadron information to a CSV file.
 
@@ -342,41 +385,65 @@ class JetAnalysis:
         bool
             False if successful.
         """
+        if self.jet_pT_range_ is None:
+            raise TypeError(
+                "'jet_pT_range_' is None. It must be initialized before calling the 'write_jet_output' function."
+            )
+
         # jet data from reconstruction
         jet_status = 10
         jet_pid = 10
         output_list = []
 
-        if jet.perp() < self.jet_pt_range_[1]:
-            output_list = [[0, jet.perp(), jet.eta(), jet.phi(), jet_status,
-                            jet_pid, jet.e(), event]]
+        if jet.perp() < self.jet_pT_range_[1]:
+            output_list = [
+                [
+                    0,
+                    jet.perp(),
+                    jet.eta(),
+                    jet.phi(),
+                    jet_status,
+                    jet_pid,
+                    jet.e(),
+                    event,
+                ]
+            ]
 
             # associated hadrons
             for i, associated in enumerate(associated_hadrons, start=1):
-                pseudo_jet = fj.PseudoJet(associated.px, associated.py,
-                                          associated.pz, associated.E)
-                output = [i, pseudo_jet.perp(), pseudo_jet.eta(),
-                          pseudo_jet.phi(), associated.status,
-                          associated.pdg, associated.E, event]
+                pseudo_jet = fj.PseudoJet(
+                    associated.px, associated.py, associated.pz, associated.E
+                )
+                output = [
+                    i,
+                    pseudo_jet.perp(),
+                    pseudo_jet.eta(),
+                    pseudo_jet.phi(),
+                    associated.status,
+                    associated.pdg,
+                    associated.E,
+                    event,
+                ]
                 output_list.append(output)
 
-        mode = 'a' if not new_file else 'w'
+        mode = "a" if not new_file else "w"
 
-        with open(output_filename, mode, newline='') as f:
+        with open(output_filename, mode, newline="") as f:
             writer = csv.writer(f)
             writer.writerows(output_list)
 
         return False
 
     def perform_jet_finding(
-            self,
-            hadron_data,
-            jet_R,
-            jet_eta_range,
-            jet_pt_range,
-            output_filename,
-            assoc_only_charged=True,
-            jet_algorithm=fj.antikt_algorithm):
+        self,
+        hadron_data: List[List[Particle]],
+        jet_R: float,
+        jet_eta_range: Tuple[Optional[float], Optional[float]],
+        jet_pT_range: Tuple[Optional[float], Optional[float]],
+        output_filename: str,
+        assoc_only_charged: bool = True,
+        jet_algorithm: fj = fj.antikt_algorithm,
+    ) -> None:
         """
         Perform the jet analysis for multiple events. The function generates a
         file containing the jets consisting of a leading particle and associated
@@ -394,7 +461,7 @@ class JetAnalysis:
             Minimum and maximum pseudorapidity for jet selection.
             :code:`None` values are allowed and are exchanged
             by :math:`-\\infty` or :math:`+\\infty` automatically.
-        jet_pt_range: tuple
+        jet_pT_range: tuple
             Minimum transverse momentum for jet finding algorithm and maximum
             transverse momentum to write out the jet to a file.
             Values can be :code:`None`, then the lower bound is set to zero
@@ -415,17 +482,36 @@ class JetAnalysis:
 
         """
         self.__initialize_and_check_parameters(
-            hadron_data, jet_R, jet_eta_range, jet_pt_range)
+            hadron_data, jet_R, jet_eta_range, jet_pT_range
+        )
+        if self.hadron_data_ is None:
+            raise TypeError(
+                "'hadron_data_' is None. It must be initialized before calling the 'perform_jet_finding' function."
+            )
+        if self.jet_eta_range_ is None:
+            raise TypeError(
+                "'jet_eta_range_' is None. It must be initialized before calling the 'perform_jet_finding' function."
+            )
+        if self.jet_pT_range_ is None:
+            raise TypeError(
+                "'jet_pT_range_' is None. It must be initialized before calling the 'perform_jet_finding' function."
+            )
+
         for event, hadron_data_event in enumerate(self.hadron_data_):
             new_file = False
             event_PseudoJets = self.create_fastjet_PseudoJets(hadron_data_event)
-            if jet_algorithm == fj.ee_genkt_algorithm or jet_algorithm == fj.genkt_algorithm:
+            if (
+                jet_algorithm == fj.ee_genkt_algorithm
+                or jet_algorithm == fj.genkt_algorithm
+            ):
                 jet_definition = fj.JetDefinition(
-                    jet_algorithm, self.jet_R_, -1.0)
+                    jet_algorithm, self.jet_R_, -1.0
+                )
             else:
                 jet_definition = fj.JetDefinition(jet_algorithm, self.jet_R_)
             jet_selector = fj.SelectorEtaRange(
-                self.jet_eta_range_[0], self.jet_eta_range_[1])
+                self.jet_eta_range_[0], self.jet_eta_range_[1]
+            )
 
             if event == 0:
                 print("jet definition is:", jet_definition)
@@ -436,22 +522,30 @@ class JetAnalysis:
             # perform the jet finding algorithm
             cluster = fj.ClusterSequence(event_PseudoJets, jet_definition)
             jets = fj.sorted_by_pt(
-                cluster.inclusive_jets(
-                    self.jet_pt_range_[0]))
+                cluster.inclusive_jets(self.jet_pT_range_[0])
+            )
             jets = jet_selector(jets)
 
             # get the associated particles in the jet cone
             for jet in jets:
                 holes_in_jet = self.fill_associated_particles(
-                    jet, event, status_selection='negative', only_charged=assoc_only_charged)
+                    jet,
+                    event,
+                    status_selection="negative",
+                    only_charged=assoc_only_charged,
+                )
                 associated_particles = self.fill_associated_particles(
-                    jet, event, status_selection='positive', only_charged=assoc_only_charged)
+                    jet,
+                    event,
+                    status_selection="positive",
+                    only_charged=assoc_only_charged,
+                )
                 jet = self.jet_hole_subtraction(jet, holes_in_jet)
-                new_file = self.write_jet_output(output_filename, jet,
-                                                 associated_particles,
-                                                 event, new_file)
+                new_file = self.write_jet_output(
+                    output_filename, jet, associated_particles, event, new_file
+                )
 
-    def read_jet_data(self, input_filename):
+    def read_jet_data(self, input_filename: str) -> None:
         """
         Read the jet data from a CSV file and store it in the
         :code:`JetAnalysis` object.
@@ -462,22 +556,31 @@ class JetAnalysis:
             Filename of the CSV file containing the jet data.
         """
         jet_data = []
-        current_jet = []
-        with open(input_filename, 'r', newline='') as f:
+        current_jet: List[List[Any]] = []
+        with open(input_filename, "r", newline="") as f:
             reader = csv.reader(f)
             for row in reader:
                 jet_index = int(row[0])
                 if jet_index == 0 and current_jet:
                     jet_data.append(current_jet)
                     current_jet = []
-                current_jet.append([int(row[0]), float(row[1]), float(row[2]),
-                                    float(row[3]), int(row[4]), int(row[5]),
-                                    float(row[6]), int(row[7])])
+                current_jet.append(
+                    [
+                        int(row[0]),
+                        float(row[1]),
+                        float(row[2]),
+                        float(row[3]),
+                        int(row[4]),
+                        int(row[5]),
+                        float(row[6]),
+                        int(row[7]),
+                    ]
+                )
             if current_jet:
                 jet_data.append(current_jet)
         self.jet_data_ = jet_data
 
-    def get_jets(self):
+    def get_jets(self) -> List[List[float]]:
         """
         Get a list of jets from the jet data.
 
@@ -487,9 +590,14 @@ class JetAnalysis:
             List of jets. Contains all data of the jet output file rows in each
             element of the list (:code:`[jet][column]`).
         """
+        if self.jet_data_ is None:
+            raise TypeError(
+                "'jet_data_' is None. It must be initialized before calling this function."
+            )
+
         return [jet[0] for jet in self.jet_data_]
 
-    def get_associated_particles(self):
+    def get_associated_particles(self) -> List[List[List[float]]]:
         """
         Get a list of associated particles for all jets.
 
@@ -499,6 +607,11 @@ class JetAnalysis:
             List of associated particles for each jet in each element
             (:code:`[jet][associated_particle][column]`).
         """
+        if self.jet_data_ is None:
+            raise TypeError(
+                "'jet_data_' is None. It must be initialized before calling this function."
+            )
+
         associated_particles_list = []
         for jet in self.jet_data_:
             associated_particles = jet[1:]
