@@ -9,11 +9,12 @@
 
 import numpy as np
 import warnings
-from scipy.interpolate import interpn
-from scipy.stats import multivariate_normal
-
+from scipy.interpolate import interpn  # type: ignore
+from scipy.stats import multivariate_normal  # type: ignore
+from typing import Optional, List, Union, Tuple, Callable
+from sparkx.Particle import Particle
 import matplotlib.pyplot as plt
-from mpl_toolkits.mplot3d import Axes3D
+from mpl_toolkits.mplot3d import Axes3D  # type: ignore
 import matplotlib.cm as cm
 
 
@@ -145,19 +146,20 @@ class Lattice3D:
     """
 
     def __init__(
-            self,
-            x_min,
-            x_max,
-            y_min,
-            y_max,
-            z_min,
-            z_max,
-            num_points_x,
-            num_points_y,
-            num_points_z,
-            n_sigma_x=None,
-            n_sigma_y=None,
-            n_sigma_z=None):
+        self,
+        x_min: float,
+        x_max: float,
+        y_min: float,
+        y_max: float,
+        z_min: float,
+        z_max: float,
+        num_points_x: int,
+        num_points_y: int,
+        num_points_z: int,
+        n_sigma_x: Optional[float] = None,
+        n_sigma_y: Optional[float] = None,
+        n_sigma_z: Optional[float] = None,
+    ) -> None:
         self.x_min_ = x_min
         self.x_max_ = x_max
         self.y_min_ = y_min
@@ -167,30 +169,45 @@ class Lattice3D:
         self.num_points_x_ = num_points_x
         self.num_points_y_ = num_points_y
         self.num_points_z_ = num_points_z
-        self.cell_volume_ = abs((x_max - x_min) * (y_max - y_min) * \
-                                (z_max - z_min) / (num_points_x * num_points_y * num_points_z))
+        self.cell_volume_: float = abs(
+            (x_max - x_min)
+            * (y_max - y_min)
+            * (z_max - z_min)
+            / (num_points_x * num_points_y * num_points_z)
+        )
 
-        self.x_values_ = np.linspace(x_min, x_max, num_points_x)
-        self.y_values_ = np.linspace(y_min, y_max, num_points_y)
-        self.z_values_ = np.linspace(z_min, z_max, num_points_z)
+        self.x_values_: np.ndarray = np.linspace(x_min, x_max, num_points_x)
+        self.y_values_: np.ndarray = np.linspace(y_min, y_max, num_points_y)
+        self.z_values_: np.ndarray = np.linspace(z_min, z_max, num_points_z)
 
-        self.grid_ = np.zeros((num_points_x, num_points_y, num_points_z))
+        self.grid_: np.ndarray = np.zeros(
+            (num_points_x, num_points_y, num_points_z)
+        )
 
         self.n_sigma_x_ = float(n_sigma_x) if n_sigma_x is not None else 3
         self.n_sigma_y_ = float(n_sigma_y) if n_sigma_y is not None else 3
         self.n_sigma_z_ = float(n_sigma_z) if n_sigma_z is not None else 3
 
-        self.spacing_x_ = self.x_values_[
-            1] - self.x_values_[0] if num_points_x > 1 else None
-        self.spacing_y_ = self.y_values_[
-            1] - self.y_values_[0] if num_points_y > 1 else None
-        self.spacing_z_ = self.z_values_[
-            1] - self.z_values_[0] if num_points_z > 1 else None
-        self.density_x_ = (self.x_max_ - self.x_min_) / self.num_points_x_
-        self.density_y_ = (self.y_max_ - self.y_min_) / self.num_points_y_
-        self.density_z_ = (self.z_max_ - self.z_min_) / self.num_points_z_
+        self.spacing_x_: Optional[float] = (
+            self.x_values_[1] - self.x_values_[0] if num_points_x > 1 else None
+        )
+        self.spacing_y_: Optional[float] = (
+            self.y_values_[1] - self.y_values_[0] if num_points_y > 1 else None
+        )
+        self.spacing_z_: Optional[float] = (
+            self.z_values_[1] - self.z_values_[0] if num_points_z > 1 else None
+        )
+        self.density_x_: float = (
+            self.x_max_ - self.x_min_
+        ) / self.num_points_x_
+        self.density_y_: float = (
+            self.y_max_ - self.y_min_
+        ) / self.num_points_y_
+        self.density_z_: float = (
+            self.z_max_ - self.z_min_
+        ) / self.num_points_z_
 
-    def __is_valid_index(self, i, j, k):
+    def __is_valid_index(self, i: int, j: int, k: int) -> bool:
         """
         Check if the given indices (i, j, k) are valid within the defined bounds.
 
@@ -208,11 +225,15 @@ class Lattice3D:
         bool
             True if the indices are valid, False otherwise.
         """
-        return (0 <= i < self.num_points_x_) and \
-               (0 <= j < self.num_points_y_) and \
-               (0 <= k < self.num_points_z_)
+        return (
+            (0 <= i < self.num_points_x_)
+            and (0 <= j < self.num_points_y_)
+            and (0 <= k < self.num_points_z_)
+        )
 
-    def set_value_by_index(self, i, j, k, value):
+    def set_value_by_index(
+        self, i: int, j: int, k: int, value: Union[int, float]
+    ) -> None:
         """
         Set the value at the specified indices (i, j, k) in the grid.
 
@@ -236,7 +257,9 @@ class Lattice3D:
         else:
             self.grid_[i, j, k] = value
 
-    def get_value_by_index(self, i, j, k):
+    def get_value_by_index(
+        self, i: int, j: int, k: int
+    ) -> Optional[Union[int, float]]:
         """
         Get the value at the specified indices (i, j, k) in the grid.
 
@@ -260,7 +283,9 @@ class Lattice3D:
         else:
             return self.grid_[i, j, k]
 
-    def __get_index(self, value, values, num_points):
+    def __get_index(
+        self, value: Union[int, float], values: Union[np.ndarray, List[float]]
+    ) -> int:
         """
         Get the index corresponding to the given value within a specified range.
 
@@ -270,8 +295,6 @@ class Lattice3D:
             The value for which the index is to be determined.
         values : list or numpy.ndarray
             The list or array containing the range of values, sorted ascending.
-        num_points : int
-            The number of points in the range.
 
         Returns
         -------
@@ -286,13 +309,15 @@ class Lattice3D:
         if value < values[0] or value > values[-1]:
             raise ValueError("Value is outside the specified range.")
 
-        index = np.searchsorted(values, value, side='right')
+        index = np.searchsorted(values, value, side="right")
         if index == 0:
             index += 1
 
-        return index - 1
+        return int(index - 1)
 
-    def __get_index_nearest_neighbor(self, value, values):
+    def __get_index_nearest_neighbor(
+        self, value: Union[int, float], values: Union[np.ndarray, List[float]]
+    ) -> int:
         """
         Get the index corresponding to the nearest neighbor of a given value
         within a specified range.
@@ -322,7 +347,9 @@ class Lattice3D:
 
         return index
 
-    def __get_indices(self, x, y, z):
+    def __get_indices(
+        self, x: Union[int, float], y: Union[int, float], z: Union[int, float]
+    ) -> Tuple[int, int, int]:
         """
         Get the indices corresponding to the given coordinates within the lattice.
 
@@ -346,12 +373,14 @@ class Lattice3D:
         ValueError
             If any of the coordinates are outside the specified ranges.
         """
-        i = self.__get_index(x, self.x_values_, self.num_points_x_)
-        j = self.__get_index(y, self.y_values_, self.num_points_y_)
-        k = self.__get_index(z, self.z_values_, self.num_points_z_)
+        i = self.__get_index(x, self.x_values_)
+        j = self.__get_index(y, self.y_values_)
+        k = self.__get_index(z, self.z_values_)
         return i, j, k
 
-    def __get_indices_nearest_neighbor(self, x, y, z):
+    def __get_indices_nearest_neighbor(
+        self, x: Union[int, float], y: Union[int, float], z: Union[int, float]
+    ) -> Tuple[int, int, int]:
         """
         Get the indices corresponding to the nearest neighbor
         given coordinates within the lattice.
@@ -381,7 +410,13 @@ class Lattice3D:
         k = self.__get_index_nearest_neighbor(z, self.z_values_)
         return i, j, k
 
-    def set_value(self, x, y, z, value):
+    def set_value(
+        self,
+        x: Union[int, float],
+        y: Union[int, float],
+        z: Union[int, float],
+        value: Union[int, float],
+    ) -> None:
         """
         Set the value at the specified coordinates within the lattice.
 
@@ -404,7 +439,13 @@ class Lattice3D:
         i, j, k = self.__get_indices(x, y, z)
         self.set_value_by_index(i, j, k, value)
 
-    def set_value_nearest_neighbor(self, x, y, z, value):
+    def set_value_nearest_neighbor(
+        self,
+        x: Union[int, float],
+        y: Union[int, float],
+        z: Union[int, float],
+        value: Union[int, float],
+    ) -> None:
         """
         Set the value at the nearest neighbor of the
         specified coordinates within the lattice.
@@ -428,7 +469,9 @@ class Lattice3D:
         i, j, k = self.__get_indices_nearest_neighbor(x, y, z)
         self.set_value_by_index(i, j, k, value)
 
-    def get_value(self, x, y, z):
+    def get_value(
+        self, x: Union[int, float], y: Union[int, float], z: Union[int, float]
+    ) -> Optional[Union[int, float]]:
         """
         Get the value at the specified coordinates within the lattice.
 
@@ -454,7 +497,9 @@ class Lattice3D:
         i, j, k = self.__get_indices(x, y, z)
         return self.get_value_by_index(i, j, k)
 
-    def get_value_nearest_neighbor(self, x, y, z):
+    def get_value_nearest_neighbor(
+        self, x: Union[int, float], y: Union[int, float], z: Union[int, float]
+    ) -> Optional[Union[int, float]]:
         """
         Get the value of the nearest neighbor at the specified coordinates
         within the lattice.
@@ -481,7 +526,9 @@ class Lattice3D:
         i, j, k = self.__get_indices_nearest_neighbor(x, y, z)
         return self.get_value_by_index(i, j, k)
 
-    def __get_value(self, index, values, num_points):
+    def __get_value(
+        self, index: int, values: np.ndarray, num_points: int
+    ) -> float:
         """
         Retrieve the value associated with the given index.
 
@@ -509,7 +556,9 @@ class Lattice3D:
             raise ValueError("Index is outside the specified range.")
         return values[index]
 
-    def get_coordinates(self, i, j, k):
+    def get_coordinates(
+        self, i: int, j: int, k: int
+    ) -> Tuple[float, float, float]:
         """
         Retrieve the coordinates associated with the given indices.
 
@@ -539,7 +588,9 @@ class Lattice3D:
         z = self.__get_value(k, self.z_values_, self.num_points_z_)
         return x, y, z
 
-    def __find_closest_index(self, value, values):
+    def __find_closest_index(
+        self, value: float, values: Union[np.ndarray, List[float]]
+    ) -> int:
         """
         Find the index of the closest value to the given value in a list of values.
 
@@ -556,10 +607,13 @@ class Lattice3D:
             The index of the closest value in the list of values.
 
         """
-        index = np.argmin(np.abs(values - value))
-        return index
+        if isinstance(values, list):
+            values = np.array(values, dtype=float)
 
-    def __is_within_range(self, x, y, z):
+        index = np.argmin(np.abs(values - value))
+        return int(index)
+
+    def __is_within_range(self, x: float, y: float, z: float) -> bool:
         """
         Check if the given coordinates are within the defined range.
 
@@ -578,11 +632,15 @@ class Lattice3D:
             True if the coordinates are within the defined range, False otherwise.
 
         """
-        return (self.x_min_ <= x <= self.x_max_) and \
-               (self.y_min_ <= y <= self.y_max_) and \
-               (self.z_min_ <= z <= self.z_max_)
+        return (
+            (self.x_min_ <= x <= self.x_max_)
+            and (self.y_min_ <= y <= self.y_max_)
+            and (self.z_min_ <= z <= self.z_max_)
+        )
 
-    def find_closest_indices(self, x, y, z):
+    def find_closest_indices(
+        self, x: float, y: float, z: float
+    ) -> Tuple[int, int, int]:
         """
         Find the closest indices in the lattice corresponding to the given coordinates.
 
@@ -610,7 +668,9 @@ class Lattice3D:
         k = self.__find_closest_index(z, self.z_values_)
         return i, j, k
 
-    def interpolate_value(self, x, y, z, method='nearest'):
+    def interpolate_value(
+        self, x: float, y: float, z: float, method: str = "nearest"
+    ) -> float:
         """
         Interpolate the value at the specified position using (up to) trilinear
         interpolation.
@@ -634,20 +694,22 @@ class Lattice3D:
 
         """
         if not self.__is_within_range(x, y, z):
-            warnings.warn("Provided position is outside the lattice range.")
-            return None
+            raise TypeError("Provided position is outside the lattice range.")
 
         # Perform interpolation
         xi = [x, y, z]
         return interpn(
-            (self.x_values_,
-             self.y_values_,
-             self.z_values_),
+            (self.x_values_, self.y_values_, self.z_values_),
             self.grid_,
             xi,
-            method=method)[0]
+            method=method,
+        )[0]
 
-    def __operate_on_lattice(self, other, operation):
+    def __operate_on_lattice(
+        self,
+        other: "Lattice3D",
+        operation: Callable[[np.ndarray, np.ndarray], np.ndarray],
+    ) -> "Lattice3D":
         """
         Apply a binary operation on two Lattice3D objects element-wise.
 
@@ -674,7 +736,8 @@ class Lattice3D:
         """
         if not isinstance(other, Lattice3D):
             raise TypeError(
-                "Unsupported operand type. The operand must be of type 'Lattice3D'.")
+                "Unsupported operand type. The operand must be of type 'Lattice3D'."
+            )
 
         if self.grid_.shape != other.grid_.shape:
             raise ValueError("The lattices must have the same shape.")
@@ -688,13 +751,14 @@ class Lattice3D:
             self.z_max_,
             self.num_points_x_,
             self.num_points_y_,
-            self.num_points_z_)
+            self.num_points_z_,
+        )
 
         result.grid_ = operation(self.grid_, other.grid_)
 
         return result
 
-    def __add__(self, other):
+    def __add__(self, other: "Lattice3D") -> "Lattice3D":
         """
         Add two Lattice3D objects element-wise.
 
@@ -711,7 +775,7 @@ class Lattice3D:
         """
         return self.__operate_on_lattice(other, lambda x, y: x + y)
 
-    def __sub__(self, other):
+    def __sub__(self, other: "Lattice3D") -> "Lattice3D":
         """
         Subtract two Lattice3D objects element-wise.
 
@@ -728,7 +792,7 @@ class Lattice3D:
         """
         return self.__operate_on_lattice(other, lambda x, y: x - y)
 
-    def __mul__(self, other):
+    def __mul__(self, other: "Lattice3D") -> "Lattice3D":
         """
         Multiply two Lattice3D objects element-wise.
 
@@ -745,7 +809,7 @@ class Lattice3D:
         """
         return self.__operate_on_lattice(other, lambda x, y: x * y)
 
-    def __truediv__(self, other):
+    def __truediv__(self, other: "Lattice3D") -> "Lattice3D":
         """
         Divide two Lattice3D objects element-wise.
 
@@ -767,7 +831,7 @@ class Lattice3D:
         """
         return self.__operate_on_lattice(other, lambda x, y: x / y)
 
-    def average(self, *lattices):
+    def average(self, *lattices: "Lattice3D") -> "Lattice3D":
         """
         Compute the average of multiple Lattice3D objects element-wise.
 
@@ -814,7 +878,8 @@ class Lattice3D:
         for lattice in all_lattices:
             if not isinstance(lattice, Lattice3D):
                 raise TypeError(
-                    "Unsupported operand type. All operands must be of type 'Lattice3D'.")
+                    "Unsupported operand type. All operands must be of type 'Lattice3D'."
+                )
 
             if self.grid_.shape != lattice.grid_.shape:
                 raise ValueError("The lattices must have the same shape.")
@@ -828,14 +893,16 @@ class Lattice3D:
             self.z_max_,
             self.num_points_x_,
             self.num_points_y_,
-            self.num_points_z_)
+            self.num_points_z_,
+        )
 
         result.grid_ = np.mean(
-            [lattice.grid_ for lattice in all_lattices], axis=0)
+            [lattice.grid_ for lattice in all_lattices], axis=0
+        )
 
         return result
 
-    def rescale(self, factor):
+    def rescale(self, factor: float) -> None:
         """
         Rescale the values of the lattice by a specified factor.
 
@@ -851,7 +918,7 @@ class Lattice3D:
         """
         self.grid_ *= factor
 
-    def save_to_csv(self, filename):
+    def save_to_csv(self, filename: str) -> None:
         """
         Save the lattice data, including metadata, to a CSV file.
 
@@ -865,15 +932,19 @@ class Lattice3D:
         None
 
         """
-        metadata = np.array([self.x_min_,
-                             self.x_max_,
-                             self.y_min_,
-                             self.y_max_,
-                             self.z_min_,
-                             self.z_max_,
-                             self.num_points_x_,
-                             self.num_points_y_,
-                             self.num_points_z_])
+        metadata = np.array(
+            [
+                self.x_min_,
+                self.x_max_,
+                self.y_min_,
+                self.y_max_,
+                self.z_min_,
+                self.z_max_,
+                self.num_points_x_,
+                self.num_points_y_,
+                self.num_points_z_,
+            ]
+        )
 
         # Reshape metadata to have 1 row and as many columns as necessary
         metadata = metadata.reshape(1, -1)
@@ -883,10 +954,10 @@ class Lattice3D:
         grid_flattened = self.grid_.flatten().reshape(1, -1)
 
         data = np.hstack((metadata, grid_flattened))
-        np.savetxt(filename, data, delimiter=',')
+        np.savetxt(filename, data, delimiter=",")
 
     @classmethod
-    def load_from_csv(cls, filename):
+    def load_from_csv(cls, filename: str) -> "Lattice3D":
         """
         Load lattice data, including metadata, from a CSV file.
 
@@ -901,9 +972,19 @@ class Lattice3D:
             The loaded Lattice3D object containing the data and metadata.
 
         """
-        data = np.loadtxt(filename, delimiter=',')
+        data = np.loadtxt(filename, delimiter=",")
         metadata = data[0:9]
-        x_min, x_max, y_min, y_max, z_min, z_max, num_points_x, num_points_y, num_points_z = metadata
+        (
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            num_points_x,
+            num_points_y,
+            num_points_z,
+        ) = metadata
 
         lattice = Lattice3D(
             x_min,
@@ -914,14 +995,15 @@ class Lattice3D:
             z_max,
             int(num_points_x),
             int(num_points_y),
-            int(num_points_z))
+            int(num_points_z),
+        )
 
         grid_data = data[9:]
         lattice.grid_ = grid_data.reshape(lattice.grid_.shape)
 
         return lattice
 
-    def visualize(self):
+    def visualize(self) -> None:
         """
         Visualize the lattice data in a 3D plot.
 
@@ -939,29 +1021,31 @@ class Lattice3D:
 
         # Create a custom colormap where 0 values are white
         cmap = cm.get_cmap("PiYG").copy()
-        cmap.set_bad(color='white')
+        cmap.set_bad(color="white")
 
         # Create a 3D plot
         fig = plt.figure()
-        ax = fig.add_subplot(111, projection='3d')
+        ax: Axes3D = fig.add_subplot(111, projection="3d")
 
         # Scatter plot the lattice points
         scatter = ax.scatter(X_flat, Y_flat, Z_flat, c=values_flat, cmap=cmap)
 
         # Set plot labels
-        ax.set_xlabel('X')
-        ax.set_ylabel('Y')
-        ax.set_zlabel('Z')
-        ax.set_title('Lattice Data Visualization')
+        ax.set_xlabel("X")
+        ax.set_ylabel("Y")
+        ax.set_zlabel("Z")
+        ax.set_title("Lattice Data Visualization")
 
         # Create a color bar
         cbar = fig.colorbar(scatter)
-        cbar.set_label('Values')
+        cbar.set_label("Values")
 
         # Show the plot
         plt.show()
 
-    def extract_slice(self, axis, index):
+    def extract_slice(
+        self, axis: str, index: int
+    ) -> Tuple[np.ndarray, Tuple[np.ndarray, np.ndarray], str]:
         """
         Extract a 2D slice from the lattice along the specified axis at the
         given index.
@@ -989,33 +1073,33 @@ class Lattice3D:
             If the `axis` parameter is invalid or the `index` is out of range.
 
         """
-        if axis == 'x':
+        if axis == "x":
             if index < 0 or index >= self.num_points_x_:
                 raise ValueError("Invalid index for the X-axis.")
 
             slice_data = self.grid_[index, :, :]
             slice_values = (self.y_values_, self.z_values_)
-            slice_label = 'Y-Z Plane at X = {}'.format(self.x_values_[index])
-        elif axis == 'y':
+            slice_label = "Y-Z Plane at X = {}".format(self.x_values_[index])
+        elif axis == "y":
             if index < 0 or index >= self.num_points_y_:
                 raise ValueError("Invalid index for the Y-axis.")
 
             slice_data = self.grid_[:, index, :]
             slice_values = (self.x_values_, self.z_values_)
-            slice_label = 'X-Z Plane at Y = {}'.format(self.y_values_[index])
-        elif axis == 'z':
+            slice_label = "X-Z Plane at Y = {}".format(self.y_values_[index])
+        elif axis == "z":
             if index < 0 or index >= self.num_points_z_:
                 raise ValueError("Invalid index for the Z-axis.")
 
             slice_data = self.grid_[:, :, index]
             slice_values = (self.x_values_, self.y_values_)
-            slice_label = 'X-Y Plane at Z = {}'.format(self.z_values_[index])
+            slice_label = "X-Y Plane at Z = {}".format(self.z_values_[index])
         else:
             raise ValueError("Invalid axis. Must be 'x', 'y', or 'z'.")
 
         return slice_data, slice_values, slice_label
 
-    def save_slice_to_csv(self, axis, index, filename):
+    def save_slice_to_csv(self, axis: str, index: int, filename: str) -> None:
         """
         Save a 2D slice from the lattice along the specified axis and index to
         a CSV file.
@@ -1035,57 +1119,47 @@ class Lattice3D:
             If the `axis` parameter is invalid or the `index` is out of range.
 
         """
-        if axis == 'x':
+        if axis == "x":
             if index < 0 or index >= self.num_points_x_:
                 raise ValueError("Invalid index for the X-axis.")
 
             slice_data = self.grid_[index, :, :]
             header = "#Slice in Y-Z Plane at X = {}. Ymin: {} Ymax: {} ".format(
-                self.x_values_[index],
-                self.y_min_,
-                self.y_max_) + "ny: {} Zmin: {} Zmax: {} nz:{} ".format(
-                self.num_points_y_,
-                self.z_min_,
-                self.z_max_,
-                self.num_points_z_)
+                self.x_values_[index], self.y_min_, self.y_max_
+            ) + "ny: {} Zmin: {} Zmax: {} nz:{} ".format(
+                self.num_points_y_, self.z_min_, self.z_max_, self.num_points_z_
+            )
 
-        elif axis == 'y':
+        elif axis == "y":
             if index < 0 or index >= self.num_points_y_:
                 raise ValueError("Invalid index for the Y-axis.")
 
             slice_data = self.grid_[:, index, :]
             header = "#Slice in X-Z Plane at Y = {}. Xmin: {} Xmax: {} ".format(
-                self.y_values_[index],
-                self.x_min_,
-                self.x_max_) + "nx: {} Zmin: {} Zmax: {} nz: {} ".format(
-                self.num_points_x_,
-                self.z_min_,
-                self.z_max_,
-                self.num_points_z_)
-        elif axis == 'z':
+                self.y_values_[index], self.x_min_, self.x_max_
+            ) + "nx: {} Zmin: {} Zmax: {} nz: {} ".format(
+                self.num_points_x_, self.z_min_, self.z_max_, self.num_points_z_
+            )
+        elif axis == "z":
             if index < 0 or index >= self.num_points_z_:
                 raise ValueError("Invalid index for the Z-axis.")
 
             slice_data = self.grid_[:, :, index]
             header = "#Slice in X-Y Plane at Z = {}. Xmin: {} Xmax: {} ".format(
-                self.z_values_[index],
-                self.x_min_,
-                self.x_max_) + "nx: {} Ymin: {} Ymax: {} ny: {} ".format(
-                self.num_points_x_,
-                self.y_min_,
-                self.y_max_,
-                self.num_points_y_)
+                self.z_values_[index], self.x_min_, self.x_max_
+            ) + "nx: {} Ymin: {} Ymax: {} ny: {} ".format(
+                self.num_points_x_, self.y_min_, self.y_max_, self.num_points_y_
+            )
         else:
             raise ValueError("Invalid axis. Must be 'x', 'y', or 'z'.")
 
         np.savetxt(
-            filename,
-            slice_data,
-            delimiter=',',
-            header=header,
-            comments='')
+            filename, slice_data, delimiter=",", header=header, comments=""
+        )
 
-    def interpolate_to_lattice(self, num_points_x, num_points_y, num_points_z):
+    def interpolate_to_lattice(
+        self, num_points_x: int, num_points_y: int, num_points_z: int
+    ) -> "Lattice3D":
         """
         Interpolate the current lattice data to a new lattice with the specified
         number of points along each axis.
@@ -1122,7 +1196,8 @@ class Lattice3D:
             self.z_max_,
             num_points_x,
             num_points_y,
-            num_points_z)
+            num_points_z,
+        )
 
         # Generate the new grid coordinates
         x_new = np.linspace(self.x_min_, self.x_max_, num_points_x)
@@ -1139,16 +1214,17 @@ class Lattice3D:
         return new_lattice
 
     def interpolate_to_lattice_new_extent(
-            self,
-            num_points_x,
-            num_points_y,
-            num_points_z,
-            x_min,
-            x_max,
-            y_min,
-            y_max,
-            z_min,
-            z_max):
+        self,
+        num_points_x: int,
+        num_points_y: int,
+        num_points_z: int,
+        x_min: float,
+        x_max: float,
+        y_min: float,
+        y_max: float,
+        z_min: float,
+        z_max: float,
+    ) -> "Lattice3D":
         """
         Interpolate the current lattice data to a new lattice with the specified
         number of points and extent.
@@ -1189,8 +1265,17 @@ class Lattice3D:
         """
         # Create a new Lattice3D object with the desired number of points and
         # extent
-        new_lattice = Lattice3D(x_min, x_max, y_min, y_max, z_min, z_max,
-                                num_points_x, num_points_y, num_points_z)
+        new_lattice = Lattice3D(
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            num_points_x,
+            num_points_y,
+            num_points_z,
+        )
 
         # Generate the new grid coordinates
         x_new = np.linspace(x_min, x_max, num_points_x)
@@ -1201,12 +1286,22 @@ class Lattice3D:
         for i, x in enumerate(x_new):
             for j, y in enumerate(y_new):
                 for k, z in enumerate(z_new):
-                    value = self.interpolate_value(x, y, z)
+                    if (
+                        x < self.x_min_
+                        or x > self.x_max_
+                        or y < self.y_min_
+                        or y > self.y_max_
+                        or z < self.z_min_
+                        or z > self.z_max_
+                    ):
+                        value = 0.0
+                    else:
+                        value = self.interpolate_value(x, y, z)
                     new_lattice.set_value_by_index(i, j, k, value)
 
         return new_lattice
 
-    def reset(self):
+    def reset(self) -> None:
         """
         Reset the values of all grid points in the lattice to zero.
 
@@ -1218,7 +1313,13 @@ class Lattice3D:
         for i, j, k in np.ndindex(self.grid_.shape):
             self.grid_[i, j, k] = 0
 
-    def add_same_spaced_grid(self, other, center_x, center_y, center_z):
+    def add_same_spaced_grid(
+        self,
+        other: "Lattice3D",
+        center_x: float,
+        center_y: float,
+        center_z: float,
+    ) -> None:
         """
         Add the values of grid points of another lattice with same spacing.
 
@@ -1243,45 +1344,65 @@ class Lattice3D:
         """
         if not isinstance(other, Lattice3D):
             raise TypeError(
-                "Unsupported operand type. The operand must be of type 'Lattice3D'.")
+                "Unsupported operand type. The operand must be of type 'Lattice3D'."
+            )
+        if (
+            self.spacing_x_ is None
+            or self.spacing_y_ is None
+            or self.spacing_z_ is None
+        ):
+            raise TypeError("At least one of the 'spacing' is None.")
         # Check if both lattices have the same spacing
-        if ((other.spacing_x_ is None or abs(self.spacing_x_ -
-                                             other.spacing_x_) < 1e-3) and (other.spacing_y_ is None or abs(self.spacing_y_ -
-                                                                                                            other.spacing_y_) < 1e-3) and (other.spacing_z_ is None or abs(self.spacing_z_ -
-                                                                                                                                                                           other.spacing_z_) < 1e-3)):
+        if (
+            (
+                other.spacing_x_ is None
+                or abs(self.spacing_x_ - other.spacing_x_) < 1e-3
+            )
+            and (
+                other.spacing_y_ is None
+                or abs(self.spacing_y_ - other.spacing_y_) < 1e-3
+            )
+            and (
+                other.spacing_z_ is None
+                or abs(self.spacing_z_ - other.spacing_z_) < 1e-3
+            )
+        ):
             for i, j, k in np.ndindex(other.grid_.shape):
                 posx, posy, posz = other.get_coordinates(i, j, k)
                 posx = posx + center_x
                 posy = posy + center_y
                 posz = posz + center_z
-                if (posx < self.x_min_ or posx > self.x_max_ or posy < self.y_min_ or posy >
-                        self.y_max_ or posz < self.z_min_ or posz > self.z_max_):
+                if (
+                    posx < self.x_min_
+                    or posx > self.x_max_
+                    or posy < self.y_min_
+                    or posy > self.y_max_
+                    or posz < self.z_min_
+                    or posz > self.z_max_
+                ):
                     continue
                 else:
                     self.set_value_nearest_neighbor(
                         posx,
                         posy,
                         posz,
-                        self.get_value_nearest_neighbor(
-                            posx,
-                            posy,
-                            posz) +
-                        other.grid_[
-                            i,
-                            j,
-                            k])
+                        self.get_value_nearest_neighbor(posx, posy, posz)
+                        + other.grid_[i, j, k],
+                    )
 
         else:
             raise ValueError(
-                "The provided lattices do not have identical spacing.")
+                "The provided lattices do not have identical spacing."
+            )
 
     def add_particle_data(
-            self,
-            particle_data,
-            sigma,
-            quantity,
-            kernel="covariant",
-            add=False):
+        self,
+        particle_data: List[Particle],
+        sigma: float,
+        quantity: str,
+        kernel: str = "gaussian",
+        add: bool = False,
+    ) -> None:
         """
         Add particle data to the lattice.
 
@@ -1294,10 +1415,11 @@ class Lattice3D:
             particle data.
         quantity : str
             The quantity of the particle data to be added. Supported values are
-            'energy density', 'number', 'charge', and 'baryon number'.
-        kernel : str, optional
+            'energy_density', 'number_density', 'charge_density',
+            'baryon_density', and 'strangeness_density'.
+        kernel : str
             The type of kernel to use for smearing the particle data. Supported
-            values are 'gaussian' and 'covariant'. The default is 'covariant'.
+            values are 'gaussian' and 'covariant'. The default is 'gaussian'.
         add : bool, optional
             Specifies whether to add the particle data to the existing lattice
             values or replace them. If True, the particle data will be added to
@@ -1354,20 +1476,36 @@ class Lattice3D:
             if np.isnan(value):
                 raise ValueError("Particle data contains NaN values.")
 
-            if (kernel == "gaussian"):
+            if kernel == "gaussian":
                 # Calculate the Gaussian kernel centered at (x, y, z)
                 kernel_value = multivariate_normal(
-                    mean=[x, y, z], cov=sigma**2 * np.eye(3))
-            elif (kernel == "covariant"):
-                if np.isnan(
-                        particle.px) or np.isnan(
-                        particle.py) or np.isnan(
-                        particle.py):
+                    mean=[x, y, z], cov=sigma**2 * np.eye(3)
+                )
+            elif kernel == "covariant":
+                if (
+                    np.isnan(particle.px)
+                    or np.isnan(particle.py)
+                    or np.isnan(particle.py)
+                ):
                     raise ValueError("Particle data contains NaN values.")
                 kernel_value = multivariate_normal(
-                    mean=[0, 0], cov=sigma**2 * np.eye(2))
+                    mean=[0, 0], cov=sigma**2 * np.eye(2)
+                )
             else:
                 raise ValueError("Unknown kernel type for lattice.")
+
+            if self.spacing_x_ is None:
+                raise TypeError(
+                    "'spacing_x_' is None. It must be initialized before calling the 'add_particle_data' function."
+                )
+            if self.spacing_y_ is None:
+                raise TypeError(
+                    "'spacing_y_' is None. It must be initialized before calling the 'add_particle_data' function."
+                )
+            if self.spacing_z_ is None:
+                raise TypeError(
+                    "'spacing_z_' is None. It must be initialized before calling the 'add_particle_data' function."
+                )
 
             # Determine the range of cells within the boundary
             range_x = self.n_sigma_x_ * sigma
@@ -1385,8 +1523,17 @@ class Lattice3D:
             norm = 0
             # we have now the right spacing, but the lattice points do not
             # coincide
-            temp_lattice = Lattice3D(-range_x, range_x, -range_y, range_y, -
-                                     range_z, range_z, 2 * num_x + 1, 2 * num_y + 1, 2 * num_z + 1)
+            temp_lattice = Lattice3D(
+                -range_x,
+                range_x,
+                -range_y,
+                range_y,
+                -range_z,
+                range_z,
+                2 * num_x + 1,
+                2 * num_y + 1,
+                2 * num_z + 1,
+            )
 
             for i in range(temp_lattice.num_points_x_):
                 for j in range(temp_lattice.num_points_y_):
@@ -1395,22 +1542,30 @@ class Lattice3D:
                         xi, yj, zk = temp_lattice.get_coordinates(i, j, k)
 
                         # Calculate the value to add to the grid at (i, j, k)
-                        if (kernel == "gaussian"):
+                        if kernel == "gaussian":
                             smearing_factor = kernel_value.pdf(
-                                [xi + x, yj + y, zk + z])
+                                [xi + x, yj + y, zk + z]
+                            )
                         else:
-                            diff_space = (xi)**2 + (yj)**2 + (zk)**2
+                            diff_space = (xi) ** 2 + (yj) ** 2 + (zk) ** 2
                             gamma = np.sqrt(
-                                1 + particle.p_abs()**2 / particle.mass**2)
+                                1 + particle.p_abs() ** 2 / particle.mass**2
+                            )
                             diff_velocity = (
-                                particle.px * (xi) + particle.py * (yj) + particle.pz * (zk)) / (gamma * particle.mass)
+                                particle.px * (xi)
+                                + particle.py * (yj)
+                                + particle.pz * (zk)
+                            ) / (gamma * particle.mass)
                             smearing_factor = kernel_value.pdf(
-                                [diff_space, diff_velocity])
+                                [diff_space, diff_velocity]
+                            )
                         norm += smearing_factor
-                        value_to_add = value * smearing_factor / self.cell_volume_
+                        value_to_add = (
+                            value * smearing_factor / self.cell_volume_
+                        )
 
                         if np.isnan(value_to_add):
-                            value_to_add = 0.
+                            value_to_add = 0.0
                         # Add the value to the grid
                         temp_lattice.grid_[i, j, k] += value_to_add
 
@@ -1424,7 +1579,7 @@ class Lattice3D:
             self.add_same_spaced_grid(temp_lattice, x_grid, y_grid, z_grid)
 
 
-def print_lattice(lattice):
+def print_lattice(lattice: "Lattice3D") -> None:
     for i in range(lattice.num_points_x_):
         for j in range(lattice.num_points_y_):
             for k in range(lattice.num_points_z_):
