@@ -11,17 +11,19 @@ import numpy as np
 import math
 from particle import PDGID
 import warnings
+from typing import Optional, Union, List
 
 
 class Particle:
     """Defines a particle object.
 
     The member variables of the Particle class are the quantities in the
-    OSCAR2013/OSCAR2013Extended or JETSCAPE hadron output. If they are not set,
-    they stay `np.nan` to throw an error if one tries to access a non existing
+    Oscar2013/Oscar2013Extended/Oscar2013Extended_IC/Oscar2013Extended_Photons/
+    ASCII or JETSCAPE hadron/parton output. If they are not set,
+    they stay :code:`np.nan` to throw an error if one tries to access a non existing
     quantity.
     If a particle with an unknown PDG is provided, a warning is thrown and and
-    `np.nan` is returned for charge, spin, and spin degeneracy.
+    :code:`np.nan` is returned for charge, spin, and spin degeneracy.
 
     Attributes
     ----------
@@ -129,11 +131,11 @@ class Particle:
         Print the particle as CSV to terminal
     angular_momentum:
         Compute angular momentum
-    momentum_rapidity_Y:
+    rapidity:
         Compute momentum rapidity
     p_abs:
         Compute absolute momentum
-    pt_abs:
+    pT_abs:
         Compute absolute value of transverse momentum
     phi:
         Compute azimuthal angle
@@ -141,24 +143,40 @@ class Particle:
         Compute polar angle
     pseudorapidity:
         Compute pseudorapidity
-    spatial_rapidity:
-        Compute spatial rapidity
+    spacetime_rapidity:
+        Compute space-time rapidity
     proper_time:
         Compute proper time
-    compute_mass_from_energy_momentum:
+    mass_from_energy_momentum:
         Compute mass from energy momentum relation
-    compute_charge_from_pdg:
+    charge_from_pdg:
         Compute charge from PDG code
+    mT:
+        Compute transverse mass
+    is_quark:
+        Is the particle a quark?
+    is_lepton:
+        Is the particle a lepton?
     is_meson:
         Is the particle a meson?
     is_baryon:
         Is the particle a baryon?
     is_hadron:
         Is the particle a hadron?
-    is_strange:
-        Is the particle a strange particle?
     is_heavy_flavor:
         Is the particle a heavy flavor particle?
+    has_down:
+        Does the particle have a down quark?
+    has_up:
+        Does the particle have an up quark?
+    has_strange:
+        Does the particle have a strange quark?
+    has_charm:
+        Does the particle have a charm quark?
+    has_bottom:
+        Does the particle have a bottom quark?
+    has_top:
+        Does the particle have a top quark?
     weight:
         What is the weight of the particle?
     spin:
@@ -194,6 +212,8 @@ class Particle:
 
     * "Oscar2013Extended_Photons"
 
+    * "ASCII"
+
     * "JETSCAPE"
 
     .. highlight:: python
@@ -206,32 +226,52 @@ class Particle:
     Notes
     -----
     If a member of the Particle class is not set or a quantity should be computed
-    and the needed member variables are not set, then `np.nan` is returned by default.
-    All quantities are saved in a numpy array member variable `data_`. The datatype
+    and the needed member variables are not set, then :code:`np.nan` is returned by default.
+    All quantities are saved in a numpy array member variable :code:`data_`. The datatype
     of this array is float, therefore casting is required when int or bool values are
     required.
 
     When JETSCAPE creates particle objects, which are partons, the charge is multiplied
     by 3 to make it an integer.
-    The functions `is_strange()` and `is_heavy_flavor()` should not be used in this
-    case.
     """
-    __slots__ = ['data_']
 
-    def __init__(self, input_format=None, particle_array=None):
-        self.data_ = np.array(25 * [np.nan], dtype=float)
-        self.pdg_valid = False
+    __slots__ = ["data_"]
+
+    def __init__(
+        self,
+        input_format: Optional[str] = None,
+        particle_array: Optional[np.ndarray] = None,
+        attribute_list: List[str] = [],
+    ) -> None:
+        self.data_: np.ndarray = np.array(25 * [np.nan], dtype=float)
+        self.pdg_valid: bool = False
 
         if ((input_format is not None) and (particle_array is None)) or (
-                (input_format is None) and (particle_array is not None)):
+            (input_format is None) and (particle_array is not None)
+        ):
             raise ValueError("'input_format' or 'particle_array' not given")
 
-        if (input_format is not None) and (particle_array is not None):
-            self.__initialize_from_array(input_format, particle_array)
+        if attribute_list is None and input_format == "ASCII":
+            raise ValueError("'attribute_list' not given")
 
-    def __initialize_from_array(self, input_format, particle_array):
+        if attribute_list != [] and input_format != "ASCII":
+            raise ValueError(
+                "'ASCII' format requires 'attribute_list' to be given."
+            )
+
+        if (input_format is not None) and (particle_array is not None):
+            self.__initialize_from_array(
+                input_format, particle_array, attribute_list
+            )
+
+    def __initialize_from_array(
+        self,
+        input_format: str,
+        particle_array: np.ndarray,
+        attribute_list: List[str],
+    ) -> None:
         """
-        Initialize instance attributes based on the provided input format and array.
+        Initialize instance attributes based on the provided input format and array, and optionally an attribute list.
 
         Parameters
         ----------
@@ -241,10 +281,14 @@ class Particle:
             - "Oscar2013Extended"
             - "Oscar2013Extended_IC"
             - "Oscar2013Extended_Photons"
+            - "ASCII"
             - "JETSCAPE"
 
         particle_array : numpy.ndarray
             An array containing particle information.
+
+        attribute_list: List[str]
+            A list containing coded information which element of particle_array corresponds to which attribute of the Particle instance.
 
         Raises
         ------
@@ -266,6 +310,31 @@ class Particle:
         # first entry: index in data array
         # second entry: index in line
         attribute_mapping = {
+            "Allfields": {
+                "t": [0, 0],
+                "x": [1, 0],
+                "y": [2, 0],
+                "z": [3, 0],
+                "mass": [4, 0],
+                "E": [5, 0],
+                "px": [6, 0],
+                "py": [7, 0],
+                "pz_": [8, 0],
+                "pdg": [9, 0],
+                "ID": [11, 0],
+                "charge": [12, 0],
+                "ncoll": [13, 0],
+                "form_time": [14, 0],
+                "xsecfac": [15, 0],
+                "proc_id_origin": [16, 0],
+                "proc_type_origin": [17, 0],
+                "t_last_coll": [18, 0],
+                "pdg_mother1": [19, 0],
+                "pdg_mother2": [20, 0],
+                "status_": [21, 0],
+                "baryon_number": [22, 0],
+                "strangeness": [23, 0],
+            },
             "Oscar2013": {
                 "t_": [0, 0],
                 "x_": [1, 1],
@@ -361,19 +430,32 @@ class Particle:
                 "pz_": [8, 6],
             },
         }
-        if (input_format in attribute_mapping):
+        if input_format == "ASCII":
+            mapping_dict = {}
+            for attr in attribute_list:
+                mapping_dict[attr] = [
+                    attribute_mapping["Allfields"][attr][0],
+                    list(attribute_list).index(attr),
+                ]
+            attribute_mapping["ASCII"] = mapping_dict
+        if input_format in attribute_mapping or input_format == "ASCII":
             if (
-                len(particle_array) == len(
-                    attribute_mapping[input_format]) or (
-                    input_format in [
-                        "Oscar2013Extended",
-                        "Oscar2013Extended_IC"] and len(particle_array) <= len(
-                    attribute_mapping[input_format]) and len(particle_array) >= len(
-                        attribute_mapping[input_format]) -
-                    2)):
+                input_format == "ASCII"
+                or len(particle_array) == len(attribute_mapping[input_format])
+                or (
+                    input_format
+                    in ["Oscar2013Extended", "Oscar2013Extended_IC"]
+                    and len(particle_array)
+                    <= len(attribute_mapping[input_format])
+                    and len(particle_array)
+                    >= len(attribute_mapping[input_format]) - 2
+                )
+            ):
                 for attribute, index in attribute_mapping[input_format].items():
                     if len(particle_array) <= (index[1]):
                         continue
+                    if input_format == "ASCII":
+                        attribute = attribute + "_"
                     # Type casting for specific attributes. Although everything is saved as a float, we will only read in int data for int fields
                     # to ensure similar behaving as if we were reading in data
                     # into ints.
@@ -390,38 +472,66 @@ class Particle:
                         "form_time_",
                         "xsecfac_",
                         "t_last_coll_",
-                            "weight_"]:
+                        "weight_",
+                    ]:
                         self.data_[index[0]] = float(particle_array[index[1]])
-                    elif attribute in ["pdg_", "ID_", "ncoll_", "proc_id_origin_", "proc_type_origin_", "pdg_mother1_", "pdg_mother2_", "status_"]:
+                    elif attribute in [
+                        "pdg_",
+                        "ID_",
+                        "ncoll_",
+                        "proc_id_origin_",
+                        "proc_type_origin_",
+                        "pdg_mother1_",
+                        "pdg_mother2_",
+                        "status_",
+                    ]:
                         self.data_[index[0]] = int(particle_array[index[1]])
                     else:
                         self.data_[index[0]] = int(particle_array[index[1]])
 
                 # It is important for JETSCAPE particles to compute pdg_valid
-                # here because the compute_charge_from_pdg function depends on
+                # here because the charge_from_pdg function depends on
                 # it.
-                self.pdg_valid = PDGID(self.pdg).is_valid
+                if np.isnan(self.pdg):
+                    self.pdg_valid = False
+                else:
+                    self.pdg_valid = PDGID(self.pdg).is_valid
 
                 if input_format == "JETSCAPE":
-                    self.mass = self.compute_mass_from_energy_momentum()
-                    self.charge = self.compute_charge_from_pdg()
+                    self.mass = self.mass_from_energy_momentum()
+                    self.charge = self.charge_from_pdg()
                     if self.pdg_valid == False and np.isnan(self.charge):
-                        warnings.warn('The PDG code ' + str(int(self.pdg)) + \
-                                      ' is not known by PDGID, charge could not be computed. Consider setting it by hand.')
+                        warnings.warn(
+                            "The PDG code "
+                            + str(int(self.pdg))
+                            + " is not known by PDGID, charge could not be computed. Consider setting it by hand."
+                        )
             else:
-                raise ValueError("The input file is corrupted! " +
-                                 "A line with wrong number of columns " +
-                                 str(len(particle_array)) +
-                                 " was found.")
+                raise ValueError(
+                    "The input file is corrupted! "
+                    + "A line with wrong number of columns "
+                    + str(len(particle_array))
+                    + " was found."
+                )
         else:
             raise ValueError(f"Unsupported input format '{input_format}'")
 
-        if (not self.pdg_valid):
-            warnings.warn('The PDG code ' + str(int(self.pdg)) + ' is not valid. ' +
-                          'All properties extracted from the PDG are set to default values.')
+        if not self.pdg_valid:
+            if np.isnan(self.pdg):
+                warnings.warn(
+                    "No PDG code given! "
+                    + "All properties extracted from the PDG are set to default values."
+                )
+            else:
+                warnings.warn(
+                    "The PDG code "
+                    + str(int(self.pdg))
+                    + " is not valid. "
+                    + "All properties extracted from the PDG are set to default values."
+                )
 
     @property
-    def t(self):
+    def t(self) -> float:
         """Get or set the time of the particle.
 
         Returns
@@ -431,11 +541,11 @@ class Particle:
         return self.data_[0]
 
     @t.setter
-    def t(self, value):
+    def t(self, value: float) -> None:
         self.data_[0] = value
 
     @property
-    def x(self):
+    def x(self) -> float:
         """Get or set the x-position of the particle.
 
         Returns
@@ -445,11 +555,11 @@ class Particle:
         return self.data_[1]
 
     @x.setter
-    def x(self, value):
+    def x(self, value: float) -> None:
         self.data_[1] = value
 
     @property
-    def y(self):
+    def y(self) -> float:
         """Get or set the y-position of the particle.
 
         Returns
@@ -459,11 +569,11 @@ class Particle:
         return self.data_[2]
 
     @y.setter
-    def y(self, value):
+    def y(self, value: float) -> None:
         self.data_[2] = value
 
     @property
-    def z(self):
+    def z(self) -> float:
         """Get or set the z-position of the particle.
 
         Returns
@@ -473,11 +583,11 @@ class Particle:
         return self.data_[3]
 
     @z.setter
-    def z(self, value):
+    def z(self, value: float) -> None:
         self.data_[3] = value
 
     @property
-    def mass(self):
+    def mass(self) -> float:
         """Get or set the mass of the particle.
 
         Returns
@@ -487,11 +597,11 @@ class Particle:
         return self.data_[4]
 
     @mass.setter
-    def mass(self, value):
+    def mass(self, value: float) -> None:
         self.data_[4] = value
 
     @property
-    def E(self):
+    def E(self) -> float:
         """Get or set the energy of the particle.
 
         Returns
@@ -501,11 +611,11 @@ class Particle:
         return self.data_[5]
 
     @E.setter
-    def E(self, value):
+    def E(self, value: float) -> None:
         self.data_[5] = value
 
     @property
-    def px(self):
+    def px(self) -> float:
         """Get or set the momentum x-component of the particle.
 
         Returns
@@ -515,11 +625,11 @@ class Particle:
         return self.data_[6]
 
     @px.setter
-    def px(self, value):
+    def px(self, value: float) -> None:
         self.data_[6] = value
 
     @property
-    def py(self):
+    def py(self) -> float:
         """Get or set the momentum y-component of the particle.
 
         Returns
@@ -529,11 +639,11 @@ class Particle:
         return self.data_[7]
 
     @py.setter
-    def py(self, value):
+    def py(self, value: float) -> None:
         self.data_[7] = value
 
     @property
-    def pz(self):
+    def pz(self) -> float:
         """Get or set the momentum z-component of the particle.
 
         Returns
@@ -543,34 +653,36 @@ class Particle:
         return self.data_[8]
 
     @pz.setter
-    def pz(self, value):
+    def pz(self, value: float) -> None:
         self.data_[8] = value
 
     @property
-    def pdg(self):
+    def pdg(self) -> Union[int, float]:
         """Get or set the PDG code of the particle.
 
         Returns
         -------
         pdg : int
         """
-        if (np.isnan(self.data_[9])):
+        if np.isnan(self.data_[9]):
             return np.nan
         return int(self.data_[9])
 
     @pdg.setter
-    def pdg(self, value):
+    def pdg(self, value: float) -> None:
         self.data_[9] = value
         self.pdg_valid = PDGID(self.pdg).is_valid
 
-        if (not self.pdg_valid):
-            warnings.warn('The PDG code ' +
-                          str(int(self.pdg)) +
-                          ' is not valid. ' +
-                          'All properties extracted from the PDG are set to nan.')
+        if not self.pdg_valid:
+            warnings.warn(
+                "The PDG code "
+                + str(int(self.pdg))
+                + " is not valid. "
+                + "All properties extracted from the PDG are set to nan."
+            )
 
     @property
-    def ID(self):
+    def ID(self) -> Union[int, float]:
         """Get or set the ID of the particle.
 
         This is a unique number in SMASH.
@@ -579,28 +691,28 @@ class Particle:
         -------
         ID : int
         """
-        if (np.isnan(self.data_[11])):
+        if np.isnan(self.data_[11]):
             return np.nan
         return int(self.data_[11])
 
     @ID.setter
-    def ID(self, value):
+    def ID(self, value: float) -> None:
         self.data_[11] = value
 
     @property
-    def charge(self):
+    def charge(self) -> Union[int, float]:
         """Get or set the electrical charge of the particle.
 
         Returns
         -------
         charge : int
         """
-        if (np.isnan(self.data_[12])):
+        if np.isnan(self.data_[12]):
             return np.nan
         return int(self.data_[12])
 
     @charge.setter
-    def charge(self, value):
+    def charge(self, value: float) -> None:
         # this is for the case a parton is created from the JETSCAPE reader
         # handle quarks with 3 times the charge to make it integer
         if np.abs(value) < 1:
@@ -608,23 +720,23 @@ class Particle:
         self.data_[12] = value
 
     @property
-    def ncoll(self):
+    def ncoll(self) -> Union[int, float]:
         """Get or set the number of collisions of the particle.
 
         Returns
         -------
         ncoll : int
         """
-        if (np.isnan(self.data_[13])):
+        if np.isnan(self.data_[13]):
             return np.nan
         return int(self.data_[13])
 
     @ncoll.setter
-    def ncoll(self, value):
+    def ncoll(self, value: float) -> None:
         self.data_[13] = value
 
     @property
-    def form_time(self):
+    def form_time(self) -> float:
         """Get or set the formation time of the particle.
 
         Returns
@@ -634,11 +746,11 @@ class Particle:
         return self.data_[14]
 
     @form_time.setter
-    def form_time(self, value):
+    def form_time(self, value: float) -> None:
         self.data_[14] = value
 
     @property
-    def xsecfac(self):
+    def xsecfac(self) -> float:
         """Get or set the cross section scaling factor of the particle.
 
         Returns
@@ -648,43 +760,43 @@ class Particle:
         return self.data_[15]
 
     @xsecfac.setter
-    def xsecfac(self, value):
+    def xsecfac(self, value: float) -> None:
         self.data_[15] = value
 
     @property
-    def proc_id_origin(self):
+    def proc_id_origin(self) -> Union[int, float]:
         """Get or set the process ID of the particle's origin.
 
         Returns
         -------
         proc_id_origin : int
         """
-        if (np.isnan(self.data_[16])):
+        if np.isnan(self.data_[16]):
             return np.nan
         return int(self.data_[16])
 
     @proc_id_origin.setter
-    def proc_id_origin(self, value):
+    def proc_id_origin(self, value: float) -> None:
         self.data_[16] = value
 
     @property
-    def proc_type_origin(self):
+    def proc_type_origin(self) -> Union[int, float]:
         """Get or set the process type of the particle's origin.
 
         Returns
         -------
         proc_type_origin : int
         """
-        if (np.isnan(self.data_[17])):
+        if np.isnan(self.data_[17]):
             return np.nan
         return int(self.data_[17])
 
     @proc_type_origin.setter
-    def proc_type_origin(self, value):
+    def proc_type_origin(self, value: float) -> None:
         self.data_[17] = value
 
     @property
-    def t_last_coll(self):
+    def t_last_coll(self) -> float:
         """Get or set the last time of a collision of the particle.
 
         Returns
@@ -694,43 +806,43 @@ class Particle:
         return self.data_[18]
 
     @t_last_coll.setter
-    def t_last_coll(self, value):
+    def t_last_coll(self, value: float) -> None:
         self.data_[18] = value
 
     @property
-    def pdg_mother1(self):
+    def pdg_mother1(self) -> Union[int, float]:
         """Get the PDG code of the first mother particle.
 
         Returns
         -------
         pdg_mother1 : int
         """
-        if (np.isnan(self.data_[19])):
+        if np.isnan(self.data_[19]):
             return np.nan
         return int(self.data_[19])
 
     @pdg_mother1.setter
-    def pdg_mother1(self, value):
+    def pdg_mother1(self, value: float) -> None:
         self.data_[19] = value
 
     @property
-    def pdg_mother2(self):
+    def pdg_mother2(self) -> Union[int, float]:
         """Get the PDG code of the second mother particle.
 
         Returns
         -------
         pdg_mother2 : int
         """
-        if (np.isnan(self.data_[20])):
+        if np.isnan(self.data_[20]):
             return np.nan
         return int(self.data_[20])
 
     @pdg_mother2.setter
-    def pdg_mother2(self, value):
+    def pdg_mother2(self, value: float) -> None:
         self.data_[20] = value
 
     @property
-    def status(self):
+    def status(self) -> Union[int, float]:
         """
         Get the hadron status (stores information on the module origin of
         a JETSCAPE hadron).
@@ -739,16 +851,16 @@ class Particle:
         -------
         status : int
         """
-        if (np.isnan(self.data_[21])):
+        if np.isnan(self.data_[21]):
             return np.nan
         return int(self.data_[21])
 
     @status.setter
-    def status(self, value):
+    def status(self, value: float) -> None:
         self.data_[21] = value
 
     @property
-    def baryon_number(self):
+    def baryon_number(self) -> Union[int, float]:
         """Get the baryon number of the particle.
 
         Returns
@@ -760,11 +872,11 @@ class Particle:
         return int(self.data_[22])
 
     @baryon_number.setter
-    def baryon_number(self, value):
+    def baryon_number(self, value: float) -> None:
         self.data_[22] = value
 
     @property
-    def strangeness(self):
+    def strangeness(self) -> Union[int, float]:
         """Get the strangeness of the particle.
 
         Returns
@@ -776,11 +888,11 @@ class Particle:
         return int(self.data_[23])
 
     @strangeness.setter
-    def strangeness(self, value):
+    def strangeness(self, value: float) -> None:
         self.data_[23] = value
 
     @property
-    def weight(self):
+    def weight(self) -> float:
         """Get the weight of the particle.
 
         Returns
@@ -790,11 +902,11 @@ class Particle:
         return self.data_[24]
 
     @weight.setter
-    def weight(self, value):
+    def weight(self, value: float) -> None:
         self.data_[24] = value
 
     @property
-    def pdg_valid(self):
+    def pdg_valid(self) -> bool:
         """Get the validity of the PDG code of the particle.
 
         Returns
@@ -804,32 +916,38 @@ class Particle:
         return bool(self.data_[10])
 
     @pdg_valid.setter
-    def pdg_valid(self, value):
+    def pdg_valid(self, value: bool) -> None:
         self.data_[10] = 1 if value else 0
 
-    def print_particle(self):
+    def print_particle(self) -> None:
         """Print the whole particle information as csv string.
 
         This function prints a header line with the different quantities.
         All particle quantities are then printed in the next line separated by
         a comma.
         """
-        def int_isnan(value):
+
+        def int_isnan(value: float) -> Union[int, float]:
             if np.isnan(value):
                 return np.nan
             else:
                 return int(value)
-        print('t,x,y,z,mass,E,px,py,pz,pdg,ID,charge,ncoll,form_time,xsecfac,\
+
+        print(
+            "t,x,y,z,mass,E,px,py,pz,pdg,ID,charge,ncoll,form_time,xsecfac,\
               proc_id_origin,proc_type_origin,t_last_coll,pdg_mother1,\
-              pdg_mother2,status,baryon_number,strangeness,weight')
-        print(f'{self.t},{self.x},{self.y},{self.z},{self.mass},{self.E},\
+              pdg_mother2,status,baryon_number,strangeness,weight"
+        )
+        print(
+            f"{self.t},{self.x},{self.y},{self.z},{self.mass},{self.E},\
               {self.px},{self.py},{self.pz},{int_isnan(self.pdg)},{int_isnan(self.ID)},\
               {int_isnan(self.charge)},{int_isnan(self.ncoll)},{self.form_time},{self.xsecfac},\
               {int_isnan(self.proc_id_origin)},{int_isnan(self.proc_type_origin)}\
               ,{self.t_last_coll},{int_isnan(self.pdg_mother1)},{int_isnan(self.pdg_mother2)},\
-              {int_isnan(self.status)},{int_isnan(self.baryon_number)},{int_isnan(self.strangeness)},{self.weight}')
+              {int_isnan(self.status)},{int_isnan(self.baryon_number)},{int_isnan(self.strangeness)},{self.weight}"
+        )
 
-    def angular_momentum(self):
+    def angular_momentum(self) -> Union[np.ndarray, float]:
         """
         Compute the angular momentum :math:`\\vec{L}=\\vec{r} \\times \\vec{p}` of a particle.
 
@@ -841,18 +959,24 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
-        if np.isnan(self.x) or np.isnan(self.y) or np.isnan(self.z)\
-                or np.isnan(self.px) or np.isnan(self.py) or np.isnan(self.pz):
+        if (
+            np.isnan(self.x)
+            or np.isnan(self.y)
+            or np.isnan(self.z)
+            or np.isnan(self.px)
+            or np.isnan(self.py)
+            or np.isnan(self.pz)
+        ):
             return np.nan
         else:
             r = [self.x, self.y, self.z]
             p = [self.px, self.py, self.pz]
             return np.cross(r, p)
 
-    def momentum_rapidity_Y(self):
+    def rapidity(self) -> float:
         """
         Compute the momentum rapidity :math:`Y=\\frac{1}{2}\\ln\\left(\\frac{E+p_z}{E-p_z}\\right)` of the particle.
 
@@ -863,7 +987,7 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.E) or np.isnan(self.pz):
@@ -873,11 +997,11 @@ class Particle:
                 # Adding a small positive value
                 denominator = (self.E - self.pz) + 1e-10
             else:
-                denominator = (self.E - self.pz)
+                denominator = self.E - self.pz
 
             return 0.5 * np.log((self.E + self.pz) / denominator)
 
-    def p_abs(self):
+    def p_abs(self) -> float:
         """
         Compute the absolute momentum :math:`|\\vec{p}|=\\sqrt{p_x^2+p_y^2+p_z^2}` of the particle.
 
@@ -888,15 +1012,15 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.px) or np.isnan(self.py) or np.isnan(self.pz):
             return np.nan
         else:
-            return np.sqrt(self.px**2. + self.py**2. + self.pz**2.)
+            return np.sqrt(self.px**2.0 + self.py**2.0 + self.pz**2.0)
 
-    def pt_abs(self):
+    def pT_abs(self) -> float:
         """
         Compute the absolute transverse momentum :math:`|\\vec{p}_{\\mathrm{T}}|=\\sqrt{p_x^2+p_y^2}` of the particle.
 
@@ -907,15 +1031,15 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.px) or np.isnan(self.py):
             return np.nan
         else:
-            return np.sqrt(self.px**2. + self.py**2.)
+            return np.sqrt(self.px**2.0 + self.py**2.0)
 
-    def phi(self):
+    def phi(self) -> float:
         """
         Compute the azimuthal angle of the particle.
 
@@ -926,18 +1050,18 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.px) or np.isnan(self.py):
             return np.nan
         else:
             if (np.abs(self.px) < 1e-6) and (np.abs(self.py) < 1e-6):
-                return 0.
+                return 0.0
             else:
                 return math.atan2(self.py, self.px)
 
-    def theta(self):
+    def theta(self) -> float:
         """
         Compute the polar angle of the particle.
 
@@ -948,18 +1072,18 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.px) or np.isnan(self.py) or np.isnan(self.pz):
             return np.nan
         else:
             if self.p_abs() == 0.0:
-                return 0.
+                return 0.0
             else:
                 return np.arccos(self.pz / self.p_abs())
 
-    def pseudorapidity(self):
+    def pseudorapidity(self) -> float:
         """
         Compute the pseudorapidity :math:`\\eta=\\frac{1}{2}\\ln\\left(\\frac{|\\vec{p}|+p_z}{|\\vec{p}|-p_z}\\right)` of the particle.
 
@@ -970,32 +1094,33 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.px) or np.isnan(self.py) or np.isnan(self.pz):
             return np.nan
         else:
             if abs(self.p_abs() - self.pz) < 1e-10:
-                denominator = (self.p_abs() - self.pz) + \
-                    1e-10  # Adding a small positive value
+                denominator = (
+                    self.p_abs() - self.pz
+                ) + 1e-10  # Adding a small positive value
             else:
-                denominator = (self.p_abs() - self.pz)
+                denominator = self.p_abs() - self.pz
 
             return 0.5 * np.log((self.p_abs() + self.pz) / denominator)
 
-    def spatial_rapidity(self):
+    def spacetime_rapidity(self) -> float:
         """
-        Compute the spatial rapidity :math:`y=\\frac{1}{2}\\ln\\left(\\frac{t+z}{t-z}\\right)` of the particle.
+        Compute the space-time rapidity :math:`\\eta_s=\\frac{1}{2}\\ln\\left(\\frac{t+z}{t-z}\\right)` of the particle.
 
         Returns
         -------
         float
-            spatial rapidity
+            space-time rapidity
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.t) or np.isnan(self.z):
@@ -1006,7 +1131,7 @@ class Particle:
             else:
                 raise ValueError("|z| < t not fulfilled")
 
-    def proper_time(self):
+    def proper_time(self) -> float:
         """
         Compute the proper time :math:`\\tau=\\sqrt{t^2-z^2}` of the particle.
 
@@ -1017,22 +1142,27 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
+        If one of the needed particle quantities is not given, then :code:`np.nan`
         is returned.
         """
         if np.isnan(self.t) or np.isnan(self.z):
             return np.nan
         else:
             if self.t > np.abs(self.z):
-                return np.sqrt(self.t**2. - self.z**2.)
+                return np.sqrt(self.t**2.0 - self.z**2.0)
             else:
                 raise ValueError("|z| < t not fulfilled")
 
-    def compute_mass_from_energy_momentum(self):
+    def mass_from_energy_momentum(self) -> float:
         """
         Compute the mass from the energy momentum relation.
 
         This function is called automatically if a JETSCAPE file is read in.
+
+        We consider particles with the following PDG codes as massless:
+        photons (22), gluons (21), e-neutrinos (12, -12),
+        mu-neutrinos (14, -14), tau-neutrinos (16, -16),
+        tau-prime-neutrinos (18, -18).
 
         Returns
         -------
@@ -1041,23 +1171,31 @@ class Particle:
 
         Notes
         -----
-        If one of the needed particle quantities is not given, then `np.nan`
-        is returned.
+        If one of the needed particle quantities (four-momentum) is not given,
+        then :code:`np.nan` is returned.
         """
-        if np.isnan(
-            self.E) or np.isnan(
-            self.px) or np.isnan(
-            self.py) or np.isnan(
-                self.pz):
+        # photons and gluons are massless, consider neutrinos as massless
+        massless_pdg = [22, 21, 12, -12, 14, -14, 16, -16, 18, -18]
+        if (
+            np.isnan(self.E)
+            or np.isnan(self.px)
+            or np.isnan(self.py)
+            or np.isnan(self.pz)
+        ):
             return np.nan
+        elif self.pdg in massless_pdg:
+            return 0.0
         else:
-            if np.abs(self.E**2. - self.p_abs()**2.) > 1e-16 and\
-                    self.E**2. - self.p_abs()**2. > 0.:
-                return np.sqrt(self.E**2. - self.p_abs()**2.)
+            if abs(self.E) >= abs(self.p_abs()):
+                return np.sqrt(self.E**2.0 - self.p_abs() ** 2.0)
             else:
-                return 0.
+                warnings.warn(
+                    "|E| >= |p| not fulfilled or not within numerical precision! "
+                    "The mass is set to nan."
+                )
+                return np.nan
 
-    def compute_charge_from_pdg(self):
+    def charge_from_pdg(self) -> float:
         """
         Compute the charge from the PDG code.
 
@@ -1070,13 +1208,72 @@ class Particle:
 
         Notes
         -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
         """
         if not self.pdg_valid:
             return np.nan
         return PDGID(self.pdg).charge
 
-    def is_meson(self):
+    def mT(self) -> float:
+        """
+        Compute the transverse mass :math:`m_{T}=\\sqrt{E^2-p_z^2}` of the particle.
+
+        Returns
+        -------
+        float
+            transverse mass
+
+        Notes
+        -----
+        If one of the needed particle quantities is not given, then :code:`np.nan`
+        is returned.
+        """
+        if np.isnan(self.E) or np.isnan(self.pz):
+            return np.nan
+        elif abs(self.E) >= abs(self.pz):
+            return np.sqrt(self.E**2.0 - self.pz**2.0)
+        else:
+            warnings.warn(
+                "|E| >= |pz| not fulfilled or not within numerical precision! "
+                "The transverse mass is set to nan."
+            )
+            return np.nan
+
+    def is_quark(self) -> Union[bool, float]:
+        """
+        Is the particle a quark?
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).is_quark
+
+    def is_lepton(self) -> Union[bool, float]:
+        """
+        Is the particle a lepton?
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).is_lepton
+
+    def is_meson(self) -> Union[bool, float]:
         """
         Is the particle a meson?
 
@@ -1087,13 +1284,13 @@ class Particle:
 
         Notes
         -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
         """
         if not self.pdg_valid:
             return np.nan
         return PDGID(self.pdg).is_meson
 
-    def is_baryon(self):
+    def is_baryon(self) -> Union[bool, float]:
         """
         Is the particle a baryon?
 
@@ -1104,13 +1301,13 @@ class Particle:
 
         Notes
         -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
         """
         if not self.pdg_valid:
             return np.nan
         return PDGID(self.pdg).is_baryon
 
-    def is_hadron(self):
+    def is_hadron(self) -> Union[bool, float]:
         """
         Is the particle a hadron?
 
@@ -1121,30 +1318,13 @@ class Particle:
 
         Notes
         -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
         """
         if not self.pdg_valid:
             return np.nan
         return PDGID(self.pdg).is_hadron
 
-    def is_strange(self):
-        """
-        Does the particle contain strangeness?
-
-        Returns
-        -------
-        bool
-            True, False
-
-        Notes
-        -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
-        """
-        if not self.pdg_valid:
-            return np.nan
-        return PDGID(self.pdg).has_strange
-
-    def is_heavy_flavor(self):
+    def is_heavy_flavor(self) -> Union[bool, float]:
         """
         Is the particle a heavy flavor hadron?
 
@@ -1155,17 +1335,122 @@ class Particle:
 
         Notes
         -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
         """
         if not self.pdg_valid:
             return np.nan
-        if PDGID(self.pdg).has_charm or PDGID(self.pdg).has_bottom\
-                or PDGID(self.pdg).has_top:
+        if (
+            PDGID(self.pdg).has_charm
+            or PDGID(self.pdg).has_bottom
+            or PDGID(self.pdg).has_top
+        ):
             return True
         else:
             return False
 
-    def spin(self):
+    def has_down(self) -> Union[bool, float]:
+        """
+        Does the particle contain a down quark? Does not work with partons.
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).has_down
+
+    def has_up(self) -> Union[bool, float]:
+        """
+        Does the particle contain an up quark?  Does not work with partons.
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).has_up
+
+    def has_strange(self) -> Union[bool, float]:
+        """
+        Does the particle contain a strange quark?  Does not work with partons.
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).has_strange
+
+    def has_charm(self) -> Union[bool, float]:
+        """
+        Does the particle contain a charm quark?  Does not work with partons.
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).has_charm
+
+    def has_bottom(self) -> Union[bool, float]:
+        """
+        Does the particle contain a bottom quark?  Does not work with partons.
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).has_bottom
+
+    def has_top(self) -> Union[bool, float]:
+        """
+        Does the particle contain a top quark?  Does not work with partons.
+
+        Returns
+        -------
+        bool
+            True, False
+
+        Notes
+        -----
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
+        """
+        if not self.pdg_valid:
+            return np.nan
+        return PDGID(self.pdg).has_top
+
+    def spin(self) -> float:
         """
         Get the total spin :math:`J` of the particle.
 
@@ -1176,13 +1461,13 @@ class Particle:
 
         Notes
         -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
         """
         if not self.pdg_valid:
             return np.nan
         return PDGID(self.pdg).J
 
-    def spin_degeneracy(self):
+    def spin_degeneracy(self) -> Union[int, float]:
         """
         Get the number of all possible spin projections (:math:`2J + 1`).
 
@@ -1193,7 +1478,7 @@ class Particle:
 
         Notes
         -----
-        If the PDG ID is not known by `PDGID`, then `np.nan` is returned.
+        If the PDG ID is not known by `PDGID`, then :code:`np.nan` is returned.
         """
         if not self.pdg_valid:
             return np.nan
