@@ -8,10 +8,12 @@
 # ===================================================
 
 from sparkx.flow import FlowInterface
-import numpy as np
-from scipy import special
-import scipy.optimize as optimize
+import numpy as np  # type: ignore
+from scipy import special  # type: ignore
+import scipy.optimize as optimize  # type: ignore
 import warnings
+from sparkx.Particle import Particle
+from typing import List, Tuple, Union
 
 
 class EventPlaneFlow(FlowInterface.FlowInterface):
@@ -50,7 +52,7 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
     n : int, optional
         The value of the harmonic. Default is 2.
     weight : str, optional
-        The weight used for calculating the flow. Default is "pt2".
+        The weight used for calculating the flow. Default is :code:`pt2`.
     pseudorapidity_gap : float, optional
         The pseudorapidity gap used for dividing the particles into sub-events.
         Default is 0.0.
@@ -97,7 +99,9 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
 
     """
 
-    def __init__(self, n=2, weight="pt2", pseudorapidity_gap=0.):
+    def __init__(
+        self, n: int = 2, weight: str = "pT2", pseudorapidity_gap: float = 0.0
+    ) -> None:
         """
         Initialize the ScalarProductFlow object.
 
@@ -106,69 +110,75 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
         n : int, optional
             The value of the harmonic. Default is 2.
         weight : str, optional
-            The weight used for calculating the flow. Default is "pt2".
+            The weight used for calculating the flow. Default is "pT2".
         pseudorapidity_gap : float, optional
             The pseudorapidity gap used for dividing the particles into sub-events.
             Default is 0.0.
         """
 
         if not isinstance(n, int):
-            raise TypeError('n has to be int')
+            raise TypeError("n has to be int")
         elif n <= 0:
             raise ValueError(
-                'n-th harmonic with value n<=0 can not be computed')
+                "n-th harmonic with value n<=0 can not be computed"
+            )
         else:
             self.n_ = n
-
         if not isinstance(weight, str):
-            raise TypeError('weight has to be a string')
-        elif weight not in ["pt", "pt2", "ptn", "rapidity", "pseudorapidity"]:
+            raise TypeError("weight has to be a string")
+        elif weight not in ["pT", "pT2", "pTn", "rapidity", "pseudorapidity"]:
             raise ValueError(
-                "Invalid weight given, choose one of the following: 'pt', 'pt2', 'ptn', 'rapidity', 'pseudorapidity'")
+                "Invalid weight given, choose one of the following: 'pT', 'pT2', 'pTn', 'rapidity', 'pseudorapidity'"
+            )
         else:
             self.weight_ = weight
-
         if not isinstance(pseudorapidity_gap, (int, float)):
-            raise TypeError('n has to be int')
+            raise TypeError("n has to be int")
         elif pseudorapidity_gap < 0:
             raise ValueError(
-                'pseudorapidity value with gap < 0 can not be computed')
+                "pseudorapidity value with gap < 0 can not be computed"
+            )
         else:
             self.pseudorapidity_gap_ = pseudorapidity_gap
 
-    def __compute_particle_weights(self, particle_data):
+    def __compute_particle_weights(
+        self, particle_data: List[List[Particle]]
+    ) -> List[List[float]]:
         event_weights = []
         for event in range(len(particle_data)):
             particle_weights = []
             for particle in particle_data[event]:
-                weight = 0.
-                if self.weight_ == "pt":
-                    weight = particle.pt_abs()
-                elif self.weight_ == "pt2":
-                    weight = particle.pt_abs()**2.
-                elif self.weight_ == "ptn":
-                    weight = particle.pt_abs()**self.n_
+                weight = 0.0
+                if self.weight_ == "pT":
+                    weight = particle.pT_abs()
+                elif self.weight_ == "pT2":
+                    weight = particle.pT_abs() ** 2.0
+                elif self.weight_ == "pTn":
+                    weight = particle.pT_abs() ** self.n_
                 elif self.weight_ == "rapidity":
-                    weight = particle.momentum_rapidity_Y()
+                    weight = particle.rapidity()
                 elif self.weight_ == "pseudorapidity":
                     weight = particle.pseudorapidity()
                 particle_weights.append(weight)
             event_weights.append(particle_weights)
         return event_weights
 
-    def __compute_flow_vectors(self, particle_data, weights):
+    def __compute_flow_vectors(
+        self, particle_data: List[List[Particle]], weights: List[List[float]]
+    ) -> List[complex]:
         # Q vector whole event
         Q_vector = []
         for event in range(len(particle_data)):
-            Q_vector_val = 0. + 0.j
+            Q_vector_val = 0.0 + 0.0j
             for particle in range(len(particle_data[event])):
                 Q_vector_val += weights[event][particle] * np.exp(
-                    1.0j * float(self.n_) * particle_data[event][particle].phi())
+                    1.0j * float(self.n_) * particle_data[event][particle].phi()
+                )
             Q_vector.append(Q_vector_val)
 
         return Q_vector
 
-    def __sum_weights(self, weights):
+    def __sum_weights(self, weights: List[List[float]]) -> List[float]:
         sum_weights = []
         for event in weights:
             weight_val = np.sum(np.square(event))
@@ -176,18 +186,25 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
 
         return sum_weights
 
-    def __compute_event_angles_sub_events(self, particle_data, weights):
+    def __compute_event_angles_sub_events(
+        self, particle_data: List[List[Particle]], weights: List[List[float]]
+    ) -> Tuple[List[float], List[float]]:
         # Q vector sub-event A
         Q_vector_A = []
         relevant_weights_A = []
         for event in range(len(particle_data)):
-            Q_vector_A_val = 0. + 0.j
+            Q_vector_A_val = 0.0 + 0.0j
             relevant_weights_A_event = []
             for particle in range(len(particle_data[event])):
-                if particle_data[event][particle].pseudorapidity(
-                ) >= +self.pseudorapidity_gap_:
+                if (
+                    particle_data[event][particle].pseudorapidity()
+                    >= +self.pseudorapidity_gap_
+                ):
                     Q_vector_A_val += weights[event][particle] * np.exp(
-                        1.0j * float(self.n_) * particle_data[event][particle].phi())
+                        1.0j
+                        * float(self.n_)
+                        * particle_data[event][particle].phi()
+                    )
                     relevant_weights_A_event.append(weights[event][particle])
             Q_vector_A.append(Q_vector_A_val)
             relevant_weights_A.extend([relevant_weights_A_event])
@@ -196,13 +213,18 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
         Q_vector_B = []
         relevant_weights_B = []
         for event in range(len(particle_data)):
-            Q_vector_B_val = 0. + 0.j
+            Q_vector_B_val = 0.0 + 0.0j
             relevant_weights_B_event = []
             for particle in range(len(particle_data[event])):
-                if particle_data[event][particle].pseudorapidity(
-                ) < -self.pseudorapidity_gap_:
+                if (
+                    particle_data[event][particle].pseudorapidity()
+                    < -self.pseudorapidity_gap_
+                ):
                     Q_vector_B_val += weights[event][particle] * np.exp(
-                        1.0j * float(self.n_) * particle_data[event][particle].phi())
+                        1.0j
+                        * float(self.n_)
+                        * particle_data[event][particle].phi()
+                    )
                     relevant_weights_B_event.append(weights[event][particle])
             Q_vector_B.append(Q_vector_B_val)
             relevant_weights_B.extend([relevant_weights_B_event])
@@ -225,42 +247,41 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
 
         # compute event plane angles of sub-events
         Psi_A = [
-            (1. /
-             float(
-                 self.n_)) *
-            np.arctan2(
-                Q_vector_A[event].imag,
-                Q_vector_A[event].real) for event in range(
-                len(particle_data))]
+            (1.0 / float(self.n_))
+            * np.arctan2(Q_vector_A[event].imag, Q_vector_A[event].real)
+            for event in range(len(particle_data))
+        ]
         Psi_B = [
-            (1. /
-             float(
-                 self.n_)) *
-            np.arctan2(
-                Q_vector_B[event].imag,
-                Q_vector_B[event].real) for event in range(
-                len(particle_data))]
+            (1.0 / float(self.n_))
+            * np.arctan2(Q_vector_B[event].imag, Q_vector_B[event].real)
+            for event in range(len(particle_data))
+        ]
 
         return Psi_A, Psi_B
 
-    def __compute_u_vectors(self, particle_data):
+    def __compute_u_vectors(
+        self, particle_data: List[List[Particle]]
+    ) -> List[List[complex]]:
         u_vectors = []  # [event][particle]
         for event in particle_data:
             u_vector_event = []
             for particle in event:
                 u_vector_event.append(
-                    np.exp(
-                        1.0j *
-                        float(
-                            self.n_) *
-                        particle.phi()))
+                    np.exp(1.0j * float(self.n_) * particle.phi())
+                )
             u_vectors.extend([u_vector_event])
 
         return u_vectors
 
-    def __compute_event_plane_resolution(self, Psi_A, Psi_B):
+    def __compute_event_plane_resolution(
+        self, Psi_A: List[float], Psi_B: List[float]
+    ) -> float:
         RnSquared = np.asarray(
-            [np.cos(self.n_ * (Psi_A[event] - Psi_B[event])) for event in range(len(Psi_A))])
+            [
+                np.cos(self.n_ * (Psi_A[event] - Psi_B[event]))
+                for event in range(len(Psi_A))
+            ]
+        )
         RnSquaredSum = np.sum(RnSquared)
         Rn = np.sqrt(RnSquaredSum / len(Psi_A))
 
@@ -268,37 +289,43 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
         # x_new = sqrt(2) * x
         # alternative: if Rn < 0.5: R = Rn * sqrt(2), we don't do that
         # implements: arXiv:0809.2949
-        def resolution(x):
-            R = (np.sqrt(np.pi) / 2.) * x * np.exp(-0.5 * x * x) * \
-                (special.i0(0.5 * x * x) + special.i1(0.5 * x * x))
+        def resolution(x: float) -> float:
+            R = (
+                (np.sqrt(np.pi) / 2.0)
+                * x
+                * np.exp(-0.5 * x * x)
+                * (special.i0(0.5 * x * x) + special.i1(0.5 * x * x))
+            )
             return R
 
-        def f1(x, Rn):
+        def f1(x: float, Rn: float) -> float:
             return resolution(x) - Rn
 
-        def f1_wrapper(x):
+        def f1_wrapper(x: float) -> float:
             return f1(x, Rn)
 
         try:
             xi = optimize.root_scalar(
-                f1_wrapper, bracket=[
-                    0, 20], method='brentq').root
+                f1_wrapper, bracket=[0, 20], method="brentq"
+            ).root
         except BaseException:
             warnings.warn(
-                "Could not find solution of resolution equation. Use approximation instead.")
+                "Could not find solution of resolution equation. Use approximation instead."
+            )
             return Rn
 
         xi_new = np.sqrt(2) * xi
         return resolution(xi_new)
 
     def __compute_flow_particles(
-            self,
-            particle_data,
-            weights,
-            Q_vector,
-            u_vectors,
-            resolution,
-            self_corr):
+        self,
+        particle_data: List[List[Particle]],
+        weights: List[List[float]],
+        Q_vector: List[complex],
+        u_vectors: List[List[complex]],
+        resolution: float,
+        self_corr: bool,
+    ) -> Tuple[List[List[float]], List[List[float]]]:
         flow_values = []
         psi_values = []
         for event in range(len(particle_data)):
@@ -307,14 +334,18 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
             for particle in range(len(particle_data[event])):
                 weight_particle = np.abs(weights[event][particle])
                 Q_vector_particle = Q_vector[event]
-                if (self_corr):
-                    Q_vector_particle -= weight_particle * \
-                        u_vectors[event][particle]  # avoid autocorrelation
+                if self_corr:
+                    Q_vector_particle -= (
+                        weight_particle * u_vectors[event][particle]
+                    )  # avoid autocorrelation
 
-                Psi_n = (1. / float(self.n_)) * \
-                    np.arctan2(Q_vector_particle.imag, Q_vector_particle.real)
-                vn_obs = np.cos(float(self.n_) *
-                                (particle_data[event][particle].phi() - Psi_n))
+                Psi_n = (1.0 / float(self.n_)) * np.arctan2(
+                    Q_vector_particle.imag, Q_vector_particle.real
+                )
+                vn_obs = np.cos(
+                    float(self.n_)
+                    * (particle_data[event][particle].phi() - Psi_n)
+                )
 
                 flow_of_particle = vn_obs / resolution
                 flow_values_event.append(flow_of_particle)
@@ -323,50 +354,70 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
             psi_values.extend([psi_values_event])
         return flow_values, psi_values
 
-    def __calculate_reference(self, particle_data_event_plane):
+    def __calculate_reference(
+        self, particle_data_event_plane: List[List[Particle]]
+    ) -> Tuple[float, List[complex]]:
         event_weights_event_plane = self.__compute_particle_weights(
-            particle_data_event_plane)
+            particle_data_event_plane
+        )
         Psi_A, Psi_B = self.__compute_event_angles_sub_events(
-            particle_data_event_plane, event_weights_event_plane)
+            particle_data_event_plane, event_weights_event_plane
+        )
         resolution = self.__compute_event_plane_resolution(Psi_A, Psi_B)
         Q_vector = self.__compute_flow_vectors(
-            particle_data_event_plane, event_weights_event_plane)
+            particle_data_event_plane, event_weights_event_plane
+        )
 
         return resolution, Q_vector
 
     def __calculate_particle_flow(
-            self,
-            particle_data,
-            resolution,
-            Q_vector,
-            self_corr):
+        self,
+        particle_data: List[List[Particle]],
+        resolution: float,
+        Q_vector: List[complex],
+        self_corr: bool,
+    ) -> Tuple[List[List[float]], List[List[float]]]:
         event_weights = self.__compute_particle_weights(particle_data)
         u_vectors = self.__compute_u_vectors(particle_data)
         flow_values, psi_values = self.__compute_flow_particles(
-            particle_data, event_weights, Q_vector, u_vectors, resolution, self_corr)
+            particle_data,
+            event_weights,
+            Q_vector,
+            u_vectors,
+            resolution,
+            self_corr,
+        )
 
         return flow_values, psi_values
 
     def __calculate_flow_event_average(
-            self,
-            particle_data,
-            flow_particle_list,
-            psi_particle_list):
+        self,
+        particle_data: List[List[Particle]],
+        flow_particle_list: List[List[float]],
+        psi_particle_list: List[List[float]],
+    ) -> Tuple[float, float, float, float]:
         # compute the integrated flow
-        number_of_particles = 0
+        number_of_particles = 0.0
         flowvalue = 0.0
         flowvalue_squared = 0.0
         psivalue = 0.0
         psivalue_squared = 0.0
         for event in range(len(flow_particle_list)):
             for particle in range(len(flow_particle_list[event])):
-                weight = 1. if np.isnan(
-                    particle_data[event][particle].weight) else particle_data[event][particle].weight
+                weight = (
+                    1.0
+                    if np.isnan(particle_data[event][particle].weight)
+                    else particle_data[event][particle].weight
+                )
                 number_of_particles += weight
                 flowvalue += flow_particle_list[event][particle] * weight
-                flowvalue_squared += flow_particle_list[event][particle]**2. * weight**2.
+                flowvalue_squared += (
+                    flow_particle_list[event][particle] ** 2.0 * weight**2.0
+                )
                 psivalue += psi_particle_list[event][particle] * weight
-                psivalue_squared += psi_particle_list[event][particle]**2. * weight**2.
+                psivalue_squared += (
+                    psi_particle_list[event][particle] ** 2.0 * weight**2.0
+                )
 
         vn_integrated = 0.0
         Psi_n = 0.0
@@ -380,20 +431,21 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
         else:
             vn_integrated = flowvalue / number_of_particles
             Psi_n = psivalue / number_of_particles
-            vn_squared = flowvalue_squared / number_of_particles**2.
-            Psi_n_squared = psivalue_squared / number_of_particles**2.
-            std_deviation = np.sqrt(vn_integrated**2. - vn_squared)
-            std_deviation_Psi = np.sqrt(Psi_n**2. - Psi_n_squared)
+            vn_squared = flowvalue_squared / number_of_particles**2.0
+            Psi_n_squared = psivalue_squared / number_of_particles**2.0
+            std_deviation = np.sqrt(vn_integrated**2.0 - vn_squared)
+            std_deviation_Psi = np.sqrt(Psi_n**2.0 - Psi_n_squared)
             sigma = std_deviation / np.sqrt(number_of_particles)
             sigma_Psi = std_deviation_Psi / np.sqrt(number_of_particles)
 
         return vn_integrated, sigma, Psi_n, sigma_Psi
 
     def integrated_flow(
-            self,
-            particle_data,
-            particle_data_event_plane,
-            self_corr=True):
+        self,
+        particle_data: List[List[Particle]],
+        particle_data_event_plane: List[List[Particle]],
+        self_corr: bool = True,
+    ) -> Tuple[float, float, float, float]:
         """
         Compute the integrated flow.
 
@@ -415,21 +467,25 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
             and the corresponding uncertainty.
         """
         if not isinstance(self_corr, bool):
-            raise TypeError('self_corr has to be bool')
+            raise TypeError("self_corr has to be bool")
         resolution, Q_vector = self.__calculate_reference(
-            particle_data_event_plane)
+            particle_data_event_plane
+        )
         flow_values, psi_values = self.__calculate_particle_flow(
-            particle_data, resolution, Q_vector, self_corr)
+            particle_data, resolution, Q_vector, self_corr
+        )
         return self.__calculate_flow_event_average(
-            particle_data, flow_values, psi_values)
+            particle_data, flow_values, psi_values
+        )
 
     def differential_flow(
-            self,
-            particle_data,
-            bins,
-            flow_as_function_of,
-            particle_data_event_plane,
-            self_corr=True):
+        self,
+        particle_data: List[List[Particle]],
+        bins: Union[np.ndarray, List[float]],
+        flow_as_function_of: str,
+        particle_data_event_plane: List[List[Particle]],
+        self_corr: bool = True,
+    ) -> List[Tuple[float, float, float, float]]:
         """
         Compute the differential flow.
 
@@ -441,7 +497,7 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
             Bins used for the differential flow calculation.
         flow_as_function_of : str
             Variable on which the flow is calculated
-            ("pt", "rapidity", or "pseudorapidity").
+            ("pT", "rapidity", or "pseudorapidity").
         particle_data_event_plane : list
             List of particle data for the event plane calculation.
         self_corr : bool, optional
@@ -456,14 +512,15 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
             and the corresponding uncertainty.
         """
         if not isinstance(self_corr, bool):
-            raise TypeError('self_corr has to be bool')
+            raise TypeError("self_corr has to be bool")
         if not isinstance(bins, (list, np.ndarray)):
-            raise TypeError('bins has to be list or np.ndarray')
+            raise TypeError("bins has to be list or np.ndarray")
         if not isinstance(flow_as_function_of, str):
-            raise TypeError('flow_as_function_of is not a string')
-        if flow_as_function_of not in ["pt", "rapidity", "pseudorapidity"]:
+            raise TypeError("flow_as_function_of is not a string")
+        if flow_as_function_of not in ["pT", "rapidity", "pseudorapidity"]:
             raise ValueError(
-                "flow_as_function_of must be either 'pt', 'rapidity', 'pseudorapidity'")
+                "flow_as_function_of must be either 'pT', 'rapidity', 'pseudorapidity'"
+            )
 
         particles_bin = []
         for bin in range(len(bins) - 1):
@@ -471,11 +528,11 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
             for event in range(len(particle_data)):
                 particles_event = []
                 for particle in particle_data[event]:
-                    val = 0.
-                    if flow_as_function_of == "pt":
-                        val = particle.pt_abs()
+                    val = 0.0
+                    if flow_as_function_of == "pT":
+                        val = particle.pT_abs()
                     elif flow_as_function_of == "rapidity":
-                        val = particle.momentum_rapidity_Y()
+                        val = particle.rapidity()
                     elif flow_as_function_of == "pseudorapidity":
                         val = particle.pseudorapidity()
                     if val >= bins[bin] and val < bins[bin + 1]:
@@ -484,14 +541,18 @@ class EventPlaneFlow(FlowInterface.FlowInterface):
             particles_bin.extend([events_bin])
 
         resolution, Q_vector = self.__calculate_reference(
-            particle_data_event_plane)
+            particle_data_event_plane
+        )
 
         flow_bin = []
         for bin in range(len(bins) - 1):
             flow_values, psi_values = self.__calculate_particle_flow(
-                particles_bin[bin], resolution, Q_vector, self_corr)
+                particles_bin[bin], resolution, Q_vector, self_corr
+            )
             flow_bin.append(
                 self.__calculate_flow_event_average(
-                    particle_data, flow_values, psi_values))
+                    particle_data, flow_values, psi_values
+                )
+            )
 
         return flow_bin
