@@ -9,6 +9,7 @@
 
 import numpy as np
 import random as rd
+from typing import Dict, List
 
 
 class GenerateFlow:
@@ -38,27 +39,27 @@ class GenerateFlow:
     generate_dummy_JETSCAPE_file:
         Generate a dummy JETSCAPE file with random particle momenta resulting in
         the same flow for all transverse momenta.
-    generate_dummy_JETSCAPE_file_realistic_pt_shape:
+    generate_dummy_JETSCAPE_file_realistic_pT_shape:
         Generate a dummy JETSCAPE file with particles having flow with a more
         realistic transverse momentum distribution.
     generate_dummy_JETSCAPE_file_multi_particle_correlations:
         Generate a dummy JETSCAPE file with random particle momenta resulting in
         the same flow for all transverse momenta. With this function multi-particle
         correlations can be introduced.
-    generate_dummy_JETSCAPE_file_realistic_pt_shape_multi_particle_correlations:
+    generate_dummy_JETSCAPE_file_realistic_pT_shape_multi_particle_correlations:
         Generate a dummy JETSCAPE file with particles having flow with a more
         realistic transverse momentum distribution. With this function multi-particle
         correlations can be introduced.
     generate_dummy_OSCAR_file:
         Generate dummy flow data in OSCAR format.
-    generate_dummy_OSCAR_file_realistic_pt_shape:
+    generate_dummy_OSCAR_file_realistic_pT_shape:
         Generate a dummy OSCAR2013 file with particles having flow with a more
         realistic transverse momentum distribution.
     generate_dummy_OSCAR_file_multi_particle_correlations:
         Generate a dummy OSCAR2013 file with random particle momenta resulting in
         the same flow for all transverse momenta. With this function multi-particle
         correlations can be introduced.
-    generate_dummy_OSCAR_file_realistic_pt_shape_multi_particle_correlations:
+    generate_dummy_OSCAR_file_realistic_pT_shape_multi_particle_correlations:
         Generate a dummy OSCAR2013 file with particles having flow with a more
         realistic transverse momentum distribution. With this function multi-particle
         correlations can be introduced.
@@ -82,8 +83,8 @@ class GenerateFlow:
 
     Notes
     -----
-    If you use the :py:meth:`generate_dummy_JETSCAPE_file_realistic_pt_shape` or
-    :py:meth:`generate_dummy_OSCAR_file_realistic_pt_shape` keep in mind, that the
+    If you use the :py:meth:`generate_dummy_JETSCAPE_file_realistic_pT_shape` or
+    :py:meth:`generate_dummy_OSCAR_file_realistic_pT_shape` keep in mind, that the
     flow values given during construction are used for the saturation value of the
     flow at large transverse momentum. They do **not** reflect the value of the
     integrated flow.
@@ -93,30 +94,41 @@ class GenerateFlow:
     `event generator <https://www.physik.uni-bielefeld.de/~borghini/Software/flow_analysis_codes/generator.cc>`__.
     """
 
-    def __init__(self, *vn, **vn_kwargs):
+    def __init__(self, *vn: float, **vn_kwargs: float) -> None:
+        self.n_: np.ndarray = np.array([])
+        self.vn_: np.ndarray = np.array([])
+
         if not vn and not vn_kwargs:
-            self.n_ = self.vn_ = None
+            self.n_ = self.vn_ = np.array([])
         else:
             try:
-                vn_dictionary = {int(kw.lstrip('v'))
-                                     : val for kw, val in vn_kwargs.items()}
+                vn_dictionary: Dict[int, float] = {
+                    int(kw.lstrip("v")): val for kw, val in vn_kwargs.items()
+                }
             except ValueError:
                 raise TypeError(
                     "Input must have the form of a dictionary with 'vN' "
-                    "where N is an integer.")
+                    "where N is an integer."
+                )
             vn_dictionary.update(
-                (k, v) for k, v in enumerate(
-                    vn, start=2) if v is not None and v != 0.)
-            kwargs = dict(dtype=float, count=len(vn_dictionary))
-            self.n_ = np.fromiter(vn_dictionary.keys(), **kwargs)
-            self.vn_ = np.fromiter(vn_dictionary.values(), **kwargs)
+                (k, v)
+                for k, v in enumerate(vn, start=2)
+                if v is not None and v != 0.0
+            )
 
-        self.phi_ = []
-        self.px_ = []
-        self.py_ = []
-        self.pz_ = []
+            self.n_ = np.fromiter(
+                vn_dictionary.keys(), dtype=int, count=len(vn_dictionary)
+            )
+            self.vn_ = np.fromiter(
+                vn_dictionary.values(), dtype=float, count=len(vn_dictionary)
+            )
 
-    def __distribution_function(self, phi):
+        self.phi_: List[float] = []
+        self.px_: List[float] = []
+        self.py_: List[float] = []
+        self.pz_: List[float] = []
+
+    def __distribution_function(self, phi: float) -> float:
         """
         Calculate the distribution function value for a given angle.
 
@@ -135,15 +147,15 @@ class GenerateFlow:
             - This method calculates the value of the distribution function
             based on the harmonic components (vn) and angles (phi).
         """
-        f = 1. / (2. * np.pi)
+        f = 1.0 / (2.0 * np.pi)
         f_harmonic = 1.0
 
         for term in range(len(self.n_)):
-            f_harmonic += 2. * self.vn_[term] * np.cos(self.n_[term] * phi)
+            f_harmonic += 2.0 * self.vn_[term] * np.cos(self.n_[term] * phi)
 
         return f * f_harmonic
 
-    def __sample_angles(self, multiplicity):
+    def __sample_angles(self, multiplicity: int) -> None:
         """
         Sample angles for a given multiplicity according to a given distribution.
 
@@ -156,19 +168,19 @@ class GenerateFlow:
         -------
             None
         """
-        f_max = (1. + 2. * self.vn_.sum()) / (2. * np.pi)
-        phi = []
+        f_max = (1.0 + 2.0 * self.vn_.sum()) / (2.0 * np.pi)
+        phi: List[float] = []
 
         while len(phi) < multiplicity:
-            random_phi = rd.uniform(0., 2. * np.pi)
-            random_dist_val = rd.uniform(0., f_max)
+            random_phi = rd.uniform(0.0, 2.0 * np.pi)
+            random_dist_val = rd.uniform(0.0, f_max)
 
             if random_dist_val <= self.__distribution_function(random_phi):
                 phi.append(random_phi)
 
         self.phi_ = phi
 
-    def __thermal_distribution(self, temperature, mass):
+    def __thermal_distribution(self, temperature: float, mass: float) -> float:
         """
         Calculate the momentum magnitude from a thermal distribution.
 
@@ -184,51 +196,54 @@ class GenerateFlow:
             momentum_radial: float
                 The magnitude of the momentum.
         """
-        momentum_radial = 0
-        energy = 0.
+        momentum_radial = 0.0
+        energy = 0.0
         if temperature > 0.6 * mass:
             while True:
-                rand_values = [rd.uniform(0., 1.) for _ in range(3)]
+                rand_values = [rd.uniform(0.0, 1.0) for _ in range(3)]
                 if all(rand_values):
                     rand_a, rand_b, rand_c = rand_values
                     momentum_radial = temperature * (rand_a + rand_b + rand_c)
-                    energy = np.sqrt(momentum_radial ** 2. + mass ** 2.)
-                    if rd.uniform(
-                            0., 1.) < np.exp(
-                            (momentum_radial - energy) / temperature):
+                    energy = np.sqrt(momentum_radial**2.0 + mass**2.0)
+                    if rd.uniform(0.0, 1.0) < np.exp(
+                        (momentum_radial - energy) / temperature
+                    ):
                         break
         else:
             while True:
-                r0 = rd.uniform(0., 1.)
-                I1 = mass ** 2.
-                I2 = 2. * mass * temperature
-                I3 = 2. * temperature ** 2.
+                r0 = rd.uniform(0.0, 1.0)
+                I1 = mass**2.0
+                I2 = 2.0 * mass * temperature
+                I3 = 2.0 * temperature**2.0
                 Itot = I1 + I2 + I3
                 K = 0.0
                 if r0 < I1 / Itot:
-                    r1 = rd.uniform(0., 1.)
-                    if r1 != 0.:
+                    r1 = rd.uniform(0.0, 1.0)
+                    if r1 != 0.0:
                         K = -temperature * np.log(r1)
                 elif r0 < (I1 + I2) / Itot:
-                    r1, r2 = rd.uniform(0., 1.), rd.uniform(0., 1.)
-                    if r1 != 0. and r2 != 0.:
+                    r1, r2 = rd.uniform(0.0, 1.0), rd.uniform(0.0, 1.0)
+                    if r1 != 0.0 and r2 != 0.0:
                         K = -temperature * np.log(r1 * r2)
                 else:
-                    r1, r2, r3 = rd.uniform(
-                        0., 1.), rd.uniform(
-                        0., 1.), rd.uniform(
-                        0., 1.)
-                    if r1 != 0. and r2 != 0. and r3 != 0.:
+                    r1, r2, r3 = (
+                        rd.uniform(0.0, 1.0),
+                        rd.uniform(0.0, 1.0),
+                        rd.uniform(0.0, 1.0),
+                    )
+                    if r1 != 0.0 and r2 != 0.0 and r3 != 0.0:
                         K = -temperature * np.log(r1 * r2 * r3)
 
                 energy = K + mass
                 momentum_radial = np.sqrt((energy + mass) * (energy - mass))
-                if rd.uniform(0., 1.) < momentum_radial / energy:
+                if rd.uniform(0.0, 1.0) < momentum_radial / energy:
                     break
 
         return momentum_radial
 
-    def __sample_momenta_thermal(self, multiplicity, temperature, mass):
+    def __sample_momenta_thermal(
+        self, multiplicity: int, temperature: float, mass: float
+    ) -> None:
         """
         Sample momenta for a given multiplicity, temperature, and mass from
         a thermal distribution function.
@@ -247,37 +262,33 @@ class GenerateFlow:
             None
         """
         p_abs = [
-            self.__thermal_distribution(
-                temperature,
-                mass) for _ in range(multiplicity)]
+            self.__thermal_distribution(temperature, mass)
+            for _ in range(multiplicity)
+        ]
 
         # compute the directions
         azimuths = [self.phi_[p] for p in range(len(p_abs))]
-        costheta_values = [rd.uniform(-1., 1.) for _ in range(len(p_abs))]
+        costheta_values = [rd.uniform(-1.0, 1.0) for _ in range(len(p_abs))]
         polar_values = np.arccos(costheta_values)
 
         # convert to cartesian
         px = [
-            p_abs[p] *
-            np.sin(
-                polar_values[p]) *
-            np.cos(
-                azimuths[p]) for p in range(
-                len(p_abs))]
+            p_abs[p] * np.sin(polar_values[p]) * np.cos(azimuths[p])
+            for p in range(len(p_abs))
+        ]
         py = [
-            p_abs[p] *
-            np.sin(
-                polar_values[p]) *
-            np.sin(
-                azimuths[p]) for p in range(
-                len(p_abs))]
+            p_abs[p] * np.sin(polar_values[p]) * np.sin(azimuths[p])
+            for p in range(len(p_abs))
+        ]
         pz = [p_abs[p] * np.cos(polar_values[p]) for p in range(len(p_abs))]
 
         self.px_ = px
         self.py_ = py
         self.pz_ = pz
 
-    def __artificial_pT_distribution(self, pT, pTmin, pT0, pT1, T0):
+    def __artificial_pT_distribution(
+        self, pT: float, pTmin: float, pT0: float, pT1: float, T0: float
+    ) -> float:
         """
         Calculate the artificial pT distribution.
 
@@ -309,18 +320,20 @@ class GenerateFlow:
 
         Momenta (pT, pT0...) are given in GeV.
         """
-        value = 0.
+        value = 0.0
         if pT < pTmin:
-            value = 0.
+            value = 0.0
         elif pT <= pT0:
-            value = 1.
+            value = 1.0
         elif pT <= pT1:
             value = np.exp(-(pT - pT0) / T0)
         else:
-            value = np.exp(-(pT1 - pT0) / T0) * (pT1 / pT)**7
+            value = np.exp(-(pT1 - pT0) / T0) * (pT1 / pT) ** 7
         return value
 
-    def __artificial_flow_pT_shape(self, pT, pT0_bis, pT_sat, vn_sat):
+    def __artificial_flow_pT_shape(
+        self, pT: float, pT0_bis: float, pT_sat: float, vn_sat: float
+    ) -> float:
         """
         Calculate the artificial flow pT shape.
 
@@ -349,22 +362,25 @@ class GenerateFlow:
         - Quadratic rise again for pT_sat - pT0 < pT < pT_sat.
         - Constant value vn_sat for pT > pT_sat.
         """
-        value = 0.
+        value = 0.0
         vn_pT0_bis = 0.5 * vn_sat * pT0_bis / (pT_sat - pT0_bis)
 
         if pT < pT0_bis:
-            value = (pT / pT0_bis)**2. * vn_pT0_bis
+            value = (pT / pT0_bis) ** 2.0 * vn_pT0_bis
         elif pT < (pT_sat - pT0_bis):
-            value = (vn_sat - 2. * vn_pT0_bis) * (pT - pT0_bis) / \
-                (pT_sat - 2. * pT0_bis) + vn_pT0_bis
+            value = (vn_sat - 2.0 * vn_pT0_bis) * (pT - pT0_bis) / (
+                pT_sat - 2.0 * pT0_bis
+            ) + vn_pT0_bis
         elif pT < pT_sat:
-            value = vn_sat - ((pT - pT_sat) / pT0_bis)**2. * vn_pT0_bis
+            value = vn_sat - ((pT - pT_sat) / pT0_bis) ** 2.0 * vn_pT0_bis
         else:
             value = vn_sat
 
         return value
 
-    def __distribution_function_pT_differential(self, phi, vn_pt_list):
+    def __distribution_function_pT_differential(
+        self, phi: float, vn_pT_list: List[float]
+    ) -> float:
         """
         Calculates the pT-differential distribution function for a given
         azimuthal angle.
@@ -373,7 +389,7 @@ class GenerateFlow:
         ----------
         phi : float
             The azimuthal angle at which to calculate the distribution function.
-        vn_pt_list : list
+        vn_pT_list : list
             A list of flow harmonics vn for different transverse momenta pT.
 
         Returns
@@ -384,16 +400,17 @@ class GenerateFlow:
         f_harmonic = 1.0
         f_norm = 1.0
         for term in range(len(self.n_)):
-            f_harmonic += 2. * vn_pt_list[term] * np.cos(self.n_[term] * phi)
-            f_norm += 2. * np.abs(vn_pt_list[term])
+            f_harmonic += 2.0 * vn_pT_list[term] * np.cos(self.n_[term] * phi)
+            f_norm += 2.0 * np.abs(vn_pT_list[term])
 
         return f_harmonic / f_norm
 
     def __create_k_particle_correlations(
-            self,
-            multiplicity,
-            k_particle_correlation,
-            correlation_fraction):
+        self,
+        multiplicity: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+    ) -> None:
         """
         Generate momentum components with k-particle correlations.
 
@@ -412,9 +429,9 @@ class GenerateFlow:
             Updates the internal arrays (px_, py_, pz_) with the
             generated correlated momenta.
         """
-        px = []
-        py = []
-        pz = []
+        px: List[float] = []
+        py: List[float] = []
+        pz: List[float] = []
         idx = 0
         while len(px) <= multiplicity:
             if rd.random() <= correlation_fraction:
@@ -433,8 +450,9 @@ class GenerateFlow:
         self.py_ = py
         self.pz_ = pz
 
-    def __generate_flow_realistic_pt_distribution(
-            self, multiplicity, reaction_plane_angle):
+    def __generate_flow_realistic_pT_distribution(
+        self, multiplicity: int, reaction_plane_angle: float
+    ) -> None:
         pTmax = 4.5
         pTmin = 0.1
         pT0 = 0.5
@@ -445,32 +463,38 @@ class GenerateFlow:
         pT_sat = 1.5
 
         for particle in range(multiplicity):
-            pT_chosen = 0.
+            pT_chosen = 0.0
             need_momentum = True
             while need_momentum:
                 pT = rd.random() * pTmax
-                if rd.random() < self.__artificial_pT_distribution(pT, pTmin, pT0, pT1, T0):
+                if rd.random() < self.__artificial_pT_distribution(
+                    pT, pTmin, pT0, pT1, T0
+                ):
                     pT_chosen = pT
                     need_momentum = False
 
-            vn_pt_list = []
+            vn_pT_list = []
             for harmonic in range(len(self.n_)):
-                vn_pt_list.append(
+                vn_pT_list.append(
                     self.__artificial_flow_pT_shape(
-                        pT, pT0_bis, pT_sat, self.vn_[harmonic]))
+                        pT, pT0_bis, pT_sat, self.vn_[harmonic]
+                    )
+                )
 
-            phi_chosen = 0.
+            phi_chosen = 0.0
             need_angle = True
             while need_angle:
-                phi = 2. * np.pi * rd.random()
-                if rd.random() < self.__distribution_function_pT_differential(phi, vn_pt_list):
+                phi = 2.0 * np.pi * rd.random()
+                if rd.random() < self.__distribution_function_pT_differential(
+                    phi, vn_pT_list
+                ):
                     phi_chosen = phi
                     need_angle = False
 
-            if reaction_plane_angle != 0.:
+            if reaction_plane_angle != 0.0:
                 phi_chosen += reaction_plane_angle
-                if phi_chosen > 2. * np.pi:
-                    phi_chosen -= 2. * np.pi
+                if phi_chosen > 2.0 * np.pi:
+                    phi_chosen -= 2.0 * np.pi
 
             # convert to cartesian
             self.px_.append(pT_chosen * np.cos(phi_chosen))
@@ -478,11 +502,8 @@ class GenerateFlow:
             self.pz_.append(rd.uniform(-1, 1) * pTmax)
 
     def generate_dummy_JETSCAPE_file(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed):
+        self, output_path: str, number_events: int, multiplicity: int, seed: int
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with random particle momenta resulting in
         the same flow for all transverse momenta.
@@ -516,7 +537,8 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
 
         rd.seed(seed)
         temperature = 0.140
@@ -526,43 +548,49 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n")
+                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n"
+            )
 
             for event in range(number_events):
                 self.__sample_angles(multiplicity)
                 self.__sample_momenta_thermal(multiplicity, temperature, mass)
 
                 output.write(
-                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n")
+                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n"
+                )
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%d %d %d %g %g %g %g\n" %
-                        (particle,
-                         pdg,
-                         status,
-                         energy,
-                         self.px_[particle],
+                        "%d %d %d %g %g %g %g\n"
+                        % (
+                            particle,
+                            pdg,
+                            status,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
-                            self.pz_[particle]))
+                            self.pz_[particle],
+                        )
+                    )
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
 
             output.write("#	sigmaGen	0.0	sigmaErr	0.0")
 
-    def generate_dummy_JETSCAPE_file_realistic_pt_shape(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed,
-            random_reaction_plane=True):
+    def generate_dummy_JETSCAPE_file_realistic_pT_shape(
+        self,
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with particles having flow with a more
         realistic transverse momentum distribution.
@@ -597,7 +625,8 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
 
         rd.seed(seed)
         mass = 0.138
@@ -606,34 +635,40 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n")
+                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n"
+            )
 
             for event in range(number_events):
                 if random_reaction_plane:
-                    reaction_plane_angle = 2. * np.pi * rd.random()
+                    reaction_plane_angle = 2.0 * np.pi * rd.random()
                 else:
-                    reaction_plane_angle = 0.
-                self.__generate_flow_realistic_pt_distribution(
-                    multiplicity, reaction_plane_angle)
+                    reaction_plane_angle = 0.0
+                self.__generate_flow_realistic_pT_distribution(
+                    multiplicity, reaction_plane_angle
+                )
 
                 output.write(
-                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n")
+                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n"
+                )
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%d %d %d %g %g %g %g\n" %
-                        (particle,
-                         pdg,
-                         status,
-                         energy,
-                         self.px_[particle],
+                        "%d %d %d %g %g %g %g\n"
+                        % (
+                            particle,
+                            pdg,
+                            status,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
-                            self.pz_[particle]))
+                            self.pz_[particle],
+                        )
+                    )
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
@@ -641,13 +676,14 @@ class GenerateFlow:
             output.write("#	sigmaGen	0.0	sigmaErr	0.0")
 
     def generate_dummy_JETSCAPE_file_multi_particle_correlations(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed,
-            k_particle_correlation,
-            correlation_fraction):
+        self,
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with random particle momenta resulting in
         the same flow for all transverse momenta. A fraction of k-particle
@@ -686,14 +722,15 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
         if not isinstance(k_particle_correlation, int):
             raise TypeError("'k_particle_correlation' is not int")
         if not isinstance(correlation_fraction, float):
             raise TypeError("'correlation_fraction' is not float")
         if k_particle_correlation < 2:
             raise ValueError("'k_particle_correlation' must be at least 2")
-        if correlation_fraction < 0. or correlation_fraction > 1.:
+        if correlation_fraction < 0.0 or correlation_fraction > 1.0:
             raise ValueError("'correlation_fraction' must be between 0 and 1")
 
         rd.seed(seed)
@@ -704,47 +741,54 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n")
+                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n"
+            )
 
             for event in range(number_events):
                 self.__sample_angles(multiplicity)
                 self.__sample_momenta_thermal(multiplicity, temperature, mass)
                 self.__create_k_particle_correlations(
-                    multiplicity, k_particle_correlation, correlation_fraction)
+                    multiplicity, k_particle_correlation, correlation_fraction
+                )
 
                 output.write(
-                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n")
+                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n"
+                )
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%d %d %d %g %g %g %g\n" %
-                        (particle,
-                         pdg,
-                         status,
-                         energy,
-                         self.px_[particle],
+                        "%d %d %d %g %g %g %g\n"
+                        % (
+                            particle,
+                            pdg,
+                            status,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
-                            self.pz_[particle]))
+                            self.pz_[particle],
+                        )
+                    )
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
 
             output.write("#	sigmaGen	0.0	sigmaErr	0.0")
 
-    def generate_dummy_JETSCAPE_file_realistic_pt_shape_multi_particle_correlations(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed,
-            k_particle_correlation,
-            correlation_fraction,
-            random_reaction_plane=True):
+    def generate_dummy_JETSCAPE_file_realistic_pT_shape_multi_particle_correlations(
+        self,
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with particles having flow with a more
         realistic transverse momentum distribution. A fraction of k-particle
@@ -784,14 +828,15 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
         if not isinstance(k_particle_correlation, int):
             raise TypeError("'k_particle_correlation' is not int")
         if not isinstance(correlation_fraction, float):
             raise TypeError("'correlation_fraction' is not float")
         if k_particle_correlation < 2:
             raise ValueError("'k_particle_correlation' must be at least 2")
-        if correlation_fraction < 0. or correlation_fraction > 1.:
+        if correlation_fraction < 0.0 or correlation_fraction > 1.0:
             raise ValueError("'correlation_fraction' must be between 0 and 1")
 
         rd.seed(seed)
@@ -801,36 +846,43 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n")
+                "#	JETSCAPE_FINAL_STATE	v2	|	N	pid	status	E	Px	Py	Pz\n"
+            )
 
             for event in range(number_events):
                 if random_reaction_plane:
-                    reaction_plane_angle = 2. * np.pi * rd.random()
+                    reaction_plane_angle = 2.0 * np.pi * rd.random()
                 else:
-                    reaction_plane_angle = 0.
-                self.__generate_flow_realistic_pt_distribution(
-                    multiplicity, reaction_plane_angle)
+                    reaction_plane_angle = 0.0
+                self.__generate_flow_realistic_pT_distribution(
+                    multiplicity, reaction_plane_angle
+                )
                 self.__create_k_particle_correlations(
-                    multiplicity, k_particle_correlation, correlation_fraction)
+                    multiplicity, k_particle_correlation, correlation_fraction
+                )
 
                 output.write(
-                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n")
+                    f"# Event {event + 1} weight 1 EPangle 0 N_hadrons {multiplicity}\n"
+                )
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%d %d %d %g %g %g %g\n" %
-                        (particle,
-                         pdg,
-                         status,
-                         energy,
-                         self.px_[particle],
+                        "%d %d %d %g %g %g %g\n"
+                        % (
+                            particle,
+                            pdg,
+                            status,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
-                            self.pz_[particle]))
+                            self.pz_[particle],
+                        )
+                    )
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
@@ -838,11 +890,8 @@ class GenerateFlow:
             output.write("#	sigmaGen	0.0	sigmaErr	0.0")
 
     def generate_dummy_OSCAR_file(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed):
+        self, output_path: str, number_events: int, multiplicity: int, seed: int
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with random particle momenta
         resulting in the same flow for all transverse momenta.
@@ -876,7 +925,8 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
 
         rd.seed(seed)
         temperature = 0.140
@@ -885,9 +935,11 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n")
+                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n"
+            )
             output.write(
-                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n")
+                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n"
+            )
             output.write("# SMASH-2.2\n")
 
             for event in range(number_events):
@@ -897,40 +949,45 @@ class GenerateFlow:
                 output.write(f"# event {event} out {multiplicity}\n")
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%g %g %g %g %g %g %g %g %g %d %d %d\n" %
-                        (1,
-                         1,
-                         1,
-                         1,
-                         mass,
-                         energy,
-                         self.px_[particle],
+                        "%g %g %g %g %g %g %g %g %g %d %d %d\n"
+                        % (
+                            1,
+                            1,
+                            1,
+                            1,
+                            mass,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
                             self.pz_[particle],
                             pdg,
                             particle,
-                            1))
+                            1,
+                        )
+                    )
 
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
 
                 output.write(
-                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n")
+                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n"
+                )
 
-    def generate_dummy_OSCAR_file_realistic_pt_shape(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed,
-            random_reaction_plane=True):
+    def generate_dummy_OSCAR_file_realistic_pT_shape(
+        self,
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with particles having flow with a more
         realistic transverse momentum distribution.
@@ -965,7 +1022,8 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
 
         rd.seed(seed)
         mass = 0.138
@@ -973,57 +1031,65 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n")
+                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n"
+            )
             output.write(
-                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n")
+                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n"
+            )
             output.write("# SMASH-2.2\n")
 
             for event in range(number_events):
                 if random_reaction_plane:
-                    reaction_plane_angle = 2. * np.pi * rd.random()
+                    reaction_plane_angle = 2.0 * np.pi * rd.random()
                 else:
-                    reaction_plane_angle = 0.
-                self.__generate_flow_realistic_pt_distribution(
-                    multiplicity, reaction_plane_angle)
+                    reaction_plane_angle = 0.0
+                self.__generate_flow_realistic_pT_distribution(
+                    multiplicity, reaction_plane_angle
+                )
 
                 output.write(f"# event {event} out {multiplicity}\n")
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%g %g %g %g %g %g %g %g %g %d %d %d\n" %
-                        (1,
-                         1,
-                         1,
-                         1,
-                         mass,
-                         energy,
-                         self.px_[particle],
+                        "%g %g %g %g %g %g %g %g %g %d %d %d\n"
+                        % (
+                            1,
+                            1,
+                            1,
+                            1,
+                            mass,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
                             self.pz_[particle],
                             pdg,
                             particle,
-                            1))
+                            1,
+                        )
+                    )
 
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
 
                 output.write(
-                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n")
+                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n"
+                )
 
     def generate_dummy_OSCAR_file_multi_particle_correlations(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed,
-            k_particle_correlation,
-            correlation_fraction):
+        self,
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with random particle momenta
         resulting in the same flow for all transverse momenta. A fraction of
@@ -1062,14 +1128,15 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
         if not isinstance(k_particle_correlation, int):
             raise TypeError("'k_particle_correlation' is not int")
         if not isinstance(correlation_fraction, float):
             raise TypeError("'correlation_fraction' is not float")
         if k_particle_correlation < 2:
             raise ValueError("'k_particle_correlation' must be at least 2")
-        if correlation_fraction < 0. or correlation_fraction > 1.:
+        if correlation_fraction < 0.0 or correlation_fraction > 1.0:
             raise ValueError("'correlation_fraction' must be between 0 and 1")
 
         rd.seed(seed)
@@ -1079,56 +1146,64 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n")
+                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n"
+            )
             output.write(
-                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n")
+                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n"
+            )
             output.write("# SMASH-2.2\n")
 
             for event in range(number_events):
                 self.__sample_angles(multiplicity)
                 self.__sample_momenta_thermal(multiplicity, temperature, mass)
                 self.__create_k_particle_correlations(
-                    multiplicity, k_particle_correlation, correlation_fraction)
+                    multiplicity, k_particle_correlation, correlation_fraction
+                )
 
                 output.write(f"# event {event} out {multiplicity}\n")
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%g %g %g %g %g %g %g %g %g %d %d %d\n" %
-                        (1,
-                         1,
-                         1,
-                         1,
-                         mass,
-                         energy,
-                         self.px_[particle],
+                        "%g %g %g %g %g %g %g %g %g %d %d %d\n"
+                        % (
+                            1,
+                            1,
+                            1,
+                            1,
+                            mass,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
                             self.pz_[particle],
                             pdg,
                             particle,
-                            1))
+                            1,
+                        )
+                    )
 
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
 
                 output.write(
-                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n")
+                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n"
+                )
 
-    def generate_dummy_OSCAR_file_realistic_pt_shape_multi_particle_correlations(
-            self,
-            output_path,
-            number_events,
-            multiplicity,
-            seed,
-            k_particle_correlation,
-            correlation_fraction,
-            random_reaction_plane=True):
+    def generate_dummy_OSCAR_file_realistic_pT_shape_multi_particle_correlations(
+        self,
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with particles having flow with a more
         realistic transverse momentum distribution. A fraction of k-particle
@@ -1168,14 +1243,15 @@ class GenerateFlow:
             raise TypeError("'seed' is not int")
         if number_events < 1 or multiplicity < 1:
             raise ValueError(
-                "'number_events' and/or 'multiplicity' must be larger than 0")
+                "'number_events' and/or 'multiplicity' must be larger than 0"
+            )
         if not isinstance(k_particle_correlation, int):
             raise TypeError("'k_particle_correlation' is not int")
         if not isinstance(correlation_fraction, float):
             raise TypeError("'correlation_fraction' is not float")
         if k_particle_correlation < 2:
             raise ValueError("'k_particle_correlation' must be at least 2")
-        if correlation_fraction < 0. or correlation_fraction > 1.:
+        if correlation_fraction < 0.0 or correlation_fraction > 1.0:
             raise ValueError("'correlation_fraction' must be between 0 and 1")
 
         rd.seed(seed)
@@ -1184,47 +1260,55 @@ class GenerateFlow:
 
         with open(output_path, "w") as output:
             output.write(
-                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n")
+                "#!OSCAR2013 particle_lists t x y z mass p0 px py pz pdg ID charge\n"
+            )
             output.write(
-                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n")
+                "# Units: fm fm fm fm GeV GeV GeV GeV GeV none none e\n"
+            )
             output.write("# SMASH-2.2\n")
 
             for event in range(number_events):
                 if random_reaction_plane:
-                    reaction_plane_angle = 2. * np.pi * rd.random()
+                    reaction_plane_angle = 2.0 * np.pi * rd.random()
                 else:
-                    reaction_plane_angle = 0.
-                self.__generate_flow_realistic_pt_distribution(
-                    multiplicity, reaction_plane_angle)
+                    reaction_plane_angle = 0.0
+                self.__generate_flow_realistic_pT_distribution(
+                    multiplicity, reaction_plane_angle
+                )
                 self.__create_k_particle_correlations(
-                    multiplicity, k_particle_correlation, correlation_fraction)
+                    multiplicity, k_particle_correlation, correlation_fraction
+                )
 
                 output.write(f"# event {event} out {multiplicity}\n")
                 for particle in range(multiplicity):
                     energy = np.sqrt(
-                        self.px_[particle] ** 2.
-                        + self.py_[particle] ** 2.
-                        + self.pz_[particle] ** 2.
-                        + mass ** 2.
+                        self.px_[particle] ** 2.0
+                        + self.py_[particle] ** 2.0
+                        + self.pz_[particle] ** 2.0
+                        + mass**2.0
                     )
                     output.write(
-                        "%g %g %g %g %g %g %g %g %g %d %d %d\n" %
-                        (1,
-                         1,
-                         1,
-                         1,
-                         mass,
-                         energy,
-                         self.px_[particle],
+                        "%g %g %g %g %g %g %g %g %g %d %d %d\n"
+                        % (
+                            1,
+                            1,
+                            1,
+                            1,
+                            mass,
+                            energy,
+                            self.px_[particle],
                             self.py_[particle],
                             self.pz_[particle],
                             pdg,
                             particle,
-                            1))
+                            1,
+                        )
+                    )
 
                 self.px_.clear()
                 self.py_.clear()
                 self.pz_.clear()
 
                 output.write(
-                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n")
+                    f"# event {event} end 0 impact  -1.000 scattering_projectile_target no\n"
+                )
