@@ -8,9 +8,11 @@
 # ===================================================
 
 from sparkx.flow import FlowInterface
+from sparkx.Particle import Particle
 import numpy as np
 import warnings
 import random as rd
+from typing import Optional, Union, List, Any
 
 rd.seed(42)
 
@@ -99,75 +101,83 @@ class PCAFlow(FlowInterface.FlowInterface):
     Raises
     ------
     TypeError
-        If n, alpha, or number_subcalc is not an integer.
+        If :code:`n`, :code:`alpha`, or :code:`number_subcalc` is not an integer.
     ValueError
-        If n is not a positive integer, alpha is less than 1,
-        or number_subcalc is less than 2.
+        If :code:`n` is not a positive integer, alpha is less than 1,
+        or :code:`number_subcalc` is less than 2.
     """
 
-    def __init__(self, n=2, alpha=2, number_subcalc=4):
+    def __init__(
+        self, n: int = 2, alpha: int = 2, number_subcalc: int = 4
+    ) -> None:
         # flow harmonic to compute
         if not isinstance(n, int):
-            raise TypeError('n has to be int')
+            raise TypeError("n has to be int")
         elif n <= 0:
             raise ValueError(
-                'n-th harmonic with value n<=0 can not be computed')
+                "n-th harmonic with value n<=0 can not be computed"
+            )
         else:
             self.n_ = n
 
         # order in sub-leading flow up to which the flow is computed
         if not isinstance(alpha, int):
-            raise TypeError('alpha has to be int')
+            raise TypeError("alpha has to be int")
         elif alpha < 1:
-            raise ValueError('alpha has to be >= 1')
+            raise ValueError("alpha has to be >= 1")
         else:
             self.alpha_ = alpha
 
         # number of sub-calculations to estimate the error of the flow
         if not isinstance(number_subcalc, int):
-            raise TypeError('number_subcalc has to be int')
+            raise TypeError("number_subcalc has to be int")
         elif number_subcalc < 2:
-            raise ValueError('number_subcalc has to be >= 2')
+            raise ValueError("number_subcalc has to be >= 2")
         else:
             self.number_subcalc_ = number_subcalc
 
-        self.number_events_ = None
-        self.subcalc_counter_ = 0
+        self.number_events_: Optional[int] = None
+        self.subcalc_counter_: int = 0
 
-        self.normalization_ = None
-        self.bin_multiplicity_total_ = None
-        self.sigma_multiplicity_total_ = []
-        self.number_events_subcalc_ = None
-        self.QnRe_total_ = None
-        self.QnIm_total_ = None
-        self.SigmaQnReSub_total_ = None
-        self.SigmaQnImSub_total_ = None
-        self.VnDelta_total_ = None
-        self.SigmaVnDelta_total_ = None
+        self.normalization_: Optional[np.ndarray] = None
+        self.bin_multiplicity_total_: Optional[np.ndarray] = None
+        self.sigma_multiplicity_total_: List[np.ndarray] = []
+        self.number_events_subcalc_: Optional[np.ndarray] = None
+        self.QnRe_total_: Optional[np.ndarray] = None
+        self.QnIm_total_: Optional[np.ndarray] = None
+        self.SigmaQnReSub_total_: Optional[np.ndarray] = None
+        self.SigmaQnImSub_total_: Optional[np.ndarray] = None
+        self.VnDelta_total_: Optional[np.ndarray] = None
+        self.SigmaVnDelta_total_: Optional[np.ndarray] = None
 
-        self.FlowSubCalc_ = None
-        self.Flow_ = None
-        self.FlowUncertainty_ = None
+        self.FlowSubCalc_: Optional[np.ndarray] = None
+        self.Flow_: Optional[np.ndarray] = None
+        self.FlowUncertainty_: Optional[np.ndarray] = None
 
-        self.Pearson_r_ = None
-        self.Pearson_r_uncertainty_ = None
+        self.Pearson_r_: Optional[np.ndarray] = None
+        self.Pearson_r_uncertainty_: Optional[np.ndarray] = None
 
-    def integrated_flow(self, particle_data):
+    def integrated_flow(self, particle_data: List[List[Particle]]) -> None:
         warnings.warn("Integrated flow is not supported for PCAFlow")
         return None
 
-    def __compute_normalization(self, bins):
+    def __compute_normalization(
+        self, bins: Union[List[float], np.ndarray]
+    ) -> None:
         self.normalization_ = np.zeros((len(bins) - 1))
         bin_widths = np.diff(bins)
         for bin in range(len(bins) - 1):
-            self.normalization_[bin] = 1. / (2. * np.pi * bin_widths[bin])**2.
+            self.normalization_[bin] = (
+                1.0 / (2.0 * np.pi * bin_widths[bin]) ** 2.0
+            )
 
     def __update_event(
-            self,
-            event_data,
-            bins,
-            flow_as_function_of,
-            event_number):
+        self,
+        event_data: List[Particle],
+        bins: Union[List[float], np.ndarray],
+        flow_as_function_of: str,
+        event_number: int,
+    ) -> None:
         """
         Update the anisotropic flow calculations based on particle data from
         a single event.
@@ -179,7 +189,7 @@ class PCAFlow(FlowInterface.FlowInterface):
         bins : list or np.ndarray
             Bins used for the flow calculation.
         flow_as_function_of : str
-            Variable on which the flow is calculated ("pt", "rapidity"
+            Variable on which the flow is calculated ("pT", "rapidity"
             or "pseudorapidity").
         event_number : int
             Index of the current event.
@@ -188,6 +198,47 @@ class PCAFlow(FlowInterface.FlowInterface):
         -------
         None
         """
+        if self.normalization_ is None:
+            raise TypeError(
+                "'normalization_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.bin_multiplicity_total_ is None:
+            raise TypeError(
+                "'bin_multiplicity_total_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.sigma_multiplicity_total_ is None:
+            raise TypeError(
+                "'sigma_multiplicity_total_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.number_events_subcalc_ is None:
+            raise TypeError(
+                "'number_events_subcalc_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.QnRe_total_ is None:
+            raise TypeError(
+                "'QnRe_total_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.QnIm_total_ is None:
+            raise TypeError(
+                "'QnIm_total_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.SigmaQnReSub_total_ is None:
+            raise TypeError(
+                "'SigmaQnReSub_total_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.SigmaQnImSub_total_ is None:
+            raise TypeError(
+                "'SigmaQnImSub_total_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.VnDelta_total_ is None:
+            raise TypeError(
+                "'VnDelta_total_' is None. It must be initialized before calling the '__update_event' function."
+            )
+        if self.SigmaVnDelta_total_ is None:
+            raise TypeError(
+                "'SigmaVnDelta' is None. It must be initialized before calling the '__update_event' function."
+            )
+
         number_bins = len(bins) - 1
         bin_multiplicity_event = np.zeros(number_bins)
         QnRe = np.zeros(number_bins)
@@ -196,7 +247,8 @@ class PCAFlow(FlowInterface.FlowInterface):
         SigmaQnImSubTot = np.zeros((number_bins, self.number_subcalc_))
         VnDelta_event = np.zeros((number_bins, number_bins))
         SigmaVnDelta_event = np.zeros(
-            (number_bins, number_bins, self.number_subcalc_))
+            (number_bins, number_bins, self.number_subcalc_)
+        )
         if event_number == 0:
             self.bin_multiplicity_total_ = np.zeros(number_bins)
             for i in range(self.number_subcalc_):
@@ -205,27 +257,34 @@ class PCAFlow(FlowInterface.FlowInterface):
             self.QnRe_total_ = np.zeros(number_bins)
             self.QnIm_total_ = np.zeros(number_bins)
             self.SigmaQnReSub_total_ = np.zeros(
-                (number_bins, self.number_subcalc_))
+                (number_bins, self.number_subcalc_)
+            )
             self.SigmaQnImSub_total_ = np.zeros(
-                (number_bins, self.number_subcalc_))
+                (number_bins, self.number_subcalc_)
+            )
             self.VnDelta_total_ = np.zeros((number_bins, number_bins))
             self.SigmaVnDelta_total_ = np.zeros(
-                (number_bins, number_bins, self.number_subcalc_))
+                (number_bins, number_bins, self.number_subcalc_)
+            )
 
         # update the sub-calculation counter if needed
-        number_events_subcalc = self.number_events_ // self.number_subcalc_
-        if (event_number % number_events_subcalc == 0) and (event_number != 0) and (
-                self.subcalc_counter_ < self.number_subcalc_ - 1):
+        if self.number_events_ is not None:
+            number_events_subcalc = self.number_events_ // self.number_subcalc_
+        if (
+            (event_number % number_events_subcalc == 0)
+            and (event_number != 0)
+            and (self.subcalc_counter_ < self.number_subcalc_ - 1)
+        ):
             self.subcalc_counter_ += 1
 
-        random_reaction_plane = rd.random() * 2. * np.pi
+        random_reaction_plane = rd.random() * 2.0 * np.pi
 
         # loop over all event particles and compute the flow vectors in the bins
         for particle in event_data:
-            if flow_as_function_of == "pt":
-                val = particle.pt_abs()
+            if flow_as_function_of == "pT":
+                val = particle.pT_abs()
             elif flow_as_function_of == "rapidity":
-                val = particle.momentum_rapidity_Y()
+                val = particle.rapidity()
             elif flow_as_function_of == "pseudorapidity":
                 val = particle.pseudorapidity()
 
@@ -237,33 +296,44 @@ class PCAFlow(FlowInterface.FlowInterface):
 
             phi = particle.phi() + random_reaction_plane
 
-            bin_multiplicity_event[bin_index] += 1.
+            bin_multiplicity_event[bin_index] += 1.0
             QnRe[bin_index] += np.cos(self.n_ * phi)
             QnIm[bin_index] += np.sin(self.n_ * phi)
 
             SigmaQnReSubTot[bin_index][self.subcalc_counter_] += np.cos(
-                self.n_ * phi)
+                self.n_ * phi
+            )
             SigmaQnImSubTot[bin_index][self.subcalc_counter_] += np.sin(
-                self.n_ * phi)
+                self.n_ * phi
+            )
 
         # compute the covariance matrix
         for a in range(number_bins):
             for b in range(number_bins):
                 if a == b:  # correct with multiplicity term
-                    VnDelta_event[a][b] += (QnRe[a] * QnRe[b] + QnIm[a] * QnIm[b] -
-                                            bin_multiplicity_event[a]) / self.normalization_[a]
+                    VnDelta_event[a][b] += (
+                        QnRe[a] * QnRe[b]
+                        + QnIm[a] * QnIm[b]
+                        - bin_multiplicity_event[a]
+                    ) / self.normalization_[a]
                     SigmaVnDelta_event[a][b][self.subcalc_counter_] += (
-                        QnRe[a] * QnRe[b] + QnIm[a] * QnIm[b] - bin_multiplicity_event[a]) / self.normalization_[a]
+                        QnRe[a] * QnRe[b]
+                        + QnIm[a] * QnIm[b]
+                        - bin_multiplicity_event[a]
+                    ) / self.normalization_[a]
                 else:
-                    VnDelta_event[a][b] += (QnRe[a] * QnRe[b] +
-                                            QnIm[a] * QnIm[b]) / self.normalization_[a]
+                    VnDelta_event[a][b] += (
+                        QnRe[a] * QnRe[b] + QnIm[a] * QnIm[b]
+                    ) / self.normalization_[a]
                     SigmaVnDelta_event[a][b][self.subcalc_counter_] += (
-                        QnRe[a] * QnRe[b] + QnIm[a] * QnIm[b]) / self.normalization_[a]
+                        QnRe[a] * QnRe[b] + QnIm[a] * QnIm[b]
+                    ) / self.normalization_[a]
 
         # update the class members
         self.bin_multiplicity_total_ += bin_multiplicity_event
         self.sigma_multiplicity_total_[
-            self.subcalc_counter_] += bin_multiplicity_event
+            self.subcalc_counter_
+        ] += bin_multiplicity_event
         self.number_events_subcalc_[self.subcalc_counter_] += 1
         self.QnRe_total_ += QnRe
         self.QnIm_total_ += QnIm
@@ -272,7 +342,7 @@ class PCAFlow(FlowInterface.FlowInterface):
         self.VnDelta_total_ += VnDelta_event
         self.SigmaVnDelta_total_ += SigmaVnDelta_event
 
-    def __compute_flow_PCA(self, bins):
+    def __compute_flow_PCA(self, bins: Union[List[float], np.ndarray]) -> None:
         """
         Perform Principal Component Analysis (PCA) to compute the anisotropic
         flow.
@@ -286,23 +356,75 @@ class PCAFlow(FlowInterface.FlowInterface):
         -------
         None
         """
+        if self.VnDelta_total_ is None:
+            raise TypeError(
+                "'VnDelta_total_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.QnRe_total_ is None:
+            raise TypeError(
+                "'QnRe_total_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.QnIm_total_ is None:
+            raise TypeError(
+                "'QnIm_total_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.SigmaVnDelta_total_ is None:
+            raise TypeError(
+                "'SigmaVnDelta_total_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.number_events_subcalc_ is None:
+            raise TypeError(
+                "'number_events_subcalc_' is None. It must be initialize before calling the '__compute_flow_PCA' function."
+            )
+        if self.SigmaQnReSub_total_ is None:
+            raise TypeError(
+                "'SigmaQnReSub_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.SigmaQnImSub_total_ is None:
+            raise TypeError(
+                "'SigmaQnImSub_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.number_events_ is None:
+            raise TypeError(
+                "'number_events_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.bin_multiplicity_total_ is None:
+            raise TypeError(
+                "'bin_multiplicity_total_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+        if self.normalization_ is None:
+            raise TypeError(
+                "'normalization_' is None. It must be initialized before calling the '__compute_flow_PCA' function."
+            )
+
         number_bins = len(bins) - 1
 
         for a in range(number_bins):
             for b in range(number_bins):
                 self.VnDelta_total_[a][b] /= self.number_events_
-                self.VnDelta_total_[a][b] -= (self.QnRe_total_[a] * self.QnRe_total_[
-                                              b] + self.QnIm_total_[a] * self.QnIm_total_[b]) / self.number_events_**2.
+                self.VnDelta_total_[a][b] -= (
+                    self.QnRe_total_[a] * self.QnRe_total_[b]
+                    + self.QnIm_total_[a] * self.QnIm_total_[b]
+                ) / self.number_events_**2.0
 
                 for sub in range(self.number_subcalc_):
-                    self.SigmaVnDelta_total_[
-                        a][b][sub] /= self.number_events_subcalc_[sub]
-                    self.SigmaVnDelta_total_[a][b][sub] -= (self.SigmaQnReSub_total_[a][sub] * self.SigmaQnReSub_total_[b][sub] + self.SigmaQnImSub_total_[
-                                                            a][sub] * self.SigmaQnImSub_total_[b][sub]) / (self.number_events_subcalc_[sub]**2. * self.normalization_[a])
+                    self.SigmaVnDelta_total_[a][b][
+                        sub
+                    ] /= self.number_events_subcalc_[sub]
+                    self.SigmaVnDelta_total_[a][b][sub] -= (
+                        self.SigmaQnReSub_total_[a][sub]
+                        * self.SigmaQnReSub_total_[b][sub]
+                        + self.SigmaQnImSub_total_[a][sub]
+                        * self.SigmaQnImSub_total_[b][sub]
+                    ) / (
+                        self.number_events_subcalc_[sub] ** 2.0
+                        * self.normalization_[a]
+                    )
 
         # perform sub calculations for error estimation
         self.FlowSubCalc_ = np.zeros(
-            (self.number_subcalc_, number_bins, self.alpha_))
+            (self.number_subcalc_, number_bins, self.alpha_)
+        )
         for sub in range(self.number_subcalc_):
             VnDelta_local = np.zeros((number_bins, number_bins))
             for a in range(number_bins):
@@ -320,11 +442,16 @@ class PCAFlow(FlowInterface.FlowInterface):
                 for alpha in range(self.alpha_):
                     eval = sorted_eigenvalues[alpha]
                     evec = sorted_eigenvectors[alpha]
-                    if (eval >= 0.) and (
-                            self.sigma_multiplicity_total_[sub][bin] > 0.):
+                    if (eval >= 0.0) and (
+                        self.sigma_multiplicity_total_[sub][bin] > 0.0
+                    ):
                         # compute the flow here for the subcalc
-                        self.FlowSubCalc_[sub][bin][alpha] = np.sqrt(
-                            eval) * evec[bin] * self.number_events_subcalc_[sub] / self.sigma_multiplicity_total_[sub][bin]
+                        self.FlowSubCalc_[sub][bin][alpha] = (
+                            np.sqrt(eval)
+                            * evec[bin]
+                            * self.number_events_subcalc_[sub]
+                            / self.sigma_multiplicity_total_[sub][bin]
+                        )
                     else:
                         self.FlowSubCalc_[sub][bin][alpha] = None
 
@@ -346,13 +473,19 @@ class PCAFlow(FlowInterface.FlowInterface):
             for alpha in range(self.alpha_):
                 eval = sorted_eigenvalues[alpha]
                 evec = sorted_eigenvectors[alpha]
-                if (eval >= 0.) and (self.bin_multiplicity_total_[bin] > 0.):
-                    self.Flow_[bin][alpha] = np.sqrt(
-                        eval) * evec[bin] * self.number_events_ / self.bin_multiplicity_total_[bin]
+                if (eval >= 0.0) and (self.bin_multiplicity_total_[bin] > 0.0):
+                    self.Flow_[bin][alpha] = (
+                        np.sqrt(eval)
+                        * evec[bin]
+                        * self.number_events_
+                        / self.bin_multiplicity_total_[bin]
+                    )
                 else:
                     self.Flow_[bin][alpha] = None
 
-    def __compute_uncertainty(self, bins):
+    def __compute_uncertainty(
+        self, bins: Union[List[float], np.ndarray]
+    ) -> None:
         """
         Compute the uncertainty of the anisotropic flow for each bin and
         order alpha.
@@ -366,6 +499,15 @@ class PCAFlow(FlowInterface.FlowInterface):
         -------
         None
         """
+        if self.FlowSubCalc_ is None:
+            raise TypeError(
+                "'FlowSubCalc_' is None. It must be initialized before calling the '__compute_uncertainty' function."
+            )
+        if self.Flow_ is None:
+            raise TypeError(
+                "'Flow_' is None. It must be initialized before calling the '__compute_uncertainty' function."
+            )
+
         number_bins = len(bins) - 1
         self.FlowUncertainty_ = np.zeros((number_bins, self.alpha_))
 
@@ -373,14 +515,17 @@ class PCAFlow(FlowInterface.FlowInterface):
         for sub in range(self.number_subcalc_):
             for bin in range(number_bins):
                 for alpha in range(self.alpha_):
-                    self.FlowUncertainty_[bin][alpha] += (self.FlowSubCalc_[sub][bin][alpha] - self.Flow_[
-                                                          bin][alpha])**2. / (self.number_subcalc_ - 1)
+                    self.FlowUncertainty_[bin][alpha] += (
+                        self.FlowSubCalc_[sub][bin][alpha]
+                        - self.Flow_[bin][alpha]
+                    ) ** 2.0 / (self.number_subcalc_ - 1)
 
         # take the sqrt to obtain the standard deviation
         for bin in range(number_bins):
             for alpha in range(self.alpha_):
                 self.FlowUncertainty_[bin][alpha] = np.sqrt(
-                    self.FlowUncertainty_[bin][alpha])
+                    self.FlowUncertainty_[bin][alpha]
+                )
 
         # Conservative ansatz: If the error cannot be computed
         # (e.g., negative eigenvalue), set the flow magnitude to NaN
@@ -389,7 +534,9 @@ class PCAFlow(FlowInterface.FlowInterface):
                 if np.isnan(self.FlowUncertainty_[bin][alpha]):
                     self.Flow_[bin][alpha] = np.nan
 
-    def __compute_factorization_breaking(self, bins):
+    def __compute_factorization_breaking(
+        self, bins: Union[List[float], np.ndarray]
+    ) -> None:
         """
         Compute the factorization breaking using the Pearson correlation
         coefficient Eq. (12), Ref. [1].
@@ -403,13 +550,23 @@ class PCAFlow(FlowInterface.FlowInterface):
         -------
         None
         """
+        if self.VnDelta_total_ is None:
+            raise TypeError(
+                "'VnDelta_total_' is None. It must be initialized before calling the '__compute_factorization_breaking' function."
+            )
+        if self.SigmaVnDelta_total_ is None:
+            raise TypeError(
+                "'SigmaVnDelta_total_' is None. It must be initialized before calling the '__compute_factorization_breaking' function."
+            )
+
         # compute the Pearson coefficient Eq. (12), Ref. [1]
         number_bins = len(bins) - 1
         r = np.zeros((number_bins, number_bins))
         for a in range(number_bins):
             for b in range(number_bins):
-                r[a][b] = self.VnDelta_total_[a][b] / \
-                    np.sqrt(self.VnDelta_total_[a][a] * self.VnDelta_total_[b][b])
+                r[a][b] = self.VnDelta_total_[a][b] / np.sqrt(
+                    self.VnDelta_total_[a][a] * self.VnDelta_total_[b][b]
+                )
 
         self.Pearson_r_ = r
 
@@ -418,16 +575,21 @@ class PCAFlow(FlowInterface.FlowInterface):
         for sub in range(self.number_subcalc_):
             for a in range(number_bins):
                 for b in range(number_bins):
-                    r_sub[a][b][sub] = self.SigmaVnDelta_total_[a][b][sub] / np.sqrt(
-                        self.SigmaVnDelta_total_[a][a][sub] * self.SigmaVnDelta_total_[b][b][sub])
+                    r_sub[a][b][sub] = self.SigmaVnDelta_total_[a][b][
+                        sub
+                    ] / np.sqrt(
+                        self.SigmaVnDelta_total_[a][a][sub]
+                        * self.SigmaVnDelta_total_[b][b][sub]
+                    )
 
         r_err = np.zeros((number_bins, number_bins))
         # sum the squared deviation from the mean
         for sub in range(self.number_subcalc_):
             for a in range(number_bins):
                 for b in range(number_bins):
-                    r_err[a][b] += (r_sub[a][b][sub] - r[a][b]
-                                    )**2. / (self.number_subcalc_ - 1)
+                    r_err[a][b] += (r_sub[a][b][sub] - r[a][b]) ** 2.0 / (
+                        self.number_subcalc_ - 1
+                    )
 
         # take the sqrt to obtain the standard deviation
         for a in range(number_bins):
@@ -436,7 +598,12 @@ class PCAFlow(FlowInterface.FlowInterface):
 
         self.Pearson_r_uncertainty_ = r_err
 
-    def differential_flow(self, particle_data, bins, flow_as_function_of):
+    def differential_flow(
+        self,
+        particle_data: List[List[Particle]],
+        bins: Union[List[float], np.ndarray],
+        flow_as_function_of: str,
+    ) -> List[Optional[np.ndarray]]:
         """
         Compute the differential flow.
 
@@ -447,7 +614,7 @@ class PCAFlow(FlowInterface.FlowInterface):
         bins : list or np.ndarray
             Bins used for the differential flow calculation.
         flow_as_function_of : str
-            Variable on which the flow is calculated ("pt", "rapidity"
+            Variable on which the flow is calculated ("pT", "rapidity"
             or "pseudorapidity").
 
         Returns
@@ -461,26 +628,25 @@ class PCAFlow(FlowInterface.FlowInterface):
 
         Notes
         -----
-        - The flow or the uncertainty can be accessed by: `function_return[bin][alpha]`
-        - If a bin has no events or the uncertainty could not be computed, the corresponding element in the result list is set to `np.nan`.
+        - The flow or the uncertainty can be accessed by: :code:`function_return[bin][alpha]`
+        - If a bin has no events or the uncertainty could not be computed, the corresponding element in the result list is set to :code:`np.nan`.
         """
         if not isinstance(bins, (list, np.ndarray)):
-            raise TypeError('bins has to be list or np.ndarray')
+            raise TypeError("bins has to be list or np.ndarray")
         if not isinstance(flow_as_function_of, str):
-            raise TypeError('flow_as_function_of is not a string')
+            raise TypeError("flow_as_function_of is not a string")
         if flow_as_function_of not in ["pt", "rapidity", "pseudorapidity"]:
             raise ValueError(
-                "flow_as_function_of must be either 'pt', 'rapidity', 'pseudorapidity'")
+                "flow_as_function_of must be either 'pT', 'rapidity', 'pseudorapidity'"
+            )
 
         self.__compute_normalization(bins)
         self.number_events_ = len(particle_data)
 
         for event in range(self.number_events_):
             self.__update_event(
-                particle_data[event],
-                bins,
-                flow_as_function_of,
-                event)
+                particle_data[event], bins, flow_as_function_of, event
+            )
 
         self.__compute_flow_PCA(bins)
         self.__compute_uncertainty(bins)
@@ -488,7 +654,7 @@ class PCAFlow(FlowInterface.FlowInterface):
 
         return [self.Flow_, self.FlowUncertainty_]
 
-    def Pearson_correlation(self):
+    def Pearson_correlation(self) -> List[Optional[np.ndarray]]:
         """
         Retrieve the Pearson correlation coefficient and its uncertainty
         (Eq. (12), Ref. [1]).
