@@ -19,9 +19,8 @@ from sparkx.Particle import Particle
 @pytest.fixture
 def oscar_file_path():
     return os.path.join(
-        os.path.dirname(__file__),
-        'test_files',
-        'particle_lists_extended.oscar')
+        os.path.dirname(__file__), "test_files", "particle_lists_extended.oscar"
+    )
 
 
 @pytest.fixture
@@ -34,23 +33,24 @@ def oscar_particle_objects_list(oscar_file_path):
 @pytest.fixture
 def test_Lattice3D_instance():
     lattice = Lattice3D(
-        x_min=-3.,
-        x_max=3.,
-        y_min=-3.,
-        y_max=3.,
-        z_min=-3.,
-        z_max=3.,
+        x_min=-3.0,
+        x_max=3.0,
+        y_min=-3.0,
+        y_max=3.0,
+        z_min=-3.0,
+        z_max=3.0,
         num_points_x=15,
         num_points_y=15,
-        num_points_z=15)
-    lattice.set_value_nearest_neighbor(0., 1., 0., 1.)
-    lattice.set_value_nearest_neighbor(0., -1., 0., 1.)
+        num_points_z=15,
+    )
+    lattice.set_value_nearest_neighbor(0.0, 1.0, 0.0, 1.0)
+    lattice.set_value_nearest_neighbor(0.0, -1.0, 0.0, 1.0)
     return lattice
 
 
 def test_EventCharacteristics_initialization(
-        oscar_particle_objects_list,
-        test_Lattice3D_instance):
+    oscar_particle_objects_list, test_Lattice3D_instance
+):
     eve_char = EventCharacteristics(oscar_particle_objects_list)
     assert eve_char.has_lattice_ == False
     assert eve_char.event_data_ == oscar_particle_objects_list
@@ -60,42 +60,42 @@ def test_EventCharacteristics_initialization(
         EventCharacteristics(add_number_to_list)
 
     with pytest.raises(TypeError):
-        add_string = 'test'
+        add_string = "test"
         EventCharacteristics(add_string)
 
     eve_char = EventCharacteristics(test_Lattice3D_instance)
     assert eve_char.has_lattice_
     assert np.array_equal(
-        eve_char.event_data_.grid_,
-        test_Lattice3D_instance.grid_)
+        eve_char.event_data_.grid_, test_Lattice3D_instance.grid_
+    )
 
 
 @pytest.fixture
 def fake_particle_list_eccentricity():
     particle_list = []
     p1 = Particle()
-    p1.x = 0.
-    p1.y = 1.
-    p1.z = 0.
-    p1.E = 1.
-    p1.px = 1.
-    p1.py = 0.
-    p1.pz = 0.
-    p1.charge = 1.
-    p1.strangeness = 1.
-    p1.baryon_number = 1.
+    p1.x = 0.0
+    p1.y = 1.0
+    p1.z = 0.0
+    p1.E = 1.0
+    p1.px = 1.0
+    p1.py = 0.0
+    p1.pz = 0.0
+    p1.charge = 1.0
+    p1.strangeness = 1.0
+    p1.baryon_number = 1.0
     particle_list.append(p1)
     p2 = Particle()
-    p2.E = 1.
-    p2.x = 0.
-    p2.y = -1.
-    p1.z = 0.
-    p2.px = 1.
-    p2.py = 0.
-    p2.pz = 0.
-    p2.charge = 0.
-    p2.strangeness = 0.
-    p2.baryon_number = 0.
+    p2.E = 1.0
+    p2.x = 0.0
+    p2.y = -1.0
+    p1.z = 0.0
+    p2.px = 1.0
+    p2.py = 0.0
+    p2.pz = 0.0
+    p2.charge = 0.0
+    p2.strangeness = 0.0
+    p2.baryon_number = 0.0
     particle_list.append(p2)
     return particle_list
 
@@ -104,32 +104,32 @@ def fake_particle_list_eccentricity():
 def initial_particles():
     particle_list = []
     p1 = Particle()
-    p1.t = 1.
-    p1.x = 0.
-    p1.y = 1.
-    p1.z = 0.
-    p1.E = 1.
-    p1.mass = 1.
-    p1.px = 1.
-    p1.py = 0.
-    p1.pz = 0.
-    p1.charge = 1.
-    p1.strangeness = 1.
-    p1.baryon_number = 1.
+    p1.t = 1.0
+    p1.x = 0.0
+    p1.y = 1.0
+    p1.z = 0.0
+    p1.E = 1.0
+    p1.mass = 1.0
+    p1.px = 1.0
+    p1.py = 0.0
+    p1.pz = 0.0
+    p1.charge = 1.0
+    p1.strangeness = 1.0
+    p1.baryon_number = 1.0
     particle_list.append(p1)
     p2 = Particle()
-    p2.E = 1.
-    p2.mass = 1.
-    p2.t = 1.
-    p2.x = 0.
-    p2.y = -1.
-    p2.z = 0.
-    p2.px = 1.
-    p2.py = 0.
-    p2.pz = 0.
-    p2.charge = 0.
-    p2.strangeness = 0.
-    p2.baryon_number = 0.
+    p2.E = 1.0
+    p2.mass = 1.0
+    p2.t = 1.0
+    p2.x = 0.0
+    p2.y = -1.0
+    p2.z = 0.0
+    p2.px = 1.0
+    p2.py = 0.0
+    p2.pz = 0.0
+    p2.charge = 0.0
+    p2.strangeness = 0.0
+    p2.baryon_number = 0.0
     particle_list.append(p2)
     return particle_list
 
@@ -137,10 +137,10 @@ def initial_particles():
 def test_eccentricity_from_particles(fake_particle_list_eccentricity):
     eve_char = EventCharacteristics(fake_particle_list_eccentricity)
     eps2 = eve_char.eccentricity(2, weight_quantity="number")
-    assert np.allclose(eps2, (1. - 0j), atol=1e-10)
+    assert np.allclose(eps2, (1.0 - 0j), atol=1e-10)
 
     eps3 = eve_char.eccentricity(3, weight_quantity="number")
-    assert np.allclose(eps3, (0. - 0j), atol=1e-10)
+    assert np.allclose(eps3, (0.0 - 0j), atol=1e-10)
 
     with pytest.raises(ValueError):
         eve_char.eccentricity(0, weight_quantity="number")
@@ -152,10 +152,10 @@ def test_eccentricity_from_particles(fake_particle_list_eccentricity):
 def test_eccentricity_from_lattice(test_Lattice3D_instance):
     eve_char = EventCharacteristics(test_Lattice3D_instance)
     eps2 = eve_char.eccentricity(2, weight_quantity="number")
-    np.allclose(eps2, (1. - 0j), atol=1e-10)
+    np.allclose(eps2, (1.0 - 0j), atol=1e-10)
 
     eps3 = eve_char.eccentricity(3, weight_quantity="number")
-    np.allclose(eps3, (0. - 0j), atol=1e-10)
+    np.allclose(eps3, (0.0 - 0j), atol=1e-10)
 
     with pytest.raises(ValueError):
         eve_char.eccentricity(0, weight_quantity="number")
@@ -165,19 +165,22 @@ def test_eccentricity_from_lattice(test_Lattice3D_instance):
 
 
 def test_eBQS_densities_Milne_from_OSCAR_IC(
-        tmp_path,
-        oscar_particle_objects_list,
-        test_Lattice3D_instance,
-        fake_particle_list_eccentricity,
-        initial_particles):
-    x_min = y_min = z_min = -3.
-    x_max = y_max = z_max = 3.
+    tmp_path,
+    oscar_particle_objects_list,
+    test_Lattice3D_instance,
+    fake_particle_list_eccentricity,
+    initial_particles,
+):
+    x_min = y_min = z_min = -3.0
+    x_max = y_max = z_max = 3.0
     Nx = Ny = Nz = 100
     n_sigma_x = n_sigma_y = n_sigma_z = 2
     sigma_smear = 0.3
+    kernel = "covariant"
     eta_range = [-1, 1, 10]
     output_filename = os.path.join(
-        tmp_path, 'test_eBQS_densities_Milne_from_OSCAR_IC.dat')
+        tmp_path, "test_eBQS_densities_Milne_from_OSCAR_IC.dat"
+    )
 
     # test error if IC_info is not None and string
     with pytest.raises(TypeError):
@@ -198,7 +201,9 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             sigma_smear,
             eta_range,
             output_filename,
-            IC_info=1)
+            kernel,
+            IC_info=1,
+        )
 
     # test error if class is initialized with lattice
     with pytest.raises(TypeError):
@@ -218,7 +223,9 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             eta_range,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     # Check error if x_min, x_max, y_min, y_max, z_min, z_max are not floats
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -237,7 +244,9 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             eta_range,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     # Check error if Nx, Ny, Nz are not positive integers
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -256,7 +265,9 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             eta_range,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
         eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
@@ -274,7 +285,9 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             eta_range,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     # Check error if n_sigma_x, n_sigma_y, n_sigma_z are not floats
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -293,7 +306,9 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             eta_range,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
         eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
@@ -311,7 +326,50 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             eta_range,
-            output_filename)
+            output_filename,
+            kernel,
+        )
+    # Check error if kernel is not a string 'covariant' or 'gaussian'
+    with pytest.raises(TypeError):
+        eve_char = EventCharacteristics(oscar_particle_objects_list)
+        eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            Nx,
+            Ny,
+            Nz,
+            n_sigma_x,
+            n_sigma_y,
+            n_sigma_z,
+            sigma_smear,
+            eta_range,
+            output_filename,
+            4,
+        )
+    with pytest.raises(TypeError):
+        eve_char = EventCharacteristics(oscar_particle_objects_list)
+        eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            Nx,
+            Ny,
+            Nz,
+            n_sigma_x,
+            n_sigma_y,
+            n_sigma_z,
+            sigma_smear,
+            eta_range,
+            output_filename,
+            "invalid",
+        )
     # Check error if eta_range is not a list with 3 float entries
     with pytest.raises(ValueError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -329,20 +387,30 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            [
-                1,
-                2],
-            output_filename)
+            [1, 2],
+            output_filename,
+            kernel,
+        )
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
-        eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(x_min, x_max,
-                                                             y_min, y_max,
-                                                             z_min, z_max,
-                                                             Nx, Ny, Nz,
-                                                             n_sigma_x, n_sigma_y,
-                                                             n_sigma_z, sigma_smear,
-                                                             [1, "invalid", 2],
-                                                             output_filename)
+        eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            Nx,
+            Ny,
+            Nz,
+            n_sigma_x,
+            n_sigma_y,
+            n_sigma_z,
+            sigma_smear,
+            [1, "invalid", 2],
+            output_filename,
+            kernel,
+        )
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
         eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
@@ -360,7 +428,9 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             3,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     # Check error if output_filename is not a string
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -379,27 +449,56 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             eta_range,
-            1)
+            1,
+            kernel,
+        )
 
     # Check error if not enough data for computing tau
     with pytest.raises(ValueError):
         eve_char = EventCharacteristics(fake_particle_list_eccentricity)
         eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
-            x_min, x_max, y_min, y_max, z_min, z_max, Nx, Ny, Nz,
-            n_sigma_x, n_sigma_y, n_sigma_z, sigma_smear, eta_range,
-            output_filename
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            Nx,
+            Ny,
+            Nz,
+            n_sigma_x,
+            n_sigma_y,
+            n_sigma_z,
+            sigma_smear,
+            eta_range,
+            output_filename,
+            kernel,
         )
 
     # Generate densities
     eve_char = EventCharacteristics(initial_particles)
     eve_char.generate_eBQS_densities_Milne_from_OSCAR_IC(
-        x_min, x_max, y_min, y_max, z_min, z_max, Nx, Ny, Nz,
-        n_sigma_x, n_sigma_y, n_sigma_z, sigma_smear, eta_range,
-        output_filename, IC_info="test"
+        x_min,
+        x_max,
+        y_min,
+        y_max,
+        z_min,
+        z_max,
+        Nx,
+        Ny,
+        Nz,
+        n_sigma_x,
+        n_sigma_y,
+        n_sigma_z,
+        sigma_smear,
+        eta_range,
+        output_filename,
+        kernel,
+        IC_info="test",
     )
 
     # Load the generated data
-    with open(output_filename, 'r') as f:
+    with open(output_filename, "r") as f:
         lines = f.readlines()
 
     # Check if the file contains data
@@ -407,39 +506,56 @@ def test_eBQS_densities_Milne_from_OSCAR_IC(
 
     # Check header information if available
     header_info = lines[:5]  # Assuming the header is 5 lines
-    assert all(isinstance(line, str)
-               for line in header_info), "Not all lines of the header are strings"
+    assert all(
+        isinstance(line, str) for line in header_info
+    ), "Not all lines of the header are strings"
 
     # Check the content of the data
     data_lines = lines[5:]  # Assuming data starts from line 6
-    integral = 0.
+    integral = 0.0
     for line in data_lines:
         # Split the line into columns
         values = line.split()
         # Extract relevant information
-        tau, x, y, eta, energy_density, baryon_density, charge_density, strangeness_density = map(
-            float, values)
+        (
+            tau,
+            x,
+            y,
+            eta,
+            energy_density,
+            baryon_density,
+            charge_density,
+            strangeness_density,
+        ) = map(float, values)
         assert x_min <= x <= x_max, "x coordinate out of range"
         assert y_min <= y <= y_max, "y coordinate out of range"
         assert z_min <= eta <= z_max, "eta coordinate out of range"
         assert energy_density >= 0, "Negative energy density"
-        integral += energy_density * tau * \
-            (x_max - x_min) * (y_max - y_min) * (eta_range[1] - eta_range[0]) / (Nx * Ny * eta_range[2])
+        integral += (
+            energy_density
+            * tau
+            * (x_max - x_min)
+            * (y_max - y_min)
+            * (eta_range[1] - eta_range[0])
+            / (Nx * Ny * eta_range[2])
+        )
     assert np.isclose(
-        integral, 2., atol=1e-1), "Energy density integral is not close to 2"
+        integral, 2.0, atol=1e-1
+    ), "Energy density integral is not close to 2"
 
 
 def test_eBQS_densities_Minkowski_from_OSCAR_IC(
-        tmp_path,
-        oscar_particle_objects_list,
-        test_Lattice3D_instance):
-    x_min = y_min = z_min = -5.
-    x_max = y_max = z_max = 5.
+    tmp_path, oscar_particle_objects_list, test_Lattice3D_instance
+):
+    x_min = y_min = z_min = -5.0
+    x_max = y_max = z_max = 5.0
     Nx = Ny = Nz = 50
     n_sigma_x = n_sigma_y = n_sigma_z = 3
     sigma_smear = 0.2
+    kernel = "covariant"
     output_filename = os.path.join(
-        tmp_path, 'test_eBQS_densities_Minkowski_from_OSCAR_IC.dat')
+        tmp_path, "test_eBQS_densities_Minkowski_from_OSCAR_IC.dat"
+    )
 
     # test error if IC_info is not None and string
     with pytest.raises(TypeError):
@@ -459,7 +575,9 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_z,
             sigma_smear,
             output_filename,
-            IC_info=1)
+            kernel,
+            IC_info=1,
+        )
     # test error if class is initialized with lattice
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(test_Lattice3D_instance)
@@ -477,8 +595,10 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            output_filename)
-     # Check error if x_min, x_max, y_min, y_max, z_min, z_max are not floats
+            output_filename,
+            kernel,
+        )
+    # Check error if x_min, x_max, y_min, y_max, z_min, z_max are not floats
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
         eve_char.generate_eBQS_densities_Minkowski_from_OSCAR_IC(
@@ -495,7 +615,9 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     # Check error if Nx, Ny, Nz are not positive integers
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -513,7 +635,9 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
         eve_char.generate_eBQS_densities_Minkowski_from_OSCAR_IC(
@@ -530,7 +654,9 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     # Check error if n_sigma_x, n_sigma_y, n_sigma_z are not positive floats
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -548,7 +674,9 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            output_filename)
+            output_filename,
+            kernel,
+        )
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
         eve_char.generate_eBQS_densities_Minkowski_from_OSCAR_IC(
@@ -565,7 +693,48 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            output_filename)
+            output_filename,
+            kernel,
+        )
+    # Check error if kernel is not a string 'covariant' or 'gaussian'
+    with pytest.raises(TypeError):
+        eve_char = EventCharacteristics(oscar_particle_objects_list)
+        eve_char.generate_eBQS_densities_Minkowski_from_OSCAR_IC(
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            Nx,
+            Ny,
+            Nz,
+            n_sigma_x,
+            n_sigma_y,
+            n_sigma_z,
+            sigma_smear,
+            output_filename,
+            4,
+        )
+    with pytest.raises(TypeError):
+        eve_char = EventCharacteristics(oscar_particle_objects_list)
+        eve_char.generate_eBQS_densities_Minkowski_from_OSCAR_IC(
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            Nx,
+            Ny,
+            Nz,
+            n_sigma_x,
+            n_sigma_y,
+            n_sigma_z,
+            sigma_smear,
+            output_filename,
+            "invalid",
+        )
     # Check error if output_filename is not a string
     with pytest.raises(TypeError):
         eve_char = EventCharacteristics(oscar_particle_objects_list)
@@ -583,4 +752,6 @@ def test_eBQS_densities_Minkowski_from_OSCAR_IC(
             n_sigma_y,
             n_sigma_z,
             sigma_smear,
-            1)
+            1,
+            kernel,
+        )
