@@ -99,7 +99,10 @@ def test_pdg():
         assert q.pdg_valid is False
 
     # Check if a warning is issued
-    with pytest.warns(UserWarning, match=r'The PDG code 99999999 is not valid. All properties extracted from the PDG are set to nan.'):
+    with pytest.warns(
+        UserWarning,
+        match=r"The PDG code 99999999 is not valid. All properties extracted from the PDG are set to nan.",
+    ):
         q.pdg = 99999999
 
 
@@ -218,8 +221,9 @@ def test_initialize_from_array_valid_formats():
     assert p1.charge == 1
 
     format2 = "Oscar2013"
-    array2 = np.array([0.0, 1.0, 2.0, 3.0, 0.138, 4.0,
-                      5.0, 6.0, 7.0, 211, 100, 1])
+    array2 = np.array(
+        [0.0, 1.0, 2.0, 3.0, 0.138, 4.0, 5.0, 6.0, 7.0, 211, 100, 1]
+    )
 
     p2 = Particle(input_format=format2, particle_array=array2)
     assert p2.t == 0.0
@@ -236,8 +240,12 @@ def test_initialize_from_array_valid_formats():
     assert p2.charge == 1
 
     format3 = "Oscar2013Extended"
-    array3 = np.array([0.0, 1.0, 2.0, 3.0, 0.138, 4.0, 5.0, 6.0,
-                      7.0, 211, 100, 1, 0, 0.1, 0.3, 101, 2, 2.5, 321, 2212, 1, -1])
+    # fmt: off
+    array3 = np.array(
+        [0.0, 1.0, 2.0, 3.0, 0.138, 4.0, 5.0, 6.0, 7.0, 211, 100, 1, 0, 0.1,
+            0.3, 101, 2, 2.5, 321, 2212, 1, -1]
+    )
+    # fmt: on
 
     p3 = Particle(input_format=format3, particle_array=array3)
     assert p3.t == 0.0
@@ -291,27 +299,12 @@ def test_initialize_from_array_valid_formats():
     assert p4.strangeness == -1
 
     format5 = "Oscar2013Extended_Photons"
-    array5 = np.array([0.0,
-                       1.0,
-                       2.0,
-                       3.0,
-                       0.138,
-                       4.0,
-                       5.0,
-                       6.0,
-                       7.0,
-                       211,
-                       100,
-                       1,
-                       0,
-                       0.1,
-                       0.3,
-                       101,
-                       2,
-                       2.5,
-                       321,
-                       2212,
-                       0.75])
+    # fmt: off
+    array5 = np.array(
+        [0.0, 1.0, 2.0, 3.0, 0.138, 4.0, 5.0, 6.0, 7.0, 211, 100, 1, 0, 0.1,
+            0.3, 101, 2, 2.5, 321, 2212, 0.75]
+    )
+    # fmt: on
     p5 = Particle(input_format=format5, particle_array=array5)
     assert p5.t == 0.0
     assert p5.x == 1.0
@@ -341,14 +334,20 @@ def test_initialize_from_array_Jetscape_invalid_pdg_warning():
     array1 = np.array([1, 99999999, 27, 4.36557, 3.56147, 0.562961, 2.45727])
 
     # check that a warning is issued
-    with pytest.warns(UserWarning, match=r"The PDG code 99999999 is not known by PDGID, charge could not be computed. Consider setting it by hand."):
+    with pytest.warns(
+        UserWarning,
+        match=r"The PDG code 99999999 is not known by PDGID, charge could not be computed. Consider setting it by hand.",
+    ):
         Particle(input_format=format1, particle_array=array1)
 
 
 def test_initialize_from_array_invalid_format():
-    with pytest.raises(ValueError, match=r"Unsupported input format 'InvalidFormat'"):
-        Particle(input_format="InvalidFormat",
-                 particle_array=np.array([1, 2, 3]))
+    with pytest.raises(
+        ValueError, match=r"Unsupported input format 'InvalidFormat'"
+    ):
+        Particle(
+            input_format="InvalidFormat", particle_array=np.array([1, 2, 3])
+        )
 
 
 def test_initialize_from_array_corrupted_data():
@@ -360,10 +359,13 @@ def test_initialize_from_array_corrupted_data():
 
 def test_initialize_from_array_warning_invalid_pdg():
     format1 = "Oscar2013"
-    array1 = np.array([0.0, 1.0, 2.0, 3.0, 0.138, 4.0,
-                      5.0, 6.0, 7.0, 99999999, 100, 1])
+    array1 = np.array(
+        [0.0, 1.0, 2.0, 3.0, 0.138, 4.0, 5.0, 6.0, 7.0, 99999999, 100, 1]
+    )
 
-    with pytest.warns(UserWarning, match=r"The PDG code 99999999 is not valid."):
+    with pytest.warns(
+        UserWarning, match=r"The PDG code 99999999 is not valid."
+    ):
         p1 = Particle(input_format=format1, particle_array=array1)
 
     assert p1.pdg_valid is False
@@ -434,34 +436,34 @@ def test_angular_momentum_zero_position():
     assert np.array_equal(result, [0.0, 0.0, 0.0])
 
 
-def test_momentum_rapidity_Y_valid_values():
+def test_rapidity_valid_values():
     p = Particle()
     p.E = 10.0
     p.pz = 5.0
 
-    result = p.momentum_rapidity_Y()
+    result = p.rapidity()
 
     expected_result = 0.5 * np.log((10.0 + 5.0) / (10.0 - 5.0))
 
     assert np.isclose(result, expected_result)
 
 
-def test_momentum_rapidity_Y_zero_denominator():
+def test_rapidity_zero_denominator():
     p = Particle()
     p.E = 5.0
     p.pz = 5.0  # E == pz, should add a small positive value
 
-    result = p.momentum_rapidity_Y()
+    result = p.rapidity()
     expected_result = 0.5 * np.log((5.0 + 5.0) / (1e-10))
 
     assert np.isclose(result, expected_result)
 
 
-def test_momentum_rapidity_Y_missing_values():
+def test_rapidity_missing_values():
     p = Particle()
     # Leave some values as np.nan
 
-    result = p.momentum_rapidity_Y()
+    result = p.rapidity()
 
     assert np.isnan(result)
 
@@ -499,33 +501,33 @@ def test_p_abs_zero_momentum():
     assert np.isclose(result, 0.0)
 
 
-def test_pt_abs_valid_values():
+def test_pT_abs_valid_values():
     p = Particle()
     p.px = 1.0
     p.py = 2.0
 
-    result = p.pt_abs()
+    result = p.pT_abs()
 
     expected_result = np.sqrt(1.0**2 + 2.0**2)
 
     assert np.isclose(result, expected_result)
 
 
-def test_pt_abs_missing_values():
+def test_pT_abs_missing_values():
     p = Particle()
     # Leave some values as np.nan
 
-    result = p.pt_abs()
+    result = p.pT_abs()
 
     assert np.isnan(result)
 
 
-def test_pt_abs_zero_transverse_momentum():
+def test_pT_abs_zero_transverse_momentum():
     p = Particle()
     p.px = 0.0
     p.py = 0.0
 
-    result = p.pt_abs()
+    result = p.pT_abs()
 
     assert np.isclose(result, 0.0)
 
@@ -612,8 +614,10 @@ def test_pseudorapidity_valid_values():
 
     result = p.pseudorapidity()
 
-    expected_result = 0.5 * \
-        np.log((np.sqrt(1.0**2 + 2.0**2 + 3.0**2) + 3.0) / (np.sqrt(1.0**2 + 2.0**2 + 3.0**2) - 3.0))
+    expected_result = 0.5 * np.log(
+        (np.sqrt(1.0**2 + 2.0**2 + 3.0**2) + 3.0)
+        / (np.sqrt(1.0**2 + 2.0**2 + 3.0**2) - 3.0)
+    )
 
     assert np.isclose(result, expected_result)
 
@@ -634,39 +638,39 @@ def test_pseudorapidity_zero_momentum():
     p.pz = 1.0
 
     result = p.pseudorapidity()
-    expected_result = 0.5 * np.log(2. / 1e-10)
+    expected_result = 0.5 * np.log(2.0 / 1e-10)
 
     assert np.isclose(result, expected_result)
 
 
-def test_spatial_rapidity_valid_values():
+def test_spacetime_rapidity_valid_values():
     p = Particle()
     p.t = 5.0
     p.z = 3.0
 
-    result = p.spatial_rapidity()
+    result = p.spacetime_rapidity()
 
     expected_result = 0.5 * np.log((5.0 + 3.0) / (5.0 - 3.0))
 
     assert np.isclose(result, expected_result)
 
 
-def test_spatial_rapidity_missing_values():
+def test_spacetime_rapidity_missing_values():
     p = Particle()
     # Leave some values as np.nan
 
-    result = p.spatial_rapidity()
+    result = p.spacetime_rapidity()
 
     assert np.isnan(result)
 
 
-def test_spatial_rapidity_invalid_values():
+def test_spacetime_rapidity_invalid_values():
     p = Particle()
     p.t = 3.0
     p.z = 5.0
 
     with pytest.raises(ValueError, match=r"|z| < t not fulfilled"):
-        p.spatial_rapidity()
+        p.spacetime_rapidity()
 
 
 def test_proper_time_valid_values():
@@ -699,58 +703,137 @@ def test_proper_time_invalid_values():
         p.proper_time()
 
 
-def test_compute_mass_from_energy_momentum_valid_values():
+def test_mass_from_energy_momentum_valid_values():
     p = Particle()
     p.E = 3.0
     p.px = 1.0
     p.py = 2.0
     p.pz = 2.0
 
-    result = p.compute_mass_from_energy_momentum()
+    result = p.mass_from_energy_momentum()
 
     expected_result = np.sqrt(3.0**2 - (1.0**2 + 2.0**2 + 2.0**2))
 
     assert np.isclose(result, expected_result)
 
 
-def test_compute_mass_from_energy_momentum_missing_values():
+def test_compute_mass_from_energy_momentum_invalid_values():
+    p = Particle()
+    p.E = 3.0
+    p.px = 1.0
+    p.py = 2.0
+    p.pz = 5.0
+
+    with pytest.warns(
+        UserWarning,
+        match=r"|E| >= |p| not fulfilled or not within numerical precision! The mass is set to nan.",
+    ):
+        assert np.isnan(p.mass_from_energy_momentum())
+
+
+def test_mass_from_energy_momentum_missing_values():
     p = Particle()
     # Leave some values as np.nan
 
-    result = p.compute_mass_from_energy_momentum()
+    result = p.mass_from_energy_momentum()
 
     assert np.isnan(result)
 
 
-def test_compute_mass_from_energy_momentum_zero_energy():
+def test_mass_from_energy_momentum_zero_energy():
     p = Particle()
     p.E = 0.0
     p.px = 0.0
     p.py = 0.0
     p.pz = 0.0
 
-    result = p.compute_mass_from_energy_momentum()
+    result = p.mass_from_energy_momentum()
 
     assert np.isclose(result, 0.0)
 
 
-def test_compute_charge_from_pdg_valid_values():
+def test_mT_missing_values():
+    p = Particle()
+    # Leave some values as np.nan
+
+    result = p.mT()
+
+    assert np.isnan(result)
+
+
+def test_mT_invalid_values():
+    p = Particle()
+    p.E = 11.2
+    p.pz = 11.3
+
+    with pytest.warns(
+        UserWarning,
+        match=r"|E| >= |pz| not fulfilled or not within numerical precision! The transverse mass is set to nan.",
+    ):
+        assert np.isnan(p.mT())
+
+
+def test_mT_valid_values():
+    p = Particle()
+
+    # Test for zero
+    p.E = 0.0
+    p.pz = 0.0
+    assert np.isclose(p.mT(), 0.0)
+
+    # Test for a few random values
+    for i in range(10):
+        energy = np.random.uniform(5.0, 10.0)
+        p_z = np.random.uniform(0.1, 5.0)
+
+        p.E = energy
+        p.pz = p_z
+
+        result = p.mT()
+        expected_result = np.sqrt(energy**2 - p_z**2)
+
+        assert np.isclose(result, expected_result)
+
+
+def test_charge_from_pdg_valid_values():
     p = Particle()
     p.pdg = 211  # Assuming PDG code for a positive pion
 
-    result = p.compute_charge_from_pdg()
+    result = p.charge_from_pdg()
 
     expected_result = PDGID(211).charge
 
     assert result == expected_result
 
 
-def test_compute_charge_from_pdg_invalid_values():
+def test_charge_from_pdg_invalid_values():
     p = Particle()
     p.pdg_valid = False
     # Leave pdg as an invalid value
 
-    result = p.compute_charge_from_pdg()
+    result = p.charge_from_pdg()
+
+    assert np.isnan(result)
+
+
+def test_is_quark_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 1  # Assuming PDG code for a positive pion
+
+    result = p.is_quark()
+
+    expected_result = PDGID(1).is_quark
+
+    assert result == expected_result
+    assert result == True
+
+
+def test_is_quark_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.is_quark()
 
     assert np.isnan(result)
 
@@ -764,6 +847,28 @@ def test_is_meson_from_pdg_valid_values():
     expected_result = PDGID(211).is_meson
 
     assert result == expected_result
+
+
+def test_is_lepton_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 11  # Assuming PDG code for a positron
+
+    result = p.is_lepton()
+
+    expected_result = PDGID(11).is_lepton
+
+    assert result == expected_result
+    assert result == True
+
+
+def test_is_lepton_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.is_lepton()
+
+    assert np.isnan(result)
 
 
 def test_is_meson_from_pdg_invalid_values():
@@ -818,35 +923,15 @@ def test_is_hadron_from_pdg_invalid_values():
     assert np.isnan(result)
 
 
-def test_is_strange_from_pdg_valid_values():
-    p = Particle()
-    p.pdg = 211  # Assuming PDG code for a positive pion
-
-    result = p.is_strange()
-
-    expected_result = PDGID(211).has_strange
-
-    assert result == expected_result
-
-
-def test_is_strange_from_pdg_invalid_values():
-    p = Particle()
-    p.pdg_valid = False
-    # Leave pdg as an invalid value
-
-    result = p.is_strange()
-
-    assert np.isnan(result)
-
-
 def test_is_heavy_flavor_from_pdg_valid_values():
     p = Particle()
     p.pdg = 211  # Assuming PDG code for a positive pion
 
     result = p.is_heavy_flavor()
 
-    expected_result = PDGID(211).has_charm or PDGID(211).has_bottom\
-        or PDGID(211).has_top
+    expected_result = (
+        PDGID(211).has_charm or PDGID(211).has_bottom or PDGID(211).has_top
+    )
 
     assert result == expected_result
 
@@ -857,6 +942,138 @@ def test_is_heavy_flavor_from_pdg_invalid_values():
     # Leave pdg as an invalid value
 
     result = p.is_heavy_flavor()
+
+    assert np.isnan(result)
+
+
+def test_has_down_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 211  # Assuming PDG code for a positive pion
+
+    result = p.has_down()
+
+    expected_result = PDGID(211).has_down
+
+    assert result == expected_result
+    assert result == True
+
+
+def test_has_down_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.has_down()
+
+    assert np.isnan(result)
+
+
+def test_has_up_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 211  # Assuming PDG code for a positive pion
+
+    result = p.has_up()
+
+    expected_result = PDGID(211).has_up
+
+    assert result == expected_result
+    assert result == True
+
+
+def test_has_up_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.has_up()
+
+    assert np.isnan(result)
+
+
+def test_has_strange_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 111  # Assuming PDG code for a positive pion
+
+    result = p.has_strange()
+
+    expected_result = PDGID(211).has_strange
+
+    assert result == expected_result
+    assert result == False
+
+
+def test_has_strange_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.has_strange()
+
+    assert np.isnan(result)
+
+
+def test_has_charm_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 211  # Assuming PDG code for a positive pion
+
+    result = p.has_charm()
+
+    expected_result = PDGID(211).has_charm
+
+    assert result == expected_result
+    assert result == False
+
+
+def test_has_charm_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.has_charm()
+
+    assert np.isnan(result)
+
+
+def test_has_bottom_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 211  # Assuming PDG code for a positive pion
+
+    result = p.has_bottom()
+
+    expected_result = PDGID(211).has_bottom
+
+    assert result == expected_result
+    assert result == False
+
+
+def test_has_bottom_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.has_bottom()
+
+    assert np.isnan(result)
+
+
+def test_has_top_from_pdg_valid_values():
+    p = Particle()
+    p.pdg = 211  # Assuming PDG code for a positive pion
+
+    result = p.has_top()
+
+    expected_result = PDGID(211).has_top
+
+    assert result == expected_result
+    assert result == False
+
+
+def test_has_top_from_pdg_invalid_values():
+    p = Particle()
+    p.pdg_valid = False
+    # Leave pdg as an invalid value
+
+    result = p.has_top()
 
     assert np.isnan(result)
 
@@ -905,9 +1122,11 @@ def test_spin_degeneracy_from_pdg_invalid_values():
 
 def test_create_quark_from_pdg_valid_values():
     particle_quantity_JETSCAPE = np.array(
-        [0, 1, 11, 2.01351754, 1.30688601, -0.422958786, -0.512249773])
-    particle = Particle(input_format="JETSCAPE",
-                        particle_array=particle_quantity_JETSCAPE)
+        [0, 1, 11, 2.01351754, 1.30688601, -0.422958786, -0.512249773]
+    )
+    particle = Particle(
+        input_format="JETSCAPE", particle_array=particle_quantity_JETSCAPE
+    )
 
     assert particle.charge == -1
     assert math.isclose(particle.mass, 1.38, rel_tol=1e-3)
@@ -915,8 +1134,10 @@ def test_create_quark_from_pdg_valid_values():
 
 def test_create_gluon_from_pdg_valid_values():
     particle_quantity_JETSCAPE = np.array(
-        [0, 21, 11, 2.01351754, 1.30688601, -0.422958786, -0.512249773])
-    particle = Particle(input_format="JETSCAPE",
-                        particle_array=particle_quantity_JETSCAPE)
+        [0, 21, 11, 2.01351754, 1.30688601, -0.422958786, -0.512249773]
+    )
+    particle = Particle(
+        input_format="JETSCAPE", particle_array=particle_quantity_JETSCAPE
+    )
 
     assert particle.charge == 0
